Common

Description

Documentation

type Map = [String]

Tipo que representa o mapa do jogo.

type Pos = (Int, Int)

Tipo que representa uma posição.

isRock :: Pos -> Map -> Bool

Função que verifica se existe uma pedra numa dada posição.

isBrick :: Pos -> Map -> Bool

Função que verifica se existe um tijolo numa dada posição.

isInPos :: Pos -> Map -> [Char] -> Bool

Função que verifica se existem algum Powerup, bomba, ou jogador (dependendo do terceiro argumento) numa dada posição.

>>> isInPos (3,8) ["! 3 8"] ['!'] = True

>>> isInPos (3,8) ["2 3 8"] ['0','1','2','3'] = True

>>> isInPos (3,8) ["+ 3 8"] ['*'] = False

>>> isInPos (5,2) ["+ 3 8"] ['+'] = False

getPos :: [String] -> Pos -> Char

Função que retorna o conteúdo de uma célula.

>>> getPos ["#######","#     #","# # # #",
            "# ??  #","# #?# #","#     #",
            "#######","+ 3 3","+ 3 4",
            "* 3 5 0 1 10","0 3 5 +","1 5 5"] (3,3) = "+" 

setPos :: Map -> Pos -> Char -> Map

Função que dados um mapa, uma posição e um carater, retorna uma mapa onde o carater na posição dada foi alterado para o carater no terceiro argumento.

setPosAux :: String -> Int -> Char -> String

Função que dada uma linha do mapa, o numero da coluna a alterar e um carater, retorna a linha com essa coluna alterada para o carater.

rmPW :: Map -> Pos -> Map

Função que marca um powerup numa dada posição como removido.

calcDim :: Map -> Int

Função que dado um mapa calcula a sua dimensão.

calcTicks :: Int -> Int

Função que dada a dimensão de uma mapa calcula o número de ticks necessários até que a espiral comece.

spiral :: Map -> Int -> Int -> Pos -> (Int, Int) -> Map

Função que implementa a espiral.

reactToEdge :: Map -> Pos -> (Int, Int) -> Int -> ((Int, Int), Int)

Função que muda a direção de um vetor quando a dada posição de aproxima um dado número de blocos da borda, e aumenta este último número sempre que muda de direção para cima. A funçäo determina em que direção a espiral deverá se mover no próximo tick.