Main

Description

Copyrigth: Alexandre Mendonça Pinho a82441@alunos.uminho.pt; Joel Filipe Esteves Gama a82202@alunos.uminho.pt;

O módulo contém os mecanismos para o movimento dos jogadores no mapa.

Documentation

type Pos = (Int, Int)

Tipo que representa posições no mapa.

move :: [String] -> Int -> Char -> [String]

Função que, dado o estado actual do jogo no formato acima descrito, o identificador de um jogador e o comando, deverá devolver o novo estado do jogo.

cmdU :: [String] -> Int -> [String]

Função auxiliar que define o comando U (ir para cima).

cmdD :: [String] -> Int -> [String]

Função auxiliar que define o comando D (ir para baixo).

cmdL :: [String] -> Int -> [String]

Função auxiliar que define o comando L (ir para a esquerda).

cmdR :: [String] -> Int -> [String]

Função auxiliar que define o comando R (ir para a direita).

cmdB :: [String] -> Int -> [String]

Função auxiliar que define o comando B (colocar bomba).

isBomb :: [String] -> Pos -> Bool

Função que verifica se existe uma bomba numa dada posição.

>>> isBomb ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5","1 5 5"] (3,5)
           True

>>> isBomb ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5","1 5 5"] (5,5)
           False

setBomb :: [String] -> Int -> [String]

Função que posiciona uma bomba numa dada posição se ela já não existir.

>>> setBomb ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 3 5","1 5 5"] 0
            ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5","1 5 5"]

maxBombs :: [String] -> Int -> Int

Função que retorna o número máximo de bombas que um jogador pode colocar.

>>> maxBombs ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 3 5 ++","1 5 5"] 0
             3

countBombs :: [String] -> Int -> Int

Função que retorna o número de bombas que um jogador tem colocadas no mapa.

>>> countBombs ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5","1 5 5"] 0
               1

addPWInPos :: [String] -> String -> Pos -> String

Função que adiciona, se existir, o Power Up da posição ao jogador.

>>> addPWInPos ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 3 3"] "" (3,3)
               "+" 

getPlayerPWs :: [String] -> Int -> String

Função que retorna uma String com os Power Ups do jogador x.

>>> getPlayerPWs ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"]
                 "+"

setPlayer :: [String] -> Int -> Pos -> String -> [String]

Função que altera a posição e quantidade de Power Ups de um jogador.

>>> setPlayer ["0 3 5"] 0 (4,5) "+!"
              ["0 4 5 +!"]

getPos :: [String] -> Pos -> Char

Função que retorna o conteúdo de uma célula.

>>> getPos ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"] (3,3)
           "+" 

getPlayerPos :: [String] -> Int -> Pos

Função que retorna a possição de um jogador.

>>> getPlayerPos ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"] 0
                 (3,5) 

removePW :: [String] -> Pos -> [String]

Função que, dado o estado atual do jogo e uma posição, revome, se existir, o Power Up nessa posição.

>>> removePW ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"] (3,3)
             ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"]

main :: IO ()

Função main.