T1Lib

Description

Este módulo contém as funções, da Tarefa 1, para a criação de mapas do jogo Bomberman.

Documentation

type Powerup = (Char, Int, Int)

O type Powerup, corresponde à informação num tuplo de um Power Up.

type LinhaMapa = String

O type LinhaMapa, corresponde a uma linha do mapa __Bomberman.

type Mapa = [LinhaMapa]

O type Mapa, corresponde ao mapa do Bomberman, que consiste numa lista de LinhaMapa.

type Resto_InfRandom = String

O type Resto_InfRandom, corresponde ao resto da informação do mapa.

type Inf_Mapa = (LinhaMapa, [Powerup], Resto_InfRandom)

O type Inf_Mapa, corresponde à informação do mapa.

num_celulas :: Int -> Int

Descrição:

A função num_celulas, determina o número de células que cada mapa irá ter para informação de power ups, tijolos ou simplesmente espaços vazios.

Exemplo:

>>> num_celulas 9
28

pilares_pedra :: Int -> Int

Descrição:

A função pilares_pedra e a respectiva auxiliar, calculam o número de pilares de pedra, ou seja, as pedras que se encontram na linhas e colunas pares.

Exemplos:

>>> pilares_pedra 9
9

aux_pilares_pedra :: Int -> Int -> Int -> Int

numeros_Random :: Int -> Int -> [Int]

Descrição:

A função numeros_Random, determina um número de forma aleatória para cada célula do mapa, excepto as células vazias e as que tem pedra.

Exemplo:

>>> numeros_Random 9 0
[83,93,63,38,0,87,81,1,61,86,13,50,32,80,54,25,90,31,65,92,2,76,70,25,6,29,10,99]

determina_Char :: [Int] -> [Char]

A forma como se determina o conteúdo de uma célula a partir do respectivo número aleatório é a seguinte:

0 a 1 -> Power up Bombs escondido atrás de um tijolo.

2 a 3 -> Power up Flames escondido atrás de um tijolo.

4 a 39 -> Tijolo.

40 a 99 -> vazio

determina_Char [] = [] 
determina_Char (h:t) | h >= 0 && h <= 1 = +:determina_Char t 
                     | h >= 2 && h <= 3 = !:determina_Char t 
                     | h >= 4 && h <= 39 = ?:determina_Char t 
                     | h >= 40 && h <= 99 = ' ':determina_Char t 

Descrição:

A função determina_Char, recebe uma lista do tipo Int, e através dos intervalos propostos no Guião do Projecto, obtêm-se uma lista de caracteres, em que temos + -> powerUps Bombs / ! -> powerUps Flames ' ' -> espaço / ? -> tijolo.

Exemplo:

>>> determina_Char [83,93,63,38,0,87,81,1,61,86,13,50,32,80,54,25,90,31,65,92,2,76,70,25,6,29,10,99]
"   ?+  +  ? ?  ? ?  !  ???? "

determina_PowerUp :: Char -> Bool

Descrição:

A função determina_PowerUp, recebe um caracter e determina se este é um powerUp, ou seja, se é igual a + ou ! e retorna True se for PowerUp e False caso contrário.

Exemplo:

>>> determina_PowerUp '+'
True

linhaCimaBaixo :: Int -> String

Descrição:

A função linhaCimaBaixo gera a parede superior e inferior do mapa, ou seja, quando é tudo pedra.

Exemplo:

>>> linhaCimaBaixo 9
"#########"

Usamos como base o mapa de dimensão 5 em que temos "#####".

sPilares_normal :: Int -> Int -> Int -> String -> Inf_Mapa

Descrição:

A função sPilares_normal constrói as linhas em que não temos pilares, ou seja, pedras no meio e que não tem as células vazias para os jogadores se movimentar no ínicio do jogo.

Exemplo:

>>> sPilares_normal 9 3 0 "?+  +  ? ?  ? ?  !  ???? "
("#??     #",('+',2,3),"? ?  ? ?  !  ???? ")

Usamos no exemplo uma lista de caracteres aleatória.

sPilares_Cvazios :: Int -> Int -> Int -> String -> Inf_Mapa

Descrição:

A função sPilares_Cvazios contrói as linhas em que não temos pilares, ou seja, pedras no meio, mas temos 4 células vazias para o jogador se movimentar no ínico do jogo.

Exemplo:

>>> sPilares_Cvazios 9 1 0 "?+  +  ? ?  ? ?  !  ???? "
"(#  ??   #",('+',4,1)," +  ? ?  ? ?  !  ???? ")

Usamos no exemplo uma lista de caracteres aleatória.

cPilares_Cvazios :: Int -> Int -> Int -> String -> Inf_Mapa

Descrição:

A função cPilares_Cvazios constrói as linhas em que temos pilares, ou seja, pedras no meio, e contém 2 células vazias para o jogador se movimentar no ínicio do jogo.

Exemplo:

>>> cPilares_Cvazios 9 2 0 "?+  +  ? ?  ? ?  !  ???? "
("# #?#?# #",('+',5,2),"  +  ? ?  ? ?  !  ???? ")

Usamos no exemplo uma lista de caracteres aleatória.

cPilares_normal :: Int -> Int -> Int -> String -> Inf_Mapa

Descrição:

A função cPilares_normal constrói as linhas em que temos pilares, ou seja, pedras no meio, mas não contém as células vazias para o jogador se movimentar no ínicio.

Exemplo:

>>> cPilares_normal 9 4 0 "?+  +  ? ?  ? ?  !  ???? "
(#?#?# # #,('+',3,4),"+  ? ?  ? ?  !  ???? ")

Usamos no exemplo uma lista de caracteres aleatória.

linha_tabuleiro :: Int -> Int -> String -> (Mapa, [Powerup])

Descrição:

A função linha_tabuleiro, dependendo da linha ela determina a função que deve ser usada para essa linhas.

As funções determinadas, são as que definimos anteriormente.

organizaPoweups :: [Powerup] -> ([Powerup], [Powerup])

Descrição:

A função organizaPoweups organiza os power ups, primeiro os power ups bombs e depois os power ups flames.

converte :: (Char, Int, Int) -> String

Descrição:

A função converte, converte a informação de um power up, de um tuplo para uma String.

Exemplo:

>>> converte ('+',7,1)
"+ 7 1"

mapa :: Int -> Int -> [String]

A função mapa, contrói o mapa e junta-o à informação dos power ups bombs e flames.