T2Lib

Documentation

determina_pos :: Char -> (Int, Int) -> (Int, Int)

A função determina_pos, recebe como parâmetros o comando que é um Char e as coordenadas atuais do jogador, e a partir desse comando descobrir qual a posição para onde o jogador quer, e descobrir através desse comando a direção e alterar o valor nas coordenadas.

ocorrePedraTijolo :: [String] -> Int -> Int -> Bool

Descrição:

A função ocorrePedraTijolo, tem como objectivo verificar se na posição c1 l1, existe uma pedra ou um tijolo, pois caso exista o jogador não se pode mover para essa posição, retornando nesse caso True, ou seja, existe uma pedra nessa posição.

Exemplo:

>>> ocorrePedraTijolo (mapa 11 0) 2 2
True

ocorreLinha :: String -> Int -> Int -> Bool

Descrição:

Utilizamos como função auxiliar ocorreLinha que verifica através da coluna se existe um pedra num linha x com coluna y, caso exista retorna True, caso contrário retorna False. Basicamente a função ocorrePedraTijolo determina a linha em que queremos verificar a existência de pedra ou tijolo, enquanto que a função __ocorreLinha verifica se existe uma pedra na coluna indicada, retornando assim um Bool.

verifica_bomb :: [String] -> Int -> Int -> Bool

Descrição:

A função verifica_bomb, recebe o mapa inicial e as coordenadas para verificar se existe uma bomba nessa posição.

Exemplo:

>>> verifica_bomb (mapa 11 0)++["* 7 7 1 1 0","0 3 1","1 7 7"] 7 7
True

verificaBombPlayer :: [String] -> Int -> Bool

Descrição:

A função verificaBombPlayer,verifica se um jogador tem powerups bombs para colocar mais uma bomba.

Exemplo:

>>> verificaBombPlayer ((mapa 11 0)++["* 6 7 0 1 10","0 5 7 +"]) 0
True

A função mapa está definida no T1Lib.

det_pUPbombs :: [String] -> Int -> Int

Descrição:

A função det_pUPbombs e a respectiva auxiliar, determinam a quantidade de power ups bombs que o jogador x tem.

Exemplo

>>> det_pUPbombs ((mapa 11 0)++["0 3 1 ++"]) 0
2

A função mapa está definida no T1Lib.

aux_det_pUPbombs :: String -> Int

detBomb_ativadas :: [String] -> Int -> Int

Descrição:

A função detBomb_ativadas, determina quantas bombas ativadas, um jogador x tem.

Exemplo:

>>> detBomb_ativadas ((mapa 11 0)++["* 6 7 0 1 10","0 5 7 +"]) 0
1

A função mapa está definida no T1Lib.

validar_comando :: [String] -> Char -> Int -> (Int, Int) -> Bool

Descrição:

A função validar_comando, verifica se a jogada é válida.

Nota:

Para isso usamos a recorrência da função ocorrePedraTijolo, para determinar se existe um pedra na posição para onde o jogador quer ir.

Exemplo:

>>> validar_comando ((mapa 11 0)++["* 6 7 0 1 10","0 5 7 +"]) 'U' 0 (5,7)
False

A função mapa está definida no T1Lib.

verifica_pUp :: [String] -> Int -> Int -> (Char, Bool)

Descrição:

A função verifica_pUp, verifica se existe um power up na posição indicada.

Exemplo:

>>> verifica_pUp (mapa 11 0) 7 1
('+',True)

A função mapa está definida no T1Lib.

add_powerUp :: [String] -> Int -> Char -> [String]

Descrição:

A função add_powerUp e a respectiva auxiliar, adiciona um power up bomb ou flame, de forma ordenada, ou seja, primeiro bombs e depois flames

Exemplo:

>>> add_powerUp ((mapa 11 0)++["0 3 1 !"]) 0 '+'
((mapa 11 0)++["0 3 1 +!"])

A função mapa está definida no T1Lib.

aux_add_powerUp :: String -> Char -> String

det_flames :: [String] -> Int -> Int

Descrição:

A função det_flames e a respectiva auxiliar, determina quantos power ups flames um jogador tem.

Exemplo:

>>> det_flames ((mapa 11 0)++["0 1 1 +!!"]) 0
2

A função mapa está definida no T1Lib.

aux_det_flames :: String -> Int

ativar_bomb :: [String] -> Int -> Int -> Int -> Int -> [String]

Descrição:

A função ativar_bomb, ativa um bomba de um jogador, imprimindo-a depois dos power ups e antes da informação dos jogadores, caso já haja uma bomba, a função coloca de forma ordenada, através das coordenadas.

Exemplo:

>>> ativar_bomb ((mapa 11 0)++["0 3 1 !"]) 3 1 0 1
((mapa 11 0)++["* 3 1 0 2 10","0 3 1 !"])

A função mapa está definida no T1Lib.

update_positions :: [String] -> Int -> Char -> Int -> Int -> [String]

Descrição:

A função update_positions atualiza as posições do jogador.

Exemplo:

>>> update_positions ((mapa 11 0)++["0 3 1"]) 0 'L' 2 1
((mapa 11 0)++["0 2 1"])

A função mapa está definida no T1Lib.

remove_InfpowUp :: [String] -> Char -> Int -> Int -> [String]

Descrição:

A função remove_InfpowUp remove a informação de um power up que tenha sido apanhado por um jogador.

Exemplo:

>>> remove_InfpowUp ((mapa 11 0)++["! 9 4"]) '+' 9 4
(mapa 11 0)

A função mapa está definida no T1Lib.

det_pos_jogador :: [String] -> Int -> (Int, Int)

Descrição:

A função det_pos_jogador determina através do número do jogador, a posição em que ele se encontra.

Exemplo:

>>> det_pos_jogador ((mapa 11 0)++["0 3 1"]) 0
(3,1)

A função mapa está definida no T1Lib.

converte_INF_pUPtoInt :: [String] -> (Char, Int, Int)

Descrição:

A função converte_INF_pUPtoInt converte a informação de um power up de string para um triplo.

Exemplo:

>>> converte_INF_pUPtoInt ["+","7","1"]
('+',7,1)

converte_INFplayer :: [String] -> ([Int], String)

Descrição:

A função converte_INFplayer converte a informação do jogador para um tuplo.

Exemplo:

>>> converte_INFplayer ["0","3","1"]
(0,3,1)

converte_INFcoord :: [String] -> [Int]

Descrição:

A função converte_INFcoord converte a informação das coordenadas para uma lista de inteiros.

Exemplo:

>>> converte_INFcoord ["0","1"]
[0,1]

converte_INFbomb :: [String] -> (Char, Int, Int, Int, Int, Int)

Descrição:

A função converte_INFbomb converte a informação da bomba, para um tuplo.

Exemplo:

>>> converte_INFbomb (["*","7","1","0","1","10"])
('*',7,1,0,1,10)

det_mapa_res :: [String] -> Char -> Int -> Int -> Int -> [String]

Descrição:

A função det_mapa_res determina através do comando (Char) introduzido, qual as funções que tem que utilizar, retornando assim o mapa com as alterações feitas pelo respectivo comando.

Exemplo:

>>> det_mapa_res ((mapa 11 0)++["0 3 1","1 2 1"]) 'B' 0 3 1
((mapa 11 9)++["* 3 1 0 1 10","0 3 1","1 2 1"])

A função mapa está definida no T1Lib.

move :: [String] -> Int -> Char -> [String]

Descrição:

A função move valida a jogada, caso esta seja inválida, ele retorna o mapa inicial, caso contrário ele determina através da função det_mapa_res as funções que tem que utilizar para comando.

Notas Importantes:

Caso o jogador não exista a minha função det_pos_jogador retorna-me (0,0) e assim eu retorno o mapa inicial, visto que a jogada não é válida!