T3Lib

Documentation

encode :: [String] -> String

Descrição:

A função encode remove inicialmente as pedras e depois remove os espaços do mapa.

Exemplos:

>>> encode (mapa 9 0)
"9\n7\n1??1\n2?2?1\n?2?\n1?2?2\n1??1\n2??3\nf+ 5 2\n+ 3 3\n! 5 5\n"

Notas:

A função mapa está definida na biblioteca da Tarefa1.hs denominada T1Lib.

remove_pedras :: [String] -> String

Descrição:

A função remove_pedras remove todas as pedras do mapa, incluindo as interiores.

Esta função ainda acrescenta um f antes do ínicio da informação dos power ups.

aux_remove_pedras :: [String] -> Int -> String

remove_spaces :: String -> String

Descrição:

A função remove_spaces remove os espaços do mapa, mas guarda o número de espaços na String, para depois recuperá-los na descompressão.

Notas:

Quando a função encontra f sabes que tem que parar de remover espaços.

aux_remove_spaces :: String -> Int -> String

lista_infMapa :: [String] -> String

Descrição:

A função lista_infMapa separa as coordenadas dos power ups com um '\n'.

delete_stone :: String -> String

Descrição:

A função delete_stone apaga todas as pedras do mapa

Notas:

Esta função é usada na função remove_pedras.

decode :: String -> [String]

Descrição:

A função decode recebe o mapa comprimido numa String, desenha o mapa inicial através dessa String.

Exemplos:

>>> decode (encode (mapa 9 0))
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

Notas:

Como podemos observar a função decode originou o mapa igual ao ínicial.

>>> mapa 9 0
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

A função mapa está definida na biblioteca da Tarefa1.hs denominada T1Lib.

aux_decode :: String -> t -> [String]

up_info :: String -> Int

Descrição:

A função up_info retorna o tamanho do mapa comprimido, tal como podemos ver nas funções anteriores o tamanho do mapa é guardado no início da String.

Exemplos:

>>> up_info "7\n5\n3\n1??2\n1?1\n5\nf+ 3 3\n+ 3 4\n"
7
>>> up_info "\r\n5??4\n1?2?1\n1?2?1?2?1\n1????1\n2?1????3\n?2?2\n9?1\n?1?2?\n3?1?2?2\n1?4\n2????2?2\nf+ 7 1\n+ 3 2\n+ 7 5\n+ 8 5\n! 10 3\n! 1 8\n! 4 9\n! 3 10\n! 6 11\n"
13

Como podemos ver no segundo exemplo, a função encode colocou o tamanho do mapa em caracter, como estratégia de compressão, mas a função up_info está preparada para esses casos.

take_Snumber :: String -> String

Descrição:

A função take_Snumber isola-me a String com a informação do tamanho do mapa, para depois ser usada na função up_info.

Exemplos:

>>> take_Snumber "7\n5\n3\n1??2\n1?1\n5\nf+ 3 3\n+ 3 4\n"
"7"

delete_inf :: String -> String

Descrição:

A função delete_inf elemina a informção do tamanho do mapa que se encontra no início da String comprimida.

Exemplos:

>>> delete_inf "7\n5\n3\n1??2\n1?1\n5\nf+ 3 3\n+ 3 4\n"
"s\n5\n3\n1??2\n1?1\n5\nf+ 3 3\n+ 3 4\n"

Notas:

Como podemos ver no exemplo, a função delete_inf coloca um s no início da String para saber onde começa o mapa.

impMapa :: String -> Int -> [String]

Descrição:

A função impMapa e a respectiva auxiliar, imprime o mapa através da String comprimida pela função encode, utilizando diversas funções.

Notas:

Esta função determina as funções que tem de utilizar, através da linha que é inicializada a 0, e posteriormente incrementada.

aux_impMapa :: String -> Int -> Int -> [String]

need_list :: String -> (String, String)

Descrição:

A função need_list determina a String que tem que ser enviada para as funções responsáveis por desenhar as linhas do mapa.

linhaSpilares :: String -> String

Descrição:

A função linhaSpilares desenha as linhas que não tem pedras no meio, recebendo a lista com os caracteres especiais, ou seja, tijolos ou espaços vazios.

Exemplos:

>>> linhaSpilares " ? ??  ?"
"# ? ??  ?#"

linhaCpilares :: String -> Int -> Int -> String

Descrição:

A função linhaCpilares desenha as linhas que tem pedras no meio, recebendo a lista com os caracteres especiais, ou seja, tijolos ou espaços vazios.

Exemplos:

>>> linhaCpilares " ? ? " 11 0
"# #?# #?# #"

draw_spaces :: String -> String

Descrição:

A função draw_spaces e desenhaLinha desenham os espaços vazios no mapa, através dos números que aparecem na String comprimida.

desenhaLinha :: [String] -> String

desenha_spaces :: Int -> String

Descrição:

A função desenha_spaces recebe um número inteiro e retorna uma String com o número de espaços igual a esse inteiro.

Exemplo:

>>> desenha_spaces 4
"    "

agarraLinhas :: String -> (String, String)

Descrição:

A função agarraLinhas seleciona uma linha entre new lines, que corresponde a uma linha do mapa.

Exemplos:

>>> agarraLinhas "#       #\n# # # # #\n"
("#       #","\n# # # # #\n")

Notas:

A String usada no exemplo é aleatória.

separaElem :: String -> [String]

Descrição:

A função separaElem separa num lista de Strings, todos os elementos de uma lista, mas usando a função isola_num.

Exemplo:

>>> separaElem "3??1"
["3","??","1",""]

Notas:

A String utilizada é aleatória.

isola_num :: String -> (String, String, String)

Descrição:

A função isola_num isola um número de um String, ou seja, recolhe o número, guardando depois o resto e tijolos.

Exemplos:

>>> isola_num "13??  "
("13","  ","??")

Notas:

A String utilizada é aleatória.

getIntrr :: String -> (String, String)

Descrição:

A função getIntrr junta os pontos de interrogação que encontrar seguidos e depois retorna o resto da lista que não foi usada.

Exemplos:

>>> getIntrr "???    ?"
("    ?","???")

Notas:

A String utilizada é aleatória.