T4Lib

Description

Este módulo contém as funções, da Tarefa 4, para a reação do jogo à passagem do tempo. Nesta Tarefa, verifica-se as alterações causadas pela passagem do tempo, em especial atenção, à explosão de bombas e a respetiva destruição.

Documentation

type InfGame = (Tabuleiro, [PowUp], [Bomba], Players)

Descrição:

O type InfGame, separa a informação do mapa, num tuplo, para facilitar o acesso a diferentes informações.

data PowUp

Descrição:

O tipo PowUp, contém dois construtores para selecionar qual o tipo de Power Up, e a sua posição.

Constructors

Bomb Posicao

Flame Posicao

data Player

Descrição:

O tipo Player, contém dois construtores, Vivo e Morto, em que no caso de Vivo, indicamos ainda a sua posição, número de Power Ups bombs e flames.

Constructors

Vivo Posicao NumbBomb NumbFlame

Morto

data Bloco

Descrição:

O tipo Bloco, contém três construtores, que são os diferentes tipos de Blocos (Pedra,Tijolo,Vazio), que se encontam no Mapa do Jogo Bomberman.

Constructors

Pedra

Tijolo

Vazio

data Direction

Descrição:

O tipo Direction, contém 4 construtores, das diferetes direções possíveis do flame das bombas no Bomberman.

Constructors

Cima

Baixo

Esquerda

Direita

type Players = [Player]

O type Players, corresponde à lista de Jogadores.

type Tabuleiro = [[Bloco]]

O type Tabuleiro, corresponde à lista de Blocos, ou seja, o mapa.

type Posicao = (Int, Int)

O type Posicao, corresponde às coordenadas no mapa.

type LinhaTabuleiro = [Bloco]

O type LinhaTabuleiro, corresponde a uma linha no mapa Bomberman.

type Bomba = (Posicao, Jogador, NumbFlame, Time)

O type Bomba, corresponde a um tuplo, com toda a informação necessária de uma bomba.

type NumbBomb = Int

O type NumbBomb, corresponde ao número de power ups Bombs.

type Jogador = Int

O type Jogador, corresponde ao número do jogador.

type NumbFlame = Int

O type NumbFlame, corresponde ao número de power ups Flames__.

type Time = Int

O type Time, corresponde ao tempo atual do jogo.

type Size = Int

O type Size, corresponde ao tamanho do mapa.

bloco :: InfGame -> Posicao -> Bloco

Descrição:

A função bloco, retorna-me o bloco que está na posição que eu indicar.

Exemplos:

>>> bloco (filterInfGame m1) (0,0)
Pedra
>>> bloco (filterInfGame m1) (1,1)
Vazio

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

powerUps :: InfGame -> [PowUp]

Descrição:

A função powerUps retorna-me a lista de power ups que está no InfGame.

Exemplos:

>>> powerUps (filterInfGame m1)
[Bomb (5,2),Bomb (3,3),Flame (5,5)]

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

bombs :: InfGame -> [Bomba]

Descrição:

A função bombs retorna-me a lista de bombas que está no InfGame.

Exemplos:

>>> bombs (filterInfGame m1)
[((2,7),0,1,10)]

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

filterTabuleiro :: [String] -> [String]

Descrição:

A função filterTabuleiro, recebe a informação do jogo em [String] e retorna apenas a informação do Tabuleiro.

Exemplos:

>>> filterTabuleiro m1
["#########","#       #","# # #?# #","#     ? #","# # # #?#","# ?? ?  #","# # #?# #","#       #","#########"]

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

filterPowUp :: [String] -> [String]

Descrição:

A função filterPowUp, recebe a informação do jogo em [String] e retorna apenas a informação dos Power Ups.

Exemplos:

>>> filterPowUp m1
["+ 5 2","+ 3 3","! 5 5"]

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

filterBomb :: [String] -> [String]

Descrição:

A função filterBomb, recebe a informação do jogo em [String] e retorna apenas a informação das Bombas

Exemplos:

>>> filterBomb ["#########","#       #","# # #?# #","#     ? #","# # # #?#","# ?? ?  #","# # #?# #","#       #","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]
["* 2 7 0 1 10"]

filterPlayer :: [String] -> [String]

Descrição:

A função filterPlayer, recebe a informação do jogo em [String] e retorna apenas a informação dos Jogadores.

Exemplos:

>>> filterPlayer ["#########","#       #","# # #?# #","#     ? #","# # # #?#","# ?? ?  #","# # #?# #","#       #","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]
["0 1 1","1 7 1","2 7 7","3 1 7"]

convToTabuleiro :: [String] -> Tabuleiro

Descrição:

A função convToTabuleiro, converte o mapa de [String], para o tipo Tabuleiro = [[Bloco]].

Exemplos:

>>> convToTabuleiro (mapa 5 0)
[[Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra]]
>>> mapa 5 0
["#####","#   #","# # #","#   #","#####"]

Notas:

A função mapa, está definida na biblioteca T1Lib, esta gera Tabuleiros do Bomberman, em [String].

auxconvToTabuleiro :: String -> LinhaTabuleiro

convToPowUp :: [String] -> [PowUp]

Descrição:

A função convToPowUp, converte a informação dos Power Ups de um mapa, para o tipo PowUp.

Exemplos:

>>> convToPowUp ["+ 5 2","+ 3 3","! 5 5"]
[Bomb (5,2),Bomb (3,3),Flame (5,5)]

auxConvToPowUp :: String -> PowUp

convToBomb :: [String] -> [Bomba]

Descrição:

A função convToBomb, converte a informação das Bombas de um mapa, para o type Bomba, que facilita a pesquisa de informação das Bombas.

Exemplos:

>>> convToBomb ["* 2 7 0 1 10"]
[((2,7),0,1,10)]

auxConvToBomb :: String -> Bomba

convToPlayer :: [String] -> Players

Descrição:

A função convToPlayer, converte a informação dos Jogadores, para o tipo Player, que neste caso como é uma lista, usamos o type Players = [Player].

Exemplos:

>>> convToPlayer ["0 1 1 +!","1 7 1","3 1 7 ++!!!"]
[Vivo (1,1) 1 1,Vivo (7,1) 0 0,Morto,Vivo (1,7) 2 3]

Notas:

Como podemos ver no exemplo, o jogador número 2, não existe, logo está Morto, aqui podemos ver a importância da utilização do data Player.

auxConvToPlayer :: String -> Player

counterBmbFlm :: String -> (Int, Int)

Descrição:

A função counterBmbFlm, conta quantos Power Ups Bombs e Flames,um determinado, jogador tem.

Exemplos:

>>> counterBmbFlm "3 1 7 ++!!!"
(2,3)
>>> counterBmbFlm "3 1 7"
(0,0)
>>> counterBmbFlm "3 1 7 !"
(0,1)

Notas:

Em que o primeiro elemento do tuplo, refere-se ao número de Power Up Bombs e o segundo aos Flames.

filterInfGame :: [String] -> InfGame

Descrição:

A função filterInfGame, conjuga, todas as funções (convToTabuleiro,convToPowUp,convToBomb,convToPlayer), para que seja possível, converter um mapa no tipo [String] para o tuplo de type InfGame.

filterInfGame m =
    (convToTabuleiro (filterTabuleiro m), convToPowUp (filterPowUp m), convToBomb (filterBomb m), convToPlayer (filterPlayer m))

Exemplos:

>>> filterInfGame m1
([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[((2,7),0,1,10)],[Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])
>>> filterInfGame (mapa 5 0)
([[Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra]],[],[],[Morto,Morto,Morto,Morto])

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

A função mapa, está definida na biblioteca T1Lib, esta gera Tabuleiros do Bomberman, em [String].

As funções filterTabuleiro, filterPowUp,filterBomb,filterPlayer, separam as respectivas partes do mapa inicial, ou seja, separa a informação do mapa, power ups, bombas e jogadores.

checkPosBloco :: InfGame -> Posicao -> Bloco -> InfGame

Descrição:

A função checkPosBloco, substitui um bloco na posição que recebe como parâmetro.

Exemplos:

>>> checkPosBloco (filterInfGame m1) (1,1) Tijolo
([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Tijolo,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[((2,7),0,1,10)],[Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])
>>> checkPosBloco (filterInfGame m1) (0,0) Vazio
([[Vazio,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[((2,7),0,1,10)],[Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 10","0 1 1","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

removeBlocoMap :: Tabuleiro -> Posicao -> Bloco -> Posicao -> Tabuleiro

removeBlocoLinhaMap :: LinhaTabuleiro -> Posicao -> Posicao -> Bloco -> LinhaTabuleiro

checkPosPowerUp :: [PowUp] -> Posicao -> (Bool, [PowUp])

Descrição:

A função checkPosPowerUp, verifica se existe um Power Up na posição indicada e caso exista retorna True e a lista dos Power Ups sem esse Power Up, ou seja, remove-o, caso contrário, retorna False e a lista dos Power Ups completa.

Exemplos:

>>> checkPosPowerUp [Bomb (5,2),Bomb (3,3),Flame (5,5)] (5,2)
(True,[Bomb (3,3),Flame (5,5)])
>>> checkPosPowerUp [Bomb (5,2),Bomb (3,3),Flame (5,5)] (5,4)
(False,[Bomb (5,2),Bomb (3,3),Flame (5,5)])

checkPosPlayer :: Players -> Posicao -> (Bool, Players)

Descrição:

A função checkPosPlayer, verifica se existe um jogador na posição indicada, caso exista ele vai Morrer, porque foi atingido por uma bomba. Portanto se ele for atingido, retornamos True e a lista dos jogadores, com este Morto, caso contrário, retornamos False e a lista dos jogadores tal como a recebemos.

Exemplos:

>>> checkPosPlayer [Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0] (1,1)
(True,[Morto,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])
>>> checkPosPlayer [Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0] (7,8)
(False,[Vivo (1,1) 0 0,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])

checkPosBomb :: [Bomba] -> Posicao -> (Bool, [Bomba])

Descrição:

A função checkPosBomb, determina se existe uma bomba na posição indicada caso exista, retorna True e a bomba com o tempo de explosão = 1, para que possa expludir no próximo instante, se não for atingida, retorna False e a lista das bombas tal como a recebeu.

Exemplos:

>>> checkPosBomb [((2,7),0,1,10)] (2,7)
(True,[((2,7),0,1,1)])
>>> checkPosBomb [((2,7),0,1,10)] (2,1)
(False,[((2,7),0,1,10)])

detRangeCoords :: Bomba -> Size -> InfGame -> [Posicao]

Descrição:

A função detRangeCoords, determina o range de explosão de uma bomba, retornando apenas as coordenadas que são afetadas pelo flame da bomba.

Exemplos:

>>> detRangeCoords ((2,7),0,3,1) 9 ([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[((2,7),0,3,1)],[Vivo (1,1) 0 2,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])
[(2,7),(1,7),(3,7),(4,7),(5,7)]

Notas:

O flame da Bomba não ultrapassa pedras, no caso dos Tijolos e Power Ups , não são ultrapassados, mas são destruídos, e tudo isso é avaliado nesta função.

detCoords :: Direction -> Posicao -> Posicao

Descrição:

A função detCoords, determina através de um constructor de Direction qual o tipo de direção a aplicar às coordenadas, retornando assim uma nova Posição.

Exemplos:

>>> detCoords Cima (1,2)
(1,1)
>>> detCoords Baixo (1,2)
(1,3)
>>> detCoords Esquerda (1,2)
(0,2)

checkConsequences :: [Posicao] -> InfGame -> InfGame

Descrição:

A função checkConsequences, recebe a lista das posições afetadas pelo flame de um bomba e determina as consequências da explosão, ou seja, a morte de jogadores, destruição de tijolos, power ups e ativação de bombas, retornando o mapa com as alterações feitas.

collectsBombs :: [Bomba] -> ([Bomba], [Bomba])

Descrição:

A função collectsBombs, separa as bombas que vão explodir neste instante de tempo das que vão explodir mais tarde, colocando estas com -1 segundo do tempo inicial.

Exemplos:

>>> collectsBombs [((2,7),0,1,10),((1,7),1,1,1)]
([((1,7),1,1,1)],[((2,7),0,1,9)])
>>> collectsBombs [((2,7),0,1,10),((1,7),1,1,2)]
([],[((2,7),0,1,9),((1,7),1,1,1)])
>>> collectsBombs [((2,7),0,1,1),((1,7),1,1,1)]
([((2,7),0,1,1),((1,7),1,1,1)],[])

Notas:

As bombas que não explodem neste instante, ficam com menos um segundo do tempo inicial.

rangeExplosion :: InfGame -> [Bomba] -> [Posicao]

Descrição:

A função rangeExplosion, determina todas as posições que irão ser afetadas por bombas.

Exemplos:

>>> rangeExplosion (filterInfGame m1) [((2,7),0,1,1),((1,7),1,1,1)]
[(2,7),(1,7),(3,7),(1,7),(1,6),(2,7)]

Notas:

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 1","* 1 7 1 1 1","0 2 7","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

explosionBombs :: InfGame -> InfGame

Descrição:

A função explosionBombs, verifica e faz as alterações no mapa, causadas pela explosão de bombas.

Exemplos:

>>> explosionBombs (filterInfGame m1)
([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[],[Morto,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Morto])

Notas:

Como podemos ver no exemplo, as bombas causaram a morte ao jogador 0 e 3, são estas algumas das consequências.

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 1","* 1 7 1 1 1","0 2 7","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

decodeInfGameToString :: InfGame -> [String]

Descrição:

A função decodeInfGameToString, converte o mapa do tipo InfGame, para [String].

Exemplos:

>>> decodeInfGameToString (filterInfGame (mapa 9 0))
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]
>>> mapa 9 0
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

Notas:

A função filterInfGame, converte de [String], para o data InfGame.

decodeTabuleiro :: Tabuleiro -> [String]

Descrição:

A função decodeTabuleiro, converte a informação do tabuleiro do tipo data Bloco, para [String].

decodePowerUps :: [PowUp] -> [String]

Descrição:

A função decodePowerUps, converte a informação dos Power Ups do tipo data PowUp, para [String].

decodeBombas :: [Bomba] -> [String]

Descrição:

A função decodeBombas, converte a informação das Bombas do tipo data [Bomba], para [String].

decodePlayers :: Players -> [String]

Descrição:

A função decodePlayers, converte a informação dos Jogadores do tipo data Players, para [String].

drawSpiral :: InfGame -> Time -> InfGame

Descrição:

A função drawSpiral, desenha a expiral, consoante o tempo de jogo, colocando uma pedra na posição destinada, e destruindo tudo o que estiver nessa posição, retornado assim o mapa com todas as alterações causadas pela expiral.

Exemplos:

>>> drawSpiral (filterInfGame m1) 49
([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[],[Morto,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])

Notas:

Como podemos ver no exemplo, caiu a primeira pedra da expiral, na posição (1,1) e matou o jogador 0.

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 1","* 1 7 1 1 1","0 2 7","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

checkSpiral :: [String] -> Time -> Bool

Descrição:

A função checkSpiral, verifica se o tempo recebido, é menor ou igual ao tempo de ínicio da expiral.

Exemplos:

>>> checkSpiral m1 49
True
>>> checkSpiral m1 50
False
>>> checkSpiral m1 25
True

Notas:

No mapa de tamanho 9, ou seja, como o mapa m1, o tempo do ínicio da Expiral é aos 49 segundos.

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 1","* 1 7 1 1 1","0 2 7","1 7 1","2 7 7","3 1 7"]

destructionSpiral :: Posicao -> InfGame -> InfGame

Descrição:

A função destructionSpiral, determina a destruição causada pela pedra que cai do efeito expiral, podendo destruir um tijolo, uma bomba, ou até matar um jogador.

Exemplos:

>>> destructionSpiral (1,1) (filterInfGame m1)
([[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra],[Pedra,Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Tijolo,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra],[Pedra,Vazio,Tijolo,Tijolo,Vazio,Tijolo,Vazio,Vazio,Pedra],[Pedra,Vazio,Pedra,Vazio,Pedra,Tijolo,Pedra,Vazio,Pedra],[Pedra,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Pedra],[Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra,Pedra]],[Bomb (5,2),Bomb (3,3),Flame (5,5)],[],[Morto,Vivo (7,1) 0 0,Vivo (7,7) 0 0,Vivo (1,7) 0 0])

Notas:

Como podemos ver no exemplo, caiu a primeira pedra da expiral, na posição (1,1) e matou o jogador 0.

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 2 7 0 1 1","* 1 7 1 1 1","0 2 7","1 7 1","2 7 7","3 1 7"]

A função filterInfGame, converte de [String], para o data InfGame.

removeBomb :: [Bomba] -> Posicao -> [Bomba]

Descrição:

A função removeBomb, remove uma bomba que esteja na posição indicada, que neste caso será a posição onde irá cair um pedra da expiral.

Exemplos:

>>> removeBomb [((7,6),0,1,8),((7,7),1,3,1)] (7,6)
[((7,7),1,3,1)]
>>> removeBomb [((7,6),0,1,8),((7,7),1,3,1)] (1,1)
[((7,6),0,1,8),((7,7),1,3,1)]

Notas:

Se não existir nenhuma bomba na posição indicada, a função retorna a lista das Bombas tal e qual como recebeu.

posSpiral :: Size -> [Posicao]

Descrição:

A função posSpiral, determina a lista das posições da expiral, pela ordem que ela ocorre, ou seja, com o efeito expiral, assim só será preciso saber onde é que a pedra tem que cair, como se trata de uma lista, utilizamos as posições da mesma.

Exemplos:

>>> posSpiral 9
[(1,1),(2,1),(3,1),(4,1),(5,1),(6,1),(7,1),(7,2),(7,3),(7,4),(7,5),(7,6),(7,7),(6,7),(5,7),(4,7),(3,7),(2,7),(1,7),(1,6),(1,5),(1,4),(1,3),(1,2),(2,2),(3,2),(4,2),(5,2),(6,2),(6,3),(6,4),(6,5),(6,6),(5,6),(4,6),(3,6),(2,6),(2,5),(2,4),(2,3),(3,3),(4,3),(5,3),(5,4),(5,5),(4,5),(3,5),(3,4),(4,4)]

Notas:

O único parâmetro da função é o tamanho do mapa.

aux_posSpiral :: (Eq a1, Integral a, Num a1) => (a, a) -> a -> a1 -> a -> [(a, a)]

avanca :: [String] -> Int -> [String]

Descrição:

A função avanca, recebe o estado de jogo e o tempo, e retorna o estado de jogo, com todas as alterações afetadas pela passagem tempo, ou seja, pode ocrrer explosão de bombas, ou simplesmente, a redução de 1 segunda nas bombas.

Exemplo:

>>> avanca m1 49
["#########","##      #","# # #?# #","#     ? #","# # # #?#","# ?? ?  #","# # #?# #","#       #","#########","+ 5 2","+ 3 3","! 5 5","* 4 1 0 1 1","1 7 1","2 7 7","3 1 7"]

Notas:

No exemplo podemos ver que a passagem de 1 segundo, matou um jogador, destrui-o um tijolo e caiu um pedra na posição (1,1).

m1 = ["######## ?",""," ? ?? ? ?","","#########","+ 5 2","+ 3 3","! 5 5","* 3 1 1 1 1","* 4 1 0 1 10","0 2 1","1 7 1","2 7 7","3 1 7"]

aux_avanca :: (Eq a, Num a) => [String] -> Time -> a -> [String]