Tarefa5_common

Description

Documentation

data Tile

O tipo de um dado bloco.

Constructors

Rock

Brick

Empty

RemovedTile

data Motion

O movimento de um jogador.

Constructors

StillU

StillD

StillL

StillR

MUp

MDown

MLeft

MRight

type Button = (String, Estado -> Estado)

Representação de um botão, com o seu nome, e a transformação que ocorre ao estado quando o botão é ativado.

data MainState

Estado principal, para saber se se está no menu ou a jogar.

Constructors

Menu [Button] Int

Game

type Pos = (Int, Int)

Posição

data PWType

Tipo de powerup (RemovedPW serve para saber quais aqueles que foram removidos)

Constructors

BombPW

FlamePW

RemovedPW

type PW = (PWType, Pos)

Representação de um powerup.

type Bomb = (Pos, Int, Int, Int)

Representação de uma bomba.

data BlastType

Tipo de explosão (para depois as desenhar).

Constructors

CenterBlast

VerticalBlast

HorizontalBlast

UpperBlast

LowerBlast

RightBlast

LeftBlast

type Blast = (BlastType, Pos)

Representação de uma explosão

type Player = (Int, Pos, (Int, Int), Motion, Either (Key, Key, Key, Key, Key) ([String] -> Int -> Int -> Maybe Char))

A representação de um jogador.

data Estado

A representação do estado do jogo.

Constructors

E

Fields

mainState :: MainState

pics :: [Picture]

tileMap :: [[Tile]]

pws :: [PW]

bombs :: [Bomb]

players :: [Player]

blasts :: [Blast]

time :: Float

isRock :: Pos -> [[Tile]] -> Bool

Função que verifica se existe uma pedra numa dada posição.

isBrick :: Pos -> [[Tile]] -> Bool

Função que verifica se existe um tijolo numa dada posição.

isEmpty :: Pos -> [[Tile]] -> Bool

Função que verifica se existe um espaço vazio numa dada posição.