Tarefa5_update

Description

Documentation

convertMap :: [String] -> ([[Tile]], [PW])

Função que converte mapas gerados pela função mapa no módulo Bomberman em tuplos de mapas como representados na tarefa 5 e listas de powerups.

splitPWs :: [String] -> [PW]

Função que extrai os powerups da lista de String gerada na tarefa 1.

charToTile :: Char -> Tile

Função que converte um carater num Tile.

tileToChar :: Tile -> Char

Função que converte um Tile num carater.

pwToString :: [[Tile]] -> PW -> String

Função que contrói um String que representa um powerup para utilizar quando é chamada a função do bot. Powerups não visíveis não são incluidos para não dar uma vantagem aos bots.

bombToString :: Bomb -> String

Função que contrói um String que representa uma bomba.

playerToString :: Player -> String

Função que contrói um String que representa um jogador.

mkStringMap :: [[Tile]] -> [PW] -> [Bomb] -> [Player] -> [String]

Função que converte o mapa utilizado na tarefa 5 num mapa para ser passado como argumento das funções dos bots, não sendo incluídos os powerups não visíveis.

handleEvent :: Event -> Estado -> Estado

Função que altera o estado do jogo quando acontece um evento.

handleMenuEvents :: Event -> Estado -> Estado

Função que lida com os eventos quando se está no menu principal.

handleGameEvents :: Event -> Estado -> [Player] -> Estado

Função que lida com os eventos quando o jogo está a decorrer.

handleButtonDown :: Estado -> Key -> [Player] -> (Estado, [Player])

Função que lida com os eventos de teclas pressionadas.

countBombs :: [Bomb] -> Int -> Int

Função que conta as bombas postas por um dado jogador.

handleBotCmd :: Estado -> [Player] -> (Estado, [Player])

Função que lida com os comandos dados pelos bots.

handleTime :: Float -> Estado -> Estado

Função que altera o estado do jogo quando o tempo avança n segundos.

calcTicks :: Int -> Int

Função que, dada uma dimensão, calcula quantos ticks é preciso faltarem até o mapa começar a ser preenchido pelas pedras da espiral.

calcDim :: [[Tile]] -> Int

Função que calcula a dimensão de um mapa.

takeExploding :: [Bomb] -> ([Bomb], [Bomb])

Função que, dado um mapa, retorna um par de uma lista de bombas que explodem no instante atual, e a lista das restantes bombas.

tickBombs :: [Bomb] -> [Bomb]

Função que reduz o contador das bombas em um.

spiral :: ([[Tile]], [PW], [Bomb], [Player]) -> Int -> Int -> Pos -> (Int, Int) -> ([[Tile]], [PW], [Bomb], [Player])

Função que implementa a espiral.

reactToEdge :: [[Tile]] -> Pos -> (Int, Int) -> Int -> ((Int, Int), Int)

Função que muda a direção de um vetor quando a dada posição de aproxima um dado número de blocos da borda, e aumenta este último número sempre que muda de direção para cima. A funçäo determina em que direção a espiral deverá se mover no próximo tick.

explode :: [[Tile]] -> [PW] -> [Bomb] -> [Bomb] -> [Player] -> [Blast] -> ([[Tile]], [PW], [Bomb], [Player], [Blast])

Função que dado um mapa e uma lista de bombas retorna um outro mapa onde se sentem os efeitos das explosões das bombas.

explodeBomb :: [[Tile]] -> [PW] -> Bomb -> [Bomb] -> [Player] -> [Blast] -> Int -> ([[Tile]], [PW], [Bomb], [Player], [Blast])

Função que explode a bomba na sua própria posição, e nas 4 direções em que a bomba explode.

explodeBombAux :: Int -> Int -> Pos -> ([[Tile]], [PW], [Bomb], [Player], [Blast]) -> Pos -> ([[Tile]], [PW], [Bomb], [Player], [Blast])

Função que, dado o tamanho da explosão, o tamanho atual, a posição da bomba, um mapa, e a posição atual da explosão, retorna um mapa onde se sentem os efeitos da explosão da bomba apenas na direção atual.

calcBlast :: Pos -> Pos -> Bool -> BlastType

Função que determina o tipo de explosão numa dada posição.

calcNextPos :: Pos -> Pos -> Pos

Função que dadas duas posições alinhados pelo eixo horizontal ou vertical, retorna a próxima posição, ou seja, aquela imediatamente após a segunda no sentido oposto desta para a primeira.

>>> calcNextPos (1,1) (4,1) = (5,1)

inMap :: Int -> Pos -> Bool

Função que testa se uma posição é uma posição válida para um mapa de uma dada dimensão.

isBomb :: Pos -> [Bomb] -> Bool

Função que testa se existe uma bomba numa dada posição.

isPW :: Pos -> [PW] -> Bool

Função que testa se existe um powerup numa dada posição.

setPos :: [[Tile]] -> Pos -> Tile -> [[Tile]]

Função que altera um tipo bloco de uma dada posição para outro especificado no terceiro argumento.

setPosAux :: [Tile] -> Int -> Tile -> [Tile]

Função auxiliar de setPos.

triggerBomb :: [Bomb] -> Pos -> [Bomb]

Função que, caso exista, altera o contador de uma bomba numa dada posição para 1.

rmPW :: [PW] -> Pos -> [PW]

Função que marca um powerup numa dada posição como removido.

pickUpPWs :: [PW] -> Pos -> (Int, Int) -> (Int, Int)

Função que acrescenta a contagem de powerups apanhados de um jogador se existir um powerup na sua posição.

rmBomb :: [Bomb] -> Pos -> [Bomb]

Função que, se existir, remove a bomba numa dada posição.

killPlayers :: [Player] -> Pos -> [Player]

Função que retira os jogadores que se encontram numa dada posição.

removeExploded :: [[Tile]] -> [PW] -> ([[Tile]], [PW])

Função que remove os blocos e powerups marcados como para remover.

movePlayers :: [[Tile]] -> [PW] -> [Player] -> Float -> ([PW], [Player])

Função que move os jogadores.

stopPlayer :: Player -> Player

Função que para um jogador, tendo em conta a direção do seu movimento para depois desenhar corretamente o jogador.