Tarefa6_li1g031

Description

Um módulo contendo definições Haskell para o cálculo de funções que permitam dar um certo comando (Up, Down, Left, Right ou Bomb) para o jogador efetuar.

Documentation

tVazio :: Mapa -> Vazios

A função tVazio recebe um Mapa e produz a lista das coordenadas de pontos vazios.

Por exemplo:

>>> tVazio ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]
[(1,1),(2,1),(3,1),(4,1),(5,1),(1,2),(3,2),(5,2),(1,3),(4,3),(5,3),(1,4),(5,4),(1,5),(2,5),(3,5),(4,5),(5,5),(1,7),(3,7),(1,8),(3,8)]

coordLivres :: Mapa -> Vazios

A função coordLivres recebe um Mapa e produz a lista das coordenadas de pontos vazios sem as coordenadas atingidas por bombas ou com bombas.

Por exemplo:

>>> coordLivres ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]
[(1,1),(2,1),(3,1),(4,1),(5,1),(1,2),(3,2),(5,2),(1,3),(4,3),(5,3),(1,4),(5,4),(1,5),(2,5),(3,5),(4,5),(5,5)]

coordLivres2 :: Mapa -> Vazios

A função coordLivres2 recebe um Mapa e produz a lista das coordenadas de pontos vazios sem as coordenadas com bombas.

Por exemplo:

>>> coordLivres2 ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 6 5 0 1 1"]
[(1,1),(2,1),(3,1),(4,1),(5,1),(1,2),(3,2),(5,2),(1,3),(4,3),(5,3),(1,4),(5,4),(1,5),(2,5),(3,5),(4,5),(5,5)]

deletingLivres :: Raios -> Vazios -> Vazios

A função deletingLivres recebe as coordenadas que são atingidas por bombas e as coordenadas vazias e elimina as coordenadas atingidas por bombas da lista de coordenadas vazias.

Propriedade:

deletingLivres x [] = []

deletingLivres []x = x

Por exemplo:

>>> deletingLivres [(1,1),(2,1)] [(1,1),(2,1),(3,1),(4,1),(5,1),(1,2),(3,2),(5,2),(1,3),(4,3),(5,3),(1,4),(5,4),(1,5),(2,5),(3,5),(4,5),(5,5)]
[(3,1),(4,1),(5,1),(1,2),(3,2),(5,2),(1,3),(4,3),(5,3),(1,4),(5,4),(1,5),(2,5),(3,5),(4,5),(5,5)]

coordRaios :: Bombas -> Raios

A função coordRaios a lista das Bombas e dá as coordenadas que são atingidas pelo raio das bombas.

Propriedade:

coordRaios [] = []

Por exemplo:

>>> coordRaios [(1,2,2,1)]
[(2,1),(1,2),(3,2),(2,3)]

igual :: Vazios -> BombasXY -> Vazios

A função igual elimina da lista das coordenadas vazias, as coordenadas das bombas.

Propriedade:

igual [] _ = []

igual x [] = x

Por exemplo:

>>> igual [(1,1),(2,1),(2,2)] [(1,1)]
[(2,1),(1,2)]

tuplosJxy :: String -> CoordenadaJ

A função tuplosJxy recebe uma String e origina um par de coordenadas.

Por exemplo:

>>> tuplosJxy "0 1 1"
(1,1)

tuploJ2 :: [String] -> CoordenadaJ

A função tuplosJxy recebe uma lista de String e origina um par de coordenadas.

Por exemplo:

>>> tuplosJxy ["0","1","1"]
(1,1)

moveR :: Mapa -> CoordenadaJ -> Bool

A função moveR recebe um mapa e a coordenada do jogador e verifica se ele pode ir para a direita. Esta função tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveR ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
True

moveR2 :: Mapa -> CoordenadaJ -> Bool

A função moveR2 recebe um mapa e a coordenada do jogador e verifica se ele pode ir para a direita. Esta função 'não' tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveR2 ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
True

moveL :: Mapa -> CoordenadaJ -> Bool

A função moveL recebe um mapa e a coordenada do jogador e verifica se ele pode ir para a esquerda. Esta função tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveL ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
False

moveL2 :: Mapa -> CoordenadaJ -> Bool

A função moveL2 recebe um mapa e a coordenada do jogador e verifica se ele pode ir para a esquerda. Esta função 'não' tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveL2 ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
False

moveU :: Mapa -> CoordenadaJ -> Bool

A função moveU recebe um mapa e a coordenada do jogador e verifica se ele pode ir para cima. Esta função tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveU ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
False

moveU2 :: Mapa -> CoordenadaJ -> Bool

A função moveU2 recebe um mapa e a coordenada do jogador e verifica se ele pode ir para cima. Esta função 'não' tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveU2 ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
False

moveD :: Mapa -> CoordenadaJ -> Bool

A função moveD recebe um mapa e a coordenada do jogador e verifica se ele pode ir para baixo. Esta função tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveD ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
True

moveD2 :: Mapa -> CoordenadaJ -> Bool

A função moveD2 recebe um mapa e a coordenada do jogador e verifica se ele pode ir para baixo. Esta função 'não' tem em conta as coordenadas atingidas por bombas.

Por exemplo:

>>> moveD2 ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] (1,1)
True

estaNoRaio :: CoordenadaJ -> Bombas -> Mapa -> (Char, Bool)

A função estaNoRaio recebe uma coordenada de jogador, a lista das bombas, um mapa e verifica se esse jogador está no raio de uma bomba.

Propriedade:

estaNoRaio (x,y) [] _ = ('N',False)

Por exemplo:

>>> estaNoRaio (2,1) [(1,1,1,1)] ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 1 0 1","0 2 1"]
('L',True)

emCimaDaBomba :: CoordenadaJ -> Bombas -> Bool

A função emCimaDaBomba recebe uma coordenada de um jogador e uma lista de bombas e verifica se esse jogador está por cima de uma bomba.

Propriedade:

estaNoRaio (x,y) [] = False

Por exemplo:

>>> emCimaDaBomba (1,1) [(1,1,1,1)]
True

boom :: Mapa -> CoordenadaJ -> Bool

A função boom recebe um Mapa, a coordenada de um jogador e verifica se esse jogador tem uma caixa à sua volta.

Por exemplo:

>>> boom  ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 1 0 1","0 2 1"] (2,1)
False

tPlayers :: Mapa -> JogadoresXY

A função tPlayers recebe um mapa e produz a lista com as coordenadas dos jogadores presentes nesse mapa.

Propriedade:

tPlayers [] = []

Por exemplo:

>>> tPlayers  ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 1 0 1","0 2 1"]
[(2,1)]

boomJ :: Mapa -> CoordenadaJ -> Bool

A função boomJ recebe um Mapa, uma coordenada de um Jogador =Propriedade:

tPlayers [] = []

Por exemplo:

>>> tPlayers  ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 1 0 1","0 2 1"]
[(2,1)]

ver :: Mapa -> Int -> Bool

A função ver, pegando num Mapa e no número do jogador, vai verificar se esse jogador se encontra no mapa, devolvendo um boleano - True se existir no mapa e False se não existir.

Propriedade:

ver [] _ = False

Por exemplo:

>>> ver ["#####"] 0
False

>>> ver ["#####","0 1 1"] 0
True

temSaidaR :: (Int, Int) -> Mapa -> Bool

A função temSaidaR pega na coordenada e num mapa e verifica se o jogador "não se vai meter num buraco" se for para a direita. É uma função útil para o jogador não ficar entre uma caixa e uma bomba.

Por exemplo:

>>> temSaidaR (1,1) ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 2 1","* 1 1 0 1 1","* 1 3 0 1 1"] 100
True

temSaidaL :: (Int, Int) -> Mapa -> Bool

A função temSaidaL pega na coordenada e num mapa e verifica se o jogador "não se vai meter num buraco" se for para a esquerda. É uma função útil para o jogador não ficar entre uma caixa e uma bomba.

Por exemplo:

>>> temSaidaL (3,1) ["#######","#?    #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 2 1","* 1 1 0 1 1","* 1 3 0 1 1"] 100
False

temSaidaD :: (Int, Int) -> Mapa -> Bool

A função temSaidaD pega na coordenada e num mapa e verifica se o jogador "não se vai meter num buraco" se for para baixo. É uma função útil para o jogador não ficar entre uma caixa e uma bomba.

Por exemplo:

>>> temSaidaD (3,1) ["#######","#?    #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 2 1","* 1 1 0 1 1","* 1 3 0 1 1"] 100
False

temSaidaU :: (Int, Int) -> Mapa -> Bool

A função temSaidaU pega na coordenada e num mapa e verifica se o jogador "não se vai meter num buraco" se for para cima. É uma função útil para o jogador não ficar entre uma caixa e uma bomba.

Por exemplo:

>>> temSaidaU (3,3) ["#######","#  ?  #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 2 1","* 1 1 0 1 1","* 1 3 0 1 1"] 100
False

bot :: Mapa -> Int -> Int -> Maybe Char

A função bot é a principal função da Tarefa 6. Esta função recebe um Mapa, um Jogador e um Inteiro (Instante de Tempo) e devolve um Maybe Char que poderá ser um movimento para o jogador ou Nothing.

Por exemplo:

>>> bot ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 1 2","+ 3 4","* 1 1 0 1 1","0 2 1"] 0 100
Just 'R'