Tarefa6_li1g135

Description

Este módulo contém as funções, da Tarefa 6, para a criação de um Bot.

Documentation

type Flag = Int

Descrição:

A Flage, é utilizada, para controlar situações.

type Gamer = Int

Descrição:

O Gamer é o número de um jogador, portanto, 0,1,2 e 3.

type CoordX = Int

Descrição:

A CoordX é a coordenada no eixo do xx, que no caso do Bomberman, corresponde à Coluna.

type CoordY = Int

Descrição:

A CoordY é a coordenada no eixo do yy, que no caso do Bomberman, corresponde à Linha.

type Distance = Float

Descrição:

A Distance, é a distância entre dois Pontos.

data Dir

Descrição:

O tipo de dado Dir, representa todas as direções possíveis do Bot, sendo que o Null, significa, que não faz comando.

Constructors

Upp

Downn

Leftt

Rightt

Null

data WhereIsBomb

Descrição:

O tipo de dado WhereIsBomb, representa a localização de uma bomba.

Construtores:

O construtor LinhaE, significa que a bomba está na mesma Linha que o jogador e que se encontra à Esquerda.

O construtor LinhaD, significa que a bomba está na mesma Linha que o jogador e que se encontra à Direita.

O construtor ColunaC, significa que a bomba está na mesma Coluna que o jogador e que se encontra em Cima.

O construtor LinhaD, significa que a bomba está na mesma Coluna que o jogador e que se encontra em em Baixo.

O construtor Bombsport, significa que a bomba está na mesma posição que o jogador.

Constructors

LinhaE

LinhaD

ColunaC

ColunaB

Bombspot

Nada

bot :: [String] -> Int -> Int -> Maybe Char

Descrição:

A função bot, recebe o estado de jogo, o número do jogador e o tempo atual do jogo, e retorna o comando que o bot, determinou através da situação de Jogo.

Descrição do Bot:

Inicialmente o Bot, procura o jogador mais próximo e tenta matá-lo com bombas.

Caso existam power ups, destapados, ele vai tentar capturá-los.

Quando o efeito espiral estiver a começar o bot, foca uma posição, para se proteger da mesma.

Quando estiver na posição ideal, ainda tenta matar jogadores quando o tempo for igual a 3 segundos.

Este Bot, sabe escapar às suas bombas e de outros jogadores.

whatDir :: Posicao -> Posicao -> (Dir, Dir)

Descrição:

A função whatDir, determina a direção que o Bot, deve seguir para atingir um determinado Ponto. Retornando assim duas possíveis Direções.

Exemplos:

>>> whatDir (1,1) (3,3)
(Rightt,Downn)
>>> whatDir (1,1) (3,1)
(Rightt,Null)
>>> whatDir (1,1) (1,3)
(Null,Downn)
>>> whatDir (3,1) (1,2)
(Leftt,Downn)
>>> whatDir (3,1) (3,1)
(Null,Null)

Notas:

O primeiro elemento do par, contém as direções Rightt, Leftt e Null, o segundo tem Upp, Downn, Null.

Quando só existe uma direção possível ele retorna-a e no outro elemento do tuplo, ocorre Null.

Quando a posição é a mesma, retorna em ambos os elemnentos do tuplo Null.

whatComand :: Dir -> Maybe Char

Descrição:

A função whatComand, recebe uma direção, e retorna qual o comando correspondente.

Exemplos:

>>> whatComand Rightt
Just 'R'
>>> whatComand Leftt
Just 'L'
>>> whatComand Upp
Just 'U'
>>> whatComand Downn
Just 'D'

whatPos :: Dir -> Dir -> Posicao -> (Posicao, Posicao)

Descrição:

A função whatPos, e a respetiva auxiliar recebe duas direções e a posição, e determina, quais as coordenadas das direções, partindo da coordenada onde o Bot se encontra.

Exemplos:

>>> whatPos Leftt Downn (7,1)
((6,1),(7,2))
>>> whatPos Null Null (7,1)
((7,1),(7,1))
>>> whatPos Null Downn (7,1)
((7,1),(7,2))

whatIsPlayer :: Players -> Flag -> Gamer -> Posicao

Descrição:

A função whatIsPlayer, recebe a lista de jogadores, a flag inicializada a zero e o número do jogador, e retorna a posição, desse jogador, caso ele não exista, retorna a posição (0,0).

Exemplos:

>>> whatIsPlayer [Vivo (1,1) 0 0,Vivo (9,1) 0 1,Morto,Morto] 0 0
(1,1)
>>> whatIsPlayer [Vivo (1,1) 0 0,Vivo (9,1) 0 1,Morto,Morto] 0 3
(0,0)
>>> whatIsPlayer [Vivo (1,1) 0 0,Vivo (9,1) 0 1,Morto,Morto] 0 1
(9,1)

chooseMove :: InfGame -> Gamer -> Time -> Maybe Char

Descrição:

A função chooseMove, determina o movimento que o Bot, deve fazer, consoante a situação de jogo

playerPos :: Players -> Posicao -> Flag -> Gamer -> Bool

Descrição:

A função playerPos, recebe a lista de Jogadores, um posição, uma flag inicializada a zero e o número do jogador (Bot), e verifica se existe algum jogador na posição indicada, mas não verifica o nosso jogador (Bot).

Exemplos:

>>> playerPos [Vivo (1,1) 0 0,Vivo (9,1) 0 1,Morto,Morto] (1,1) 0 1
True
>>> playerPos [Vivo (1,1) 0 0,Vivo (9,1) 0 1,Morto,Morto] (9,1) 0 1
False

Notas:

Retorna True, quando existe algum jogador na posição indicada, que não é o nosso jogador, e False, caso contrário.

checkBomb :: [Bomba] -> Posicao -> WhereIsBomb

Descrição:

A função checkBomb, recebe a lista das Bombas, uma posição e verifica se na posição indicada, existe alguma bomba capaz de atingir com o flame.

Exemplos:

>>> checkBomb [((1,2),1,1,10)] (1,1)
ColunaB
>>> checkBomb [((1,1),1,1,10)] (1,1)
Bombspot
>>> [((1,1),1,1,10)] (2,1)
LinhaE

Notas:

Esta função retorna um tipo novo, que é o WhereIsBomb, o ColunaB, significa que a bomba esta na mesma Coluna que a posição indicada e que se encontra abaixo da posição.

O Bombspot, significa que a bomba, econtra-se na mesma posição que a indicada.

Pode-se consultar toda a informação sobre o tipo WhereIsBomb nesta Documentação.

mvToUpDown :: Posicao -> InfGame -> WhereIsBomb -> Dir

Descrição:

A função mvToUpDown, recebe a posição do Bot, a informação do jogo, e a localização da bomba, e determina qual o melhor comando para a fuga, caso os comandos principais Upp, Downn, não sejam válidos, a função determina a melhor alternativa.

mvToLeftRight :: Posicao -> InfGame -> WhereIsBomb -> Dir

Descrição:

A função mvToLeftRight, recebe a posição do Bot, a informação do jogo, e a localização da bomba, e determina qual o melhor comando para a fuga, caso os comandos principais Leftt, Rightt, não sejam válidos, a função determina a melhor alternativa.

mvToUpDownLeftRight :: Posicao -> InfGame -> Dir

Descrição:

A função mvToUpDownLeftRight, recebe a posição do jogador e a informação do jogo, e determina qual das quatro direções possíveis é melhor, visto que este caso é o mais complexo, necessita-se de verificar todas as direções e escolher a melhor.

goodWay :: Posicao -> NumbFlame -> InfGame -> Dir -> Bool

Descrição:

A função goodway, recebe a posição, o número de power ups flames, a informação do jogo e por fim a direção, esta função verfica se esta direção é a melhor para a fuga da bomba, pois necessita-se de verificar a existência de obstáculos ou de bombas.

whatCoords :: Dir -> Posicao -> Posicao

Descrição:

A função whatCoords, recebe a direção e uma posição e determina, qual a posição originada por essa direção.

checkFlmBomb :: [Bomba] -> Posicao -> NumbFlame

Descrição:

A função checkFlmBomb, recebe a lista de bombas e uma posição, e retorna o flame da bomba que se encontra nessa posição.

Exemplos:

>>> checkFlmBomb [((1,2),1,4,10)] (1,2)
4
>>> checkFlmBomb [((1,2),1,4,10),((3,4),0,2,10)] (3,4)
2

escapeBomb :: InfGame -> WhereIsBomb -> Posicao -> Maybe Char

Descrição:

A função escapeBomb, determina através dos construtores do tipo WhereIsBomb, qual o melhor comando para a fuga do Bot.

Exemplos:

>>> escapeBomb (filterInfGame ((mapa 11 0)++["* 1 1 1 1 10","0 1 1","1 9 1 !"]))  Bombspot (1,1)
Just 'R'
>>> escapeBomb (filterInfGame ((mapa 11 0)++["* 1 1 1 2 10","0 1 1","1 9 1 !"]))  ColunaC (1,2)
Just 'D'
>>> escapeBomb (filterInfGame ((mapa 11 0)++["* 1 1 1 1 10","0 1 1","1 9 1 !"]))  LinhaE (1,2)
Just 'D'

Notas:

A função filterInfGame está definida na Biblioteca T4Lib, converte a informação do mapa de [String] para o tipo InfGame.

A função mapa está definida na Biblioteca T1Lib, e gera mapas do jogo Bomberman.

escapeToPoint :: InfGame -> Posicao -> CoordX -> CoordY -> Maybe Char

Descrição:

A função escapeToPoint, recebe a informação do jogador, a posição do mesmo, e as coordenadas objectivo. A função determina qual o comando que o Bot, deve efetuar, para atingir a posição indicada.

Exemplos:

>>> escapeToPoint (filterInfGame ((mapa 11 0)++["* 1 1 1 2 10","0 1 1","1 9 1 !"])) (1,1) 3 3
Just 'R'
>>> escapeToPoint (filterInfGame ((mapa 11 0)++["* 1 1 1 2 10","0 1 1","1 9 1 !"])) (1,1) 1 2
Just 'D'
>>> escapeToPoint (filterInfGame ((mapa 11 0)++["* 1 1 1 2 10","0 1 1","1 9 1 !"])) (9,1) 1 2
Just 'L'

Notas:

A função filterInfGame está definida na Biblioteca T4Lib, converte a informação do mapa de [String] para o tipo InfGame.

A função mapa está definida na Biblioteca T1Lib, e gera mapas do jogo Bomberman.

checkMyRange :: Posicao -> [Bomba] -> Bool

Descrição:

A função checkMyRange, verifica se existe alguma bomba no range do Bot, que o possa atingir.

checkMyRange :: Posicao -> [Bomba] -> Bool
checkMyRange (c,l) listaBmb = dir1 == Nada && dir2 == Nada && dir3 == Nada && dir4 == Nada
                                where dir1 = checkBomb listaBmb (c,l-1)
                                      dir2 = checkBomb listaBmb (c,l+1)
                                      dir3 = checkBomb listaBmb (c-1,l)
                                      dir4 = checkBomb listaBmb (c+1,l)

Notas1:

Como podemos ver no código, ele averigua todas as direções do range do Bot, para verficar a existência de Bombas.

Exemplos:

>>> checkMyRange (1,1) [((3,1),1,1,10)]
False
>>> checkMyRange (1,1) [((4,1),1,1,10)]
True

Notas2:

Como os movimentos possíveis do Bot, são cima, baixo, esquerda e direita, a função verifica nessas posições, a existência de Bombas.

followPlayer :: InfGame -> Posicao -> CoordX -> CoordY -> Gamer -> Maybe Char

Descrição:

A função followPlayer, recebe a informação do jogo, a posição do Bot, as coordenadas objetivo e o jogador, e determina qual o comando para perseguir e tentar matar um jogador.

selectPlayer :: Players -> Posicao -> Distance -> Posicao -> Flag -> Gamer -> Posicao

Descrição:

A função selectPlayer, recebe a lista de jogadores, a posição do Bot, o valor da distância entre pontos, que é inicializado a 3*tamanho, uma posição inicializada a (0,0) (o objetivo desta é guardar a posição do jogador mais próximo), uma flag de controlo de jogadores, e por fim o número do jogador. Esta função seleciona o jogador mais próximo do Bot.

Exemplos:

>>> selectPlayer [Vivo (1,1) 0 0,Vivo (3,3) 0 1,Vivo (7,7) 0 0,Vivo (4,3) 0 0] (1,1) 27 (0,0) 0 0
(3,3)
>>> selectPlayer [Vivo (5,5) 0 0,Vivo (3,3) 0 1,Vivo (7,7) 0 0,Vivo (4,3) 0 0] (5,5) 27 (0,0) 0 0
(4,3)

Notas:

O método utilizado nesta função, foi a fórmula matemática da distância entre dois pontos.

O tamanho do mapa utilizado nos exemplos é 9.

selectPowerUp :: [PowUp] -> Posicao -> Distance -> Posicao -> Posicao

Descrição:

A função selectPowerUp, recebe a lista de power ups, a posição atual do Bot, a distância inicializada a zero, uma posição, para guardar a posição mais próxima, caso não exista power ups, retorna (0,0), pois esta foi inicializada. Esta função retorna a posição do power up mais próximo.

Exemplos:

>>> selectPowerUp [] (5,5) (27) (0,0)
(0,0)
>>> selectPowerUp [Bomb (3,1)] (5,5) (27) (0,0)
(3,1)
>>> selectPowerUp [Bomb (3,1), Flame (3,2)] (5,5) (27) (0,0)
(3,2)
>>> selectPowerUp [Bomb (3,1), Flame (3,2), Flame (5,6)] (5,5) (27) (0,0)
(5,6)

Notas:

Caso não existam power ups, a função retorna a posição (0,0).