Tarefa6_li1g166

Description

Descrption: funções que permitem a implementação da estratégia de jogo do nosso jogador Copyright: Ana Ribeiro Jéssica Lemos

Documentation

bot :: [String] -> Int -> Int -> Maybe Char

Esta função devolve os movimentos que o jogador deverá executar. Caso este permaneça no mesmo local e o local para onde pretende ir seja um tijolo, então este deverá colocar uma bomba. Na eventualidade, de este se encontrar no canto esquerdo (superior ou inferior) e se encontrar no raio de ação de uma bomba então este deve movimentar-se para a direita. Se este se encontrar no canto direito (superior ou inferior) e se encontrar no raio de ação de uma bomba então este deve movimentar-se para a esquerda. No caso de se encontrar acima da fila do meio do mapa e no raio de ação de uma bomba deve exercutar o comando U. Se se encontrar abaixo então movimenta-se baixo. Nos outros casos, o jogador executa os movimentos previstos pela função "just".

infBomb :: [String] -> Int -> Int -> [String]

Esta função dá uma lista com a informação de todas as bombas prestes a explodir sem o símbolo da bomba e o espaço entre este e o próximo algarismo. Inicialmente percorremos a lista até encontrar uma bomba nestas condições. Quando tal se sucede, retiramos através de "drop 2" o símbolo e o espaço referidos anteriormente à informação da bomba. Fazemos de seguida a recursiva para aplicar o mesmo processo às restantes bombas nestas condições.

cooryBomb :: [String] -> Int -> Int -> [Int]

A "cooryBomb" fornece uma lista com todas as coordenadas y de todas as bombas prestes a explodir. Assim, aplicamos a função map à "infBomb". Esta percorre todos os elementos da lista e aplica-lhes a função f, que isola o primeiro elemento resultante do break.

coorxBomb :: [String] -> Int -> Int -> [Int]

A "coorxBomb" fornece uma lista com todas as coordenadas x de todas as bombas prestes a explodir. Assim, aplicamos a função map à "infBomb". Esta percorre todos os elementos da lista e aplica-lhes a função f, que isola o primeiro elemento do tuplo resultante do break aplicado à remoção da coordenada y, através de um break, e do espaço que os separava, recorrendo ao drop.

convert :: [String] -> Int -> Int -> [Int]

A função multiplica todas as coordenadas x, das bombas prontas a explodir, pela dimensão do mapa. A função "dim" indica a dimensão do mapa recebido. Para tal calculamos o tamanho da cabeça da lista, que corresponde à primeira linha do mapa constituída apenas por pedras.

convertToNum :: [String] -> Int -> Int -> [Int]

Esta converte as coordenadas das bombas prestes a explodir num número do mapa. Para tal, recorremos à função auxiliar "aux" que junta cada coordenada y mais 1 com a correspondente coordenada x, que foi anteriormente multiplicada pela dimensão do mapa, na função "convert". Note-se que conseideramos cada ponto do mapa um número sendo que o primeiro ponto do mapa corresponde ao numero 1 e assim sucessivamente.

bombaJog :: [String] -> Int -> Int -> Bool

Esta função verifica se o jogador se encontra num local do mapa sob ação de uma bomba. Assim, temos de verificar se o seu número no mapa corresponde a um dos números da lista com os números do mapa em que a chama se pode propagar.

infOurJog :: [String] -> Int -> Int -> [String]

A função "infOurJog" devolve uma lista com a informação do nosso jogador. Para tal percorremos a informação do mapa até encontrarmos uma lista cujo primeiro elemento é o dígito igual ao do nosso jogador.

tuploCoord :: [String] -> Int -> Int -> ([Char], [Char])

A função devolve um tuplo cujo primeiro elemento é o identificador do jogador e o segundo contém as coordenadas do jogador.

tuploCoord2 :: [String] -> Int -> Int -> ([Char], [Char])

A "tuploCoord2" devolve um tuplo cujo primeiro elemento contém a coordenada y e o segundo contém a coordenada x.

coordy :: [String] -> Int -> Int -> Int

Esta devolve a coordenada y do nosso jogador. Para tal, é necessário converter o primeiro elemento do tuplo num algarismo.

coordx :: [String] -> Int -> Int -> Int

Esta devolve a coordenada x do nosso jogador. Para tal, é necessário converter o segundo elemento do tuplo num algarismo.

coorOurJogNum :: [String] -> Int -> Int -> Int

A função "coorOurJogNum" devolve as coordenadas do nosso jogador.

convertS :: [String] -> Int -> Int -> Char

A função devolve o símbolo da célula para a qual o jogador pretende ir.

mOurjog :: [String] -> Int -> Int -> Int

Esta função devolve o o número do mapa associado ao movimento do jogador. Se este pretender movimentar-se para cima então à sua coordenada retiramos a dimensão. Na possibilidade, de se movimentar para baixo então aumentamos à coordenada a dimensão. Enquanto que para a esquerda diminuímos um e para a esquerda diminuímos um. Noutros casos, a sua coordenada mantém-se a mesma.

d :: [String] -> Int -> Int -> Int

Esta função devolve o tamanho do mapa.

just :: [String] -> Int -> Int -> Maybe Char

Esta função devolve o o número do mapa associado ao movimento do jogador. Se este pretender movimentar-se para cima então à sua coordenada retiramos a dimensão. Na possibilidade, de se movimentar para baixo então aumentamos à coordenada a dimensão. Enquanto que para a esquerda diminuímos um e para a esquerda diminuímos um. Noutros casos, a sua coordenada mantém-se a mesma.

indicaR :: [String] -> Int -> Int -> [Int]

A função "indicaR" devolve uma lista com o número máximo que a chama da bomba pode alcançar para a direita. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo somar ao número do mapa onde se encontra a bomba ao raio correspondente.

restricaoR :: [String] -> Int -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaR". Uma vez que a dimensão máxima da chama é o limite máximo do mapa à direita, caso o número do mapa correspondente à chama da bomba pertença à linha seguinte, ou seja, maior que

indicaU :: [String] -> Int -> Int -> [Int]

A função "indicaU" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para baixo. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo subtrair ao número do mapa onde se encontra a bomba o resultado do produto da dimensão com o raio correspondente.

restricaoU :: [String] -> Int -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaU".

indicaD :: [String] -> Int -> Int -> [Int]

A função "indicaD" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para a cima. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo somar ao número do mapa onde se encontra a bomba o resultado do produto da dimensão com o raio correspondente

restricaoD :: [String] -> Int -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaL".

indicaL :: [String] -> Int -> Int -> [Int]

A função "indicaL" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para a esquerda. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo subtrair ao número do mapa onde se encontra a bomba ao raio correspondente.

restricaoL :: [String] -> Int -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaL".

numerosR :: [String] -> Int -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Assim, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir, pelo que geramos uma lista cujo primeiro elemento é o número da bomba mais um e o último é o número máxima da chama. Recorremos à recursiva para as restantes bombas prestes a explodir.

numerosU :: [String] -> Int -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Pelo que, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir. Para tal, recorremos a uma lista de compreensão, que cada elemento da lista será o número do mapa onde se encontra a bomba menos o produto da dimensão do mapa por um "a" (o que se deve ao facto de a chama se propagar para cima) tal que este "a" vai ser cada elemento de uma lista dos números de um até à divisão da diferença dos números do mapa onde se encontra a bomba e da chama máxima pela dimensão deste. É necessária a recursiva para que o mesmo seja feito para as restantes bombas prestes a explodir.

numerosL :: [String] -> Int -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Desta forma, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir, pelo que geramos uma lista cujo primeiro elemento é o número máximo da chama e o último é o número da bomba, sendo necessário inverter esta lista uma vez que a chama inicia no local do mapa onde se encontra a bomba. Recorremos à recursiva para as restantes bombas prestes a explodir.

numerosD :: [String] -> Int -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Pelo que, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir. Para tal, recorremos a uma lista de compreensão, que cada elemento da lista será o número do mapa onde se encontra a bomba mais o produto da dimensão do mapa por um "a" (o que se deve ao facto de a chama se propagar para baixo) tal que este "a" vai ser cada elemento de uma lista dos números de um até à divisão da diferença dos números do mapa onde se encontra a bomba e da chama máxima pela dimensão deste. É necessária a recursiva para que o mesmo seja feito para as restantes bombas prestes a explodir.

raio :: [String] -> Int -> Int -> [Int]

A função raio devolve uma lista com os raios das bombas prestes a explodir (ou seja, com tempo igual a 1). Começamos por percorrer a lista até encontrar uma bomba (*) e com tempo de explosão 1. Quando esta é encontrada vamos retirar-lhe a informação do seu raio e fazemos uma recursiva para verificar a existência de mais bombas nesta situação. Para tal, usamos a função break que permitiu separar o tempo de explosão da bomba da sua restante informação. E de seguida, pegamos no segundo elemento do tuplo ao qual retiramos o primeiro elemento visto que correspondia a um ' ' (contém a restante informação da bomba) e recolhemos o primeiro elemento do resultado da função break para obtermos o raio da bomba. Caso não exista nenhuma bombas prestes a explodir devolve uma lista com raio 0.

locaisComChama :: [String] -> Int -> Int -> [Int]

Esta função devolve uma lista com todos os pontos do mapa em que as chamas se poderão propagar. Ou seja, é uma lista com todos os números do mapa em que a chama se pode propagar, em todas as diferentes direções, e o local onde está a bomba.