Tarefa6_li1g168

Description

Documentation

type Bomb = (Pos, Int, Int)

Representação de uma bomba.

bot :: [String] -> Int -> Int -> Maybe Char

Função que dado um mapa, o número do jogador a controlar, e o número de ticks até ao final do jogo, retorna um comando para o jogador.

getMovable :: [String] -> Pos -> Pos -> [Pos] -> [Pos]

Função que lista todas as posições que o jogador consegue alcançar.

isMovable :: [String] -> Pos -> Bool

Função que testa se um jogador pode ocupar uma dada posição

getBombs :: [String] -> [Bomb]

Função que retorna uma lista de bombas dado um mapa.

getCmd :: [String] -> Pos -> Int -> [(Pos, Float)] -> [(Pos, Float)] -> Maybe Char

Função que dados um mapa, a posição do jogador, o número de ticks que faltam, e as duas listas de utilidade de posições (uma para o jogador se mover, outra para por bombas), determina o comando.

getBestPos :: (Pos, Float) -> [(Pos, Float)] -> (Pos, Float)

Função que determina a melhor posição de uma lista de posições e os seus respetivos valores de utilidade.

getMove :: [String] -> Pos -> Pos -> Int -> Maybe Char

Função que determina para onde o jogador se deve mover para chegar a uma dada posição.

getPathsTo :: [String] -> Pos -> Pos -> Int -> [Pos] -> [[Pos]]

Função que retorna uma lista de caminhos possíveis de uma posição para outra, representados por uma lista de posições.

getPlayerPos :: [String] -> Int -> Pos

Função que retorna a possição de um jogador.

>>> getPlayerPos ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 3 5 0 1 10","0 3 5 +","1 5 5"] 0
                 (3,5) 

bDistFactor :: Float

Fator de desisão de proximidade de bombas.

bEFactor :: Float

Fator de desisão de explosão de tijolos.

distFromCenterFactor :: Float

Fator de desisão de proximidade do centro.

distanceSquared :: Pos -> Pos -> Float

Função que calcula o quadrado da distância entre uma posição e outra.

distanceFromCenter :: Pos -> Pos -> Int -> Float

Função que calcula o valor de utilidade de uma posição tendo em conta apenas a sua distância à posição central.

calcCommonUtility :: [String] -> Int -> Int -> [Pos] -> [(Pos, Float)]

Função que calcula os valores de utilidade em comum de todas as posições para que o jogador se consegue mover.

calcMoveUtility :: [String] -> Pos -> Int -> [(Pos, Float)] -> [(Pos, Float)]

Função que calcula os valores de utilidade para o movimento do jogador para todas as posições para que se pode mover.

bombDistance :: [String] -> Pos -> Float

Função que calcula o valor de utilidade de uma dada posição tendo em conta apenas as bombas.

bombDistanceAux :: [String] -> Pos -> [Bomb] -> Float

Função auxiliar de bombDistance.

isBlocked :: [String] -> Pos -> Pos -> (Int, Int) -> Bool

Função que testa se uma explosão está bloqueada.

calcBombUtility :: [String] -> Int -> [(Pos, Float)] -> [(Pos, Float)]

Função que calcula os valores de utilidade para a introdução de bombas pelo jogador para todas as posições para que se pode mover.

bricksExploded :: [String] -> Int -> Pos -> Float

Função que calcula o número de tijolos que uma bomba explodiria se posto numa dada posição.

countBricks :: [String] -> Pos -> (Int, Int) -> Int -> Int -> Int

Função auxiliar de bricksExploded.

getExplosionSize :: [String] -> Int -> Int

Função que determina o tamanho da explosão de uma bomba posto por um dado jogador.