Main

Description

Um módulo contendo definições Haskell para o cálculo de funções que permitam concretizar um certo comando (Up, Down, Left, Right ou Bomb) num mapa do jogo Bomberman sob a forma de lista de strings, partindo de um mapa (lista de String), um jogador e um comando.

Documentation

tirarCoord :: [String] -> Int -> String

A função tirarCoord recebe uma lista de Strings ( neste caso em específico, o mapa do Bomberman) e o um inteiro ( o número do jogador) e devolve a coordenada do jogador pedido.

Propriedade:

tirarCoord [] _ = []

Por exemplo:

>>> tirarCoord ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 2 1", "1 7 7"] 0
"0 2 1"

>>> tirarCoord ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 2 1", "1 7 7"] 1
"1 7 7"

verificaChar :: String -> Bool

A função verificaChar recebe uma String e dá um Bool consoante encontrar ou não os Char + ou !.

Propriedade:

verificaChar [] = False

Por exemplo:

>>> verificaChar "0 2 2 +"
True
>>> verificaChar "0 2 2"
False

cChar :: String -> Int

A função cChar recebe uma string e conta o número de powerups (+ e !) nessa string.

Propriedade:

cChar [] = 0

Por exemplo:

>>> cChar "0 2 2"
0
>>> cChar "0 2 2 +++!"
4

coordPowers :: String -> (Int, Int)

A função coordPowers recebe uma string e devolve um tuplo com as coordenadas dessa string.

Por exemplo:

>>> coordPowers "0 2 2"
(2,2)

>>> coordPowers "1 3 3 +++!!!"
(3,3)

tuploPowers :: [Int] -> (Int, Int)

A função tuploPowers recebe uma lista de inteiros com 2 inteiros e dá um tuplo. Tem como objetivo transformar esses números em coordenadas.

Por exemplo:

>>> tuploPowers [2,2]
(2,2)

cMais :: String -> Int

A função cMais recebe uma string e conta o número de power-ups bombas nessa string.

cMais [] = 0

Por exemplo:

>>> cMais "0 1 1 +++!"
3

>>> cMais "0 1 1"
0

cFlames :: String -> Int

A função cFlames recebe uma string e conta o número de power-ups flames nessa string.

Propriedade:

cFlames [] = 0

Por exemplo:

>>> cFlames "0 1 1 +++!"
1

>>> cFlames "0 1 1"
0

tuploString :: Int -> (Int, Int) -> String

A função tuploString recebe um inteiro, o jogador, um par de coordenadas e transforma tudo na respetiva string.

Por exemplo:

>>> tuploString 0 (1,1)
"0 1 1"

tuploStringMaisRight :: Int -> (Int, Int) -> String

A função tuploStringMaisRight recebe um inteiro, o jogador, um par de coordenadas e transforma tudo na respetiva string adicionando +1 ao x.

Por exemplo:

>>> tuploStringMaisRight 0 (1,1)
"0 2 1"

andarR :: String -> Int -> Char -> String

A função andarR recebe uma String, um Int (o jogador), um Char (comando do movimento) e dá a nova String depois de efetuar o movimento.

Propriedade:

andarR [] j m = []

Por exemplo:

>>> andarR "0 2 2" 0 'R'
"0 3 2"

tuploStringMaisLeft :: Int -> (Int, Int) -> String

A função tuploStringMaisLeft recebe um inteiro, o jogador, um par de coordenadas e transforma tudo na respetiva string subtraindo -1 ao x caso o x > 1.

Por exemplo:

>>> tuploStringMaisLeft 0 (1,1)
"0 1 1"
>>> tuploStringMaisLeft 0 (2,2)
"0 1 2" 

andarL :: String -> Int -> Char -> String

A função andarL recebe uma String, um Int (o jogador), um Char (comando do movimento) e dá a nova String depois de efetuar o movimento.

Propriedade:

andarL [] j m = []

Por exemplo:

>>> andarL "0 2 2" 0 'L'
"0 1 2"

tuploStringMaisDown :: Int -> (Int, Int) -> String

A função tuploStringMaisDown recebe um inteiro, o jogador, um par de coordenadas e transforma tudo na respetiva string adicionando +1 ao y.

Por exemplo:

>>> tuploStringMaisDown 0 (1,1)
"0 1 2"

andarD :: String -> Int -> Char -> String

A função andarD recebe uma String, um Int (o jogador), um Char (comando do movimento) e dá a nova String depois de efetuar o movimento.

Propriedade:

andarD [] j m = []

Por exemplo:

>>> andarD "0 2 2" 0 'D'
"0 2 3"

tuploStringMaisUp :: Int -> (Int, Int) -> String

A função tuploStringMaisUp recebe um inteiro, o jogador, um par de coordenadas e transforma tudo na respetiva string subtraindo -1 ao y.

Por exemplo:

>>> tuploStringMaisUp 0 (1,1)
"0 1 2"

andarU :: String -> Int -> Char -> String

A função andarU recebe uma String, um Int (o jogador), um Char (comando do movimento) e dá a nova String depois de efetuar o movimento.

Propriedade:

andarU [] j m = []

Por exemplo:

>>> andarR "0 2 2" 0 'U'
"0 2 1"

andar :: [String] -> Int -> Char -> (Int, Int)

A função andar recebe uma lista de String (o mapa), um Int (o jogador), um Char (comando do movimento) e dá um par de coordenadas para onde o jogador se moveu

Propriedade:

andar [] j m = (0,0)

Por exemplo:

>>> andar ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 2 1", "1 7 7"] 0 'R'
(3,1)

tP :: [String] -> [(Int, Int)]

A função tP recebe uma lista de String (mapa) e dá uma lista com os pares ordenados de todos os cardinais e pontos de interrogação.

Por exemplo:

>>> tP ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 2 1", "1 7 7"]
[(0,0),(1,0),(2,0),(3,0),(4,0),(5,0),(6,0),(7,0),(8,0),(0,1),(8,1),(0,2),(2,2),(3,2),(4,2),(5,2),(6,2),(8,2),(0,3),(3,3),(6,3),(8,3),(0,4),(1,4),(2,4),(4,4),
(6,4),(7,4),(8,4),(0,5),(2,5),(5,5),(8,5),(0,6),(2,6),(3,6),(4,6),(5,6),(6,6),(8,6),(0,7),(3,7),(4,7),(8,7),(0,8),(1,8),(2,8),(3,8),(4,8),(5,8),(6,8),(7,8),(8,8)]

coordB :: [String] -> [(Int, Int)]

A função coordB recebe uma lista de String (mapa) e dá uma lista com todos os pares ordenados dos power-ups flames.

Propriedade:

coordB [] = []

Por exemplo:

>>> coordB ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 2 1", "* 1 1 0 1 10"]
[(5,2),(3,3)]

tuplosB :: String -> (Int, Int)

A função tuplosB recebe uma String e dá um par de coordenadas (Int,Int)

Por exemplo:

>>> tuplosB "+ 5 5"
(5,5)

tuploStringB :: (Int, Int) -> String

A função tuploStringB recebe um par ordenado de um power-up bomba e converte na respetiva string.

Por exemplo:

>>> tuploStringB (2,2)
"+ 2 2"

coordF :: [String] -> [(Int, Int)]

A função coordF recebe uma lista de String (mapa) e dá uma lista com todos os pares ordenados dos power-ups Flames.

Propriedade:

coordB [] = []

Por exemplo:

>>> coordB ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 1 1"]
[(5,5)]

tuplosF :: String -> (Int, Int)

A função tuplosF converte uma string de um power-up Flame no respetivo par de coordenadas.

Por exemplo:

>>> tuplosF "! 2 2"
(2,2)

tuploStringF :: (Int, Int) -> String

A função tuploStringF recebe um par de coordenadas de um power-up flame e converte-o na respetiva string.

Por exemplo:

>>> tuploStringF (5,5)
"! 5 5"

tuploDeixarBomba :: Int -> (Int, Int) -> String

A função tuploDeixarBomba recebe um Int (jogador) e um par de coordenadas onde o jogador quer deixar uma bomba e imprime uma String com a bomba, as suas coordenadas, o jogador, o raio da bomba e o tempo para a sua explosão.

Por exemplo:

>>> tuploDeixarBom 0 (1,1)
"* 1 1 0 1 10"

bIniciais :: [String] -> [(Int, Int)]

A função bIniciais recebe uma lista de String (mapa) e imprime uma lista com todos os pares ordenados das bombas iniciais.

Propriedade:

bIniciais [] = []

Por exemplo:

>>> bIniciais ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 1 1"]
[(1,1)]

listajogadores :: [String] -> [String]

A função listajogadores recebe uma lista de String (mapa) e imprime uma lista com todos os pares ordenados dos jogadores iniciais.

Propriedade:

listajogadores [] = []

Por exemplo:

>>> listajogadores ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"]
["0 5 5 +!"]

listapowerups :: [String] -> [String]

A função listajogadores recebe uma lista de String (mapa) e imprime uma lista de strings com todos os power-ups do mapa.

Propriedade:

listapowerups [] = []

Por exemplo:

>>> listapowerups ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"]
["+ 5 2", "+ 3 3","! 5 5"]

cardinais :: [String] -> [String]

A função listajogadores recebe uma lista de String (mapa) e imprime uma lista de String com todos as pedras do mapa.

Propriedade:

cardinais [] = []

Por exemplo:

>>> cardinais ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]

bombas :: [String] -> [String]

A função bombas recebe uma lista de String (mapa) e imprime uma lista de String com todas as bombas desse mapa.

Propriedade:

bombas [] = []

Por exemplo:

>>> bombas ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]
["* 1 1 0 1 10"]

meterBomba :: [String] -> String -> Bool

A função meterBomba recebe uma lista de String (mapa) e uma string da nova bomba e verifica se essa bomba pode ser colocada ou não.

Propriedade:

meterBomba [] _ = True

Por exemplo:

>>> meterBomba ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"] "* 1 1 0 1 10"
False

>>> meterBomba ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "0 5 5 +!"] "* 1  0 1 10"
True

contarB :: String -> Int

A função contarB recebe uma String e conta o número de + (power-up Bombas) nessa string.

Propriedade:

contarB [] = 0

Por exemplo:

>>> contarB "0 2 2 ++++"
4

contarF :: String -> Int

A função contarB recebe uma String e conta o número de + (power-up Bombas) nessa string.

Propriedade:

contarB [] = 0

Por exemplo:

>>> contarB "0 2 2 ++++"
4

ifFlames :: [String] -> Int -> Bool

A função ifFlames recebe uma lista de Strings (mapa) e um jogador e verifica se esse jogador tem power-ups flames.

Propriedade:

ifFlames [] j = False

Por exemplo:

>>> ifFlames ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"] 0
  True

tuploDeixarBombaF :: Int -> (Int, Int) -> Int -> String

A função tuploDeixarBombaF recebe um Int (jogador), um par de coordenadas onde o jogador quer deixar uma bomba e o número de power-up Flames desse jogador e imprime uma String com a bomba, as suas coordenadas, o jogador, o raio da bomba e o tempo para a sua explosão.

Por exemplo:

>>> tuploDeixarBom 0 (1,1) 2
"* 1 1 0 3 10"

bIniciaisJ :: [String] -> Int -> Int

A função bIniciaisJ recebe uma lista de String (mapa) e um jogador e conta as bombas que esse jogador já tem no mapa.

Propriedade:

bIniciais [] j = 0

Por exemplo:

>>> bIniciais ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "0 5 5 +!"] 0
1

bombasJogador :: String -> Int -> Bool

A função bombasJogador recebe uma string e um Int (jogador) e verifica se essa bomba é do jogador j.

Propriedade;

bombasJogador [] j = False

Por exemplo:

>>> bombasJogador "* 1 1 0 1 10" 0
True

bombasJogadorMapa :: [String] -> Int -> Int

A função bombasJogadorMapa recebe uma lista de String (mapa) e um Int (jogador) e conta no número de bombas que esse jogador tem no mapa.

Propriedade:

bombasJogadorMapa [] j = 0

Por exemplo:

>>> bombasJogadorMapa ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "* 1 2 0 1 10" "0 5 5 +!"] 0
2

verificaJogador :: Int -> [String] -> Bool

A função verificaJogador recebe um Int (jogador) e uma lista de String (mapa) e verifica se esse jogador está presente no jogo.

Propriedade:

verificaJogador _ [] = False

Por exemplo:

>>> verificaJogador 2 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "* 1 2 0 1 10" "0 5 5 +!"]
False

move :: [String] -> Int -> Char -> [String]

A função move é a função principal desta tarefa e recebe uma lista de String (mapa), um Int (jogador) e um Char (comando a dar ao jogador). Imprime o novo mapa depois de efetuar esse comando.

Propriedade:

move [] j m = []

Por exemplo:

>>> move ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "* 1 2 0 1 10" "0 5 5 +!"] 0 'R'
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10", "* 1 2 0 1 10" "0 6 5 +!"]

main :: IO ()

O programa main foi-nos facultado para a realização deste projeto e pode ser usado para testar a função move, por nós definida, ou seja, após compilado, o programa aceita como parâmetros o mapa, o jogador e o comando m e depois invoca a função ​move​ imprimindo o resultado no stdout.