Main

Description

Um módulo contendo definições Haskell para o cálculo de funções que permitam transformar os mapas do jogo Bomberman em gráficos que nos permitam realizar ações, isto é, que nos permitam jogar. Para isso, foi utilizada a biblioteca Gloss.

Documentation

type Estado = ([String], Bool, Bool, Bool, Picture, Picture, Picture, Int, Picture, Picture, Int, Picture, Picture, Picture, [Int], Float, Picture, Char, Char, Char, Char)

Estado é um tipo de dados por nós definido, que consiste num tuplo constituído por uma lista de strings (mapa do jogo), por três boleanos (que nos dão informações sobre o estado de jogo, isto é, o primeiro diz-nos se o jogo está no ecrã inicial ou não, o segundo diz-nos se está no ecrã de selecionar o número de jogadores e o terceiro diz-nos se está no ecrã de seleção da dificuldade), por um inteiro (semente do mapa, que será gerada aleatoriamente), pelas imagens em formato bitmap que serão utilizadas no jogo (tijolo, caixa, power-ups, bomba e explosão), por um segundo inteiro (número de jogadores escolhido), por uma lista de inteiros (lista de jogadores que estão vivos), por um Float (número de segundos que passaram desde o ínicio do jogo) e por quatro caracteres (posição em que se encontra o boneco no jogo).

estadoInicial :: Picture -> Picture -> Picture -> Int -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Estado

A função estadoInicial vai receber todas as imagens em formato bitmap utilizadas no jogo e a semente e vai devolver um Estado correspondente ao ecrã inicial do jogo.

Por exemplo:

>>> estadoInicial p n d 1 t c f b bomba exp
([],True,False,False,p,n,d,1,t,c,0,f,b,bomba,[0,1,2,3],0,exp,'f','f','f','f')

escolherNumero :: Picture -> Picture -> Picture -> Int -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Estado

A função escolherNumero vai receber todas as imagens em formato bitmap utilizadas no jogo e a semente e vai devolver um Estado correspondente ao ecrã de seleção do número de jogadores.

Por exemplo:

>>> escolherNumero p n d 1 t c f b bomba exp
([],False,True,False,p,n,d,1,t,c,0,f,b,bomba,[0,1,2,3],0,exp,'f','f','f','f')

escolherDificuldade :: Picture -> Picture -> Picture -> Int -> Picture -> Picture -> Int -> Picture -> Picture -> Picture -> Picture -> Estado

A função escolherDificuldade vai receber todas as imagens em formato bitmap utilizadas no jogo, a semente e o número de jogadores escolhido e vai devolver um Estado correspondente ao ecrã de seleção do número de jogadores.

Por exemplo:

>>> escolherDificuldade p n d 1 t c 3 f b bomba exp
([],False,False,True,p,n,d,1,t,c,3,f,b,bomba,[0,1,2,3],0,exp,'f','f','f','f')

tuploGloss :: [String] -> (Float, Float)

A função tuploGloss recebe uma lista com duas strings e vai devolver um tuplo constituído pelos números representados nessas strings.

Por exemplo:

>>> tuploGloss ["1","10"]
(1,10)

tuplosGloss :: String -> (Float, Float)

A função tuplosGloss recebe uma string com as informações do jogo relativas aos jogadores ou aos power-ups e devolve um tuplo com as suas coordenadas, recorrendo às funções getCoord e tuploGloss.

Por exemplo:

>>> tuplosGloss "0 1 10"
(1,10)

>>> tuplosGloss "+ 1 1"
(1,1)

coordTudo :: [String] -> Char -> [(Float, Float)]

A função coordTudo recebe um estado de jogo e um caracter e devolve uma lista de tuplos com as coordenadas a que o caracter diz respeito, isto é, + corresponde aos power-ups Bombs, ! aos power-ups Flames,* às bombas, '0' ao jogador 0, '1' ao jogador 1, etc.

Propriedade:

coordTudo [] _ = []

Por exemplo:

>>> coordTudo ["#####","#   #","# # #","#   #","#####","0 1 1","1 2 1"] '0'
[(1.0,1.0)]

>>> coordTudo ["#####","#   #","# # #","#   #","#####","0 1 1","1 2 1"] '1'
[(2.0,1.0)]

porJogador :: [(Float, Float)] -> Char -> Int -> Float -> Char -> Picture

A função porJogador recebe a lista de coordenadas de um determinado jogador, definido pelo caracter inserido, a dimensão do mapa, o número de segundos que passaram desde o início do jogo e a posição do boneco e vai colocar o boneco desse jogador nos gráficos do mapa.

Propriedade:

porJogador [] _ _ _ _ = Blank

porPU :: [(Float, Float)] -> Int -> Picture -> Picture

A função porPU vai funcionar da mesma maneira que a porJogador, mas em vez de colocar os jogadores nos gráficos, vai colocar os power-ups, recebendo para isso as coordenadas dos power-ups, a dimensão do mapa e a imagem que lhe corresponde.

Propriedade:

porPU [] _ _ = Blank

genLinha :: String -> Int -> Float -> Float -> Picture -> Picture -> Picture

A função genLinha, recebendo uma linha do mapa, a sua dimensão, a posição de cada caracter na string, o número da linha e as imagens correspondentes aos tijolos e às pedras, vai devolver os gráficos que dizem respeito a essa linha do mapa.

Propriedade:

genLinha [] _ _ _ _ _ = Blank

genMapa :: [String] -> Int -> Float -> Picture -> Picture -> Picture

A função genMapa, recebendo um mapa, a sua dimensão, o número da linha e as imagens que correspondem aos tijolos e às caixas, vai aplicar a genLinha a todo o mapa.

Propriedade:

genMapa [] _ _ _ _ = Blank

corpo :: Int -> Char -> Picture

A função corpo recebe o número correspondente a cada jogador (0,1,2 ou 3) e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o corpo do boneco.

cabeca :: Char -> Picture

A função cabeca recebe a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o cabeça do boneco.

olhoE :: Picture

A função olhoE vai imprimir no ecrã de jogo o olho esquerdo do boneco.

olhoD :: Picture

A função olhoD vai imprimir no ecrã de jogo o olho direito do boneco.

boca :: Picture

A função boca vai imprimir no ecrã de jogo a boca do boneco.

cabelo :: Int -> Char -> Picture

A função cabelo recebe o número correspondente a cada jogador (0,1,2 ou 3) e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o cabelo do boneco.

bigode :: Int -> Picture

A função bigode recebe o número correspondente a cada jogador (0,1,2 ou 3) e vai imprimir no ecrã de jogo o corpo do boneco.

sobrancelhaE :: Picture

A função sobrancelhaE vai imprimir no ecrã de jogo a sobrancelha esquerda do boneco.

sobrancelhaD :: Picture

A função sobrancelhaD vai imprimir no ecrã de jogo a sobrancelha direita do boneco.

bracoE :: Float -> Char -> Picture

A função bracoE recebe o número de segundos que passaram desde o início do jogo e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o braço esquerdo do boneco.

bracoD :: Float -> Picture

A função bracoD recebe o número de segundos que passaram desde o início do jogo e vai imprimir no ecrã de jogo o braço direito do boneco.

pernaE :: Int -> Char -> Picture

A função pernaE recebe o número correspondente a cada jogador (0,1,2 ou 3) e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo a perna esquerda do boneco.

pernaD :: Int -> Picture

A função pernaD recebe o número correspondente a cada jogador (0,1,2 ou 3) e vai imprimir no ecrã de jogo a perna direita do boneco.

chapeu :: Int -> Char -> Picture

A função chapeu recebe o número correspondente a cada jogador (0,1,2 ou 3) e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o chapéu do boneco.

cor :: Int -> Color

A função cor recebe o número correspondente a cada jogador (0,1,2 ou 3) e vai devolver a cor de roupa que corresponde a cada boneco.

boneco :: Int -> Float -> Char -> Picture

A função boneco recebe o número correspondente a cada jogador (0,1,2 ou 3), o número de segundos que passaram desde o início do jogo e a posição em que o boneco se encontra (em forma de caracter: f = frente, t = trás e l = lado) e vai imprimir no ecrã de jogo o boneco completo.

tempo :: [String] -> Float -> Int

A função tempo, recebendo um mapa e o número de segundos que passaram desde o início do jogo, vai devolver o número de segundos que faltam para acabar o jogo.

vivo :: [Int] -> Int -> (Float, Float) -> (Float, Float) -> (Float, Float) -> Picture

A função vivo recebe a lista de jogadores que se encontram vivos, o número correspondente a um jogador e três coordenadas que irão ser utilizadas para mover as imagens dentro do painel lateral de forma a ficar tudo alinhado, devolvendo uma imagem que nos diz se o jogador selecionado está vivo ou morto no jogo.

Propriedade:

vivo [] _ _ _ _ = Blank

vivoBot :: [Int] -> Int -> (Float, Float) -> (Float, Float) -> (Float, Float) -> Picture

A função vivoBot irá funcionar da mesma maneira que a vivo, mas desta vez aplicando as informações aos jogadores que são bots.

Propriedade:

vivoBot [] _ _ _ _ = Blank

puBApanhados :: String -> (Float, Float) -> Picture -> Picture

A função puBApanhados, recebendo uma linha do mapa com as informações de um determinado jogador, um par de coordenadas e a imagem correspondete ao power-up Bombs, vai devolver uma imagem que lista o número de power-ups Bombs que esse jogador apanhou.

Propriedade:

puBApanhados [] _ _ = Blank

puFApanhados :: String -> (Float, Float) -> Picture -> Picture

A função puFApanhados vai atuar da mesma maneira que a puBApanhados, mas aplicando-se desta vez aos power-up Flames.

Propriedade:

puFApanhados [] _ _ = Blank

painelLateral :: Estado -> Picture

A função painelLateral, recebendo um Estado de jogo, vai devolver uma imagem com todas as informações adicionais do jogo, localizando-se no lado esquerdo do ecrã de jogo. Estas informações consistem no tempo (em segundos) que falta para o jogo terminar, nos jogadores que estão vivos os mortos, se são bots ou não e quais os power-ups que apanharam.

verRaioBomba :: [String] -> [Float]

A função verRaioBomba, recebendo um estado de jogo, vai devolver uma lista com os raios de todas as bombas colocadas no mapa.

Propriedade:

verRaioBomba [] = []

Por exemplo:

>>> verRaioBomba ["#####","#   #","# # #","#   #","#####","* 1 1 0 1 10","0 1 1"]
[1.0]

>>> verRaioBomba ["#####","#   #","# # #","#   #","#####","* 1 1 0 1 10","* 1 2 1 1 10","0 1 1","1 1 2"]
[1.0,1.0]

raios :: Float -> [Float]

A função raios, recebendo um raio, vai devolver uma lista com os números de 0 até esse raio.

Propriedade:

raios 0 [] = []

Por exemplo:

>>> raios 1
[0.0,1.0]

>>> raios 5
[0.0,1.0,2.0,3.0,4.0,5.0]

verTempoBomba :: [String] -> [Float]

A função verTempoBomba, recebendo um estado de jogo, vai devolver uma lista com os tempos de todas as bombas colocadas no mapa.

Propriedade:

verTempoBomba [] = []

Por exemplo:

>>> verTempoBomba ["#####","#   #","# # #","#   #","#####","* 1 1 0 1 10","0 1 1"]
[10.0]

>>> verTempoBomba ["#####","#   #","# # #","#   #","#####","* 1 1 0 1 9","* 1 2 1 1 10","0 1 1","1 1 2"]
[9.0,10.0]

explosaoCima :: (Float, Float) -> Int -> [Float] -> [(Float, Float)] -> [(Float, Float)] -> [(Float, Float)] -> Picture -> Picture

A função explosaoCima, recebendo as coordenadas de uma bomba, a dimensão do mapa, a lista com os números de 0 até ao seu raio (dada pela raios), a lista das coordenadas dos tijolos, a lista das coordenadas das caixas, a lista das coordenadas dos power-ups e a imagem que corresponde à explosão, vai imprimir no ecrã a explosão da bomba para cima, isto é, nas coordenadas de ordenada superior àquela onde a bomba foi colocada.

Propriedade:

explosaoCima _ _ [] _ _ _ _ = Blank

explosaoBaixo :: (Float, Float) -> Int -> [Float] -> [(Float, Float)] -> [(Float, Float)] -> [(Float, Float)] -> Picture -> Picture

A função explosaoBaixo vai atuar da mesma maneira que a explosaoCima, mas imprimindo a explosão para baixo.

Propriedade:

explosaoBaixo _ _ [] _ _ _ _ = Blank

explosaoDta :: (Float, Float) -> Int -> [Float] -> [(Float, Float)] -> [(Float, Float)] -> [(Float, Float)] -> Picture -> Picture

A função explosaoDta vai atuar da mesma maneira que a explosaoCima, mas imprimindo a explosão para a direita.

Propriedade:

explosaoDta _ _ [] _ _ _ _ = Blank

explosaoEsq :: (Float, Float) -> Int -> [Float] -> [(Float, Float)] -> [(Float, Float)] -> [(Float, Float)] -> Picture -> Picture

A função explosaoEsq vai atuar da mesma maneira que a explosaoCima, mas imprimindo a explosão para a esquerda.

Propriedade:

explosaoEsq _ _ [] _ _ _ _ = Blank

porBomba :: [String] -> [(Float, Float)] -> Int -> [Float] -> [Float] -> Picture -> Picture -> Picture

A função porBomba vai receber um estado de jogo, a lista com as coordenadas das bombas, a dimensão do mapa, a lista dos raios das bombas, a lista dos tempos das bombas e as imagens correspondentes às bombas e às explosões, colocando as bombas, e respetivas explosões, nos gráficos do jogo.

Propriedade:

porBomba _ [] _ _ _ _ _ = Blank

nJogadores :: [String] -> [Int]

A função nJogadores vai receber um estado de jogo e devolver a lista dos jogadores que estão ainda vivos.

Propriedade:

nJogadores [] = []

Por exemplo:

>>> nJogadores ["#####","#   #","# # #","#   #","#####","0 1 1"]
[0]

>>> nJogadores ["#####","#   #","# # #","#   #","#####","0 1 1","1 1 2","3 2 1"]
[0,1,3]

desenha :: Estado -> Picture

A função desenha, recebendo um Estado do jogo, vai imprimir no ecrã a imagem correspondente a esse estado.

mapaInicial :: Int -> Int -> [String]

A função mapaInicial, recebendo a dimensão do mapa desejado e a semente, vai devolver o mapa gerado pela função com o mesmo definida na Tarefa 1, acrescentando-lhe as coordenadas iniciais dos quatro jogadores.

Por exemplo:

>>> mapaInicial 7 0
["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 1 1","1 5 5","2 1 5","3 5 1"]

>>> mapaInicial 19 12
["###################","#    ?      ???   #","# # # # #?# # #?# #","#    ???   ?  ??? #","#?#?#?# # #?# #?#?#","#?    ?   ?   ?? ?#","# #?# #?# #?#?# # #","#  ???    ?? ??  ?#","#?# # # # # # #?# #","#  ??? ?  ?    ?  #","#?#?#?# #?#?#?# # #","#??   ?   ?  ??   #","#?#?# #?# #?#?#?#?#","#    ? ???      ? #","#?# #?# #?# #?# # #","#    ?        ?  ?#","# #?#?#?# # # #?# #","#   ????  ?   ??  #","###################","+ 11 4","+ 13 6","+ 11 7","+ 5 9","+ 10 11","+ 10 17","! 3 6","! 7 9","! 14 15","! 5 16","0 1 1","1 17 17","2 1 17","3 17 1"]

continua :: Event -> Estado -> Estado

A função desenha, recebendo um evento e um Estado do jogo, vai devolver o Estado que sucede à transformação do anterior acionada pelo evento.

numeroBot :: Int -> Float -> Int

A função numeroBot, recebendo o número de jogadores que estão a jogar (selecionado num dos ecrãs anteriores ao jogo em si) e o número de segundos que passaram desde o início do jogo, vai devolver o número correspondente a um bot, de forma a fazer com que a cada segundo se mova um bot diferente.

mexerBot :: [String] -> Int -> [Int] -> Float -> [String]

A função mexerBot vai receber um estado de jogo, o número de jogadores e o número de segundos que passaram desde o início do jogo e vai devolver um novo estado de jogo após o movimento (ou não movimento) de um bot.

mudarLado :: Int -> Float -> [String] -> Char -> Char

A função mudarLado vai receber o número de jogadores selecionado no menu, o número de segundos que passaram desde o início do jogo, o mapa e o caracter que diz respeito à posiçao do boneco de um bot e vai devolver um caracter que define a nova posição do boneco desse mesmo bot, de forma a fazê-lo virar-se.

Por exemplo:

>>> mudarLado 2 15 ["#######","#     #","# # # #","#     #","#######","0 1 1","1 5 5","2 1 5","3 5 1"] 'f'
'f'

reageTempo :: Float -> Estado -> Estado

A função reageTempo, recebendo o frame rate e um Estado do jogo, vai devolver o Estado alterado pela passagem do tempo.

window :: Display

A função window define o nome da janela do jogo, a sua dimensão e a sua posição no ecrã.

main :: IO ()

A função main vai imprimir no ecrã os gráficos completos do jogo Bomberman, utilizando as funções por nós definidas e recorrendo às funções da biblioteca Gloss. O frame rate por nós definido é igual a 5.