Main

Description

Um módulo contendo definições Haskell para o cálculo de funções que permitem produzir efeitos da passagem de tempo num estado de jogo. partindo de um Mapa e de um Instante de Tempo, originando o novo Mapa.

Documentation

instanteBomba :: Mapa -> [String]

A função instanteBomba recebe o Mapa e produz o efeito da passagem de tempo nas bombas originando uma lista de strings com as novas bombas e sem bombas que explodem.

Propriedade:

instanteBomba []  = []

Por exemplo:

>>> tirarCoord ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 10"]
"* 1 1 0 1 9"

>>> tirarCoord ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 1"]
[]

explodePB :: CoordenadaBomba -> Bombas -> Mapa -> Bombas

A função explodePB recebe uma Coordenada de uma Bomba, Bombas e um Mapa e produz uma lista de Coordenadas com as bombas que são atingidas pela CoordenadaBomba

Propriedade:

explodePB (r,x,y,t) [] _ = []

Por exemplo:

>>> explodePB (1,1,1,1) [(1,2,1,10)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 1", "* 2 1 0 1 10"]
[(1,2,1,10)]

menosUm :: String -> String

A função 'menosUm" recebe uma String de uma Bomba e altera o tempo dessa bomba, subtraindo 1.

Por exemplo:

>>> menosUm "* 1 1 0 1 10"
"* 1 1 0 1 9"

tuplosBr :: String -> (Int, Int, Int, Int)

A função 'tuplosBr" recebe uma String de uma Bomba e dá um tuplo de coordenadas do género (raio,x,y,tempo).

Por exemplo:

>>> tuplosBr "* 1 1 0 1 10"
"* 1 1 0 1 9"

tuploRaio2 :: [String] -> (Int, Int, Int, Int)

A função tuploRaio2 é a função auxiliar de tuplosBr que lê os elementos de uma lista de Strings e produz o tuplo (raio,x,y,tempo)

Por exemplo:

>>> tuploRaio2 ["2","2","0","1","10"]
(1,2,2,10)

tempo1 :: String -> String

A função 'tempo1" recebe uma String de uma Bomba e dá uma String da mesma Bomba com o tempo igual a 1.

Por exemplo:

>>> tempo1 "* 1 1 0 1 10"
"* 1 1 0 1 1"

tJ :: Mapa -> Jogadores

A função tJ recebe um Mapa e produz a lista com as coordenadas dos jogadores e a sua identificação.

Propriedade:

tJ [] = []

Por exemplo:

>>> tJ ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","0 1 1"]
[(0,1,1)]

tuploRaio :: [String] -> CoordenadaBomba

A função tuploRaio é a função auxiliar de tuplosBr que lê os elementos de uma lista de Strings e produz o tuplo (raio,x,y,tempo)

Por exemplo:

>>> tuploRaio ["2","2","0","1","10"]
(1,2,2,10)

tB :: Mapa -> Bombas

A função tB recebe um Mapa e produz a lista com as coordenadas das bombas do género (raio,x,y,tempo).

Propriedade:

tB [] = []

Por exemplo:

>>> tB ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 10"]
[(1,1,1,10)]

arrebentaBonecos :: Mapa -> NJogadores -> Mapa

A função arrebentaBonecos recebe um Mapa, uma lista com os jogadores e produz o novo mapa sem os jogadores da lista.

Propriedade:

arrebentaBonecos [] = []

arrebentaBonecos mapa [] = mapa

Por exemplo:

>>> arrebentaBonecos ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","0 1 1", "1 2 1"] [0]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","1 2 1"]

listajogadores3 :: Mapa -> NJogadores

A função listajogadores3 recebe um Mapa e produz a lista dos jogadores existentes no estado de jogo.

Propriedade:

listajogadores3 [] = []

Por exemplo:

>>> listajogadores3 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","0 1 1", "1 2 1"] [0]
[0,1]

explodeJ :: Jogadores -> Bombas -> Mapa -> NJogadores

A função explodeJ recebe os Jogadores, as Bombas e um Mapa e produz uma lista com os jogadores que são atingidos pelas Bombas.

Propriedade:

explodeJ jogadores [] _ = []

explodeJ [] _ _ = []

Por exemplo:

>>> explodeJ (1,1,1,1) [(1,2,1,10)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 1", "* 2 1 0 1 10"]
[(1,2,1,10)]

tPSemLimites :: PedrasEPI -> Mapa -> PedrasEPI

A função tPSemLimites recebe as coordenadas das Pedras e dos Pontos de Interrogação, o Mapa e produz as coordenadas das Pedras e PI sem as Pedras Limitantes de um Mapa.

Propriedade:

tPSemLimites [] _  = []

Por exemplo:

>>> tPSemLimites [(0,0),(1,0),(2,0),(3,0),(4,0),(5,0),(6,0),(0,1),(6,1),(0,2),(2,2),(4,2),(6,2),(0,3),(2,3),(3,3),(6,3),(0,4),(2,4),(3,4),(4,4),(6,4),(0,5),(6,5),(0,6),(1,6),(2,6),(3,6),(4,6),(5,6),(6,6)] ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]
[(2,2),(4,2),(2,3),(3,3),(2,4),(3,4),(4,4)]

esqPedra :: CoordenadaJ -> PedrasEPI -> Bool

A função esqPedra recebe as coordenadas de um Jogador e as Coordenadas das Pedras e dos PI e verifica se o jogador tem uma Pedra ou um PI à sua esquerda.

Propriedade:

esqPedra (x,y) []  = False

Por exemplo:

>>> esqPedra (3,2) [(2,2),(4,2),(2,3),(3,3),(2,4),(3,4),(4,4)]
True

dirPedra :: CoordenadaJ -> PedrasEPI -> Bool

A função dirPedra recebe as coordenadas de um Jogador e as Coordenadas das Pedras e dos PI e verifica se o jogador tem uma Pedra ou um PI à sua direita.

Propriedade:

dirPedra (x,y) []  = False

Por exemplo:

>>> dirPedra (3,2) [(2,2),(4,2),(2,3),(3,3),(2,4),(3,4),(4,4)]
True

baixoPedra :: CoordenadaJ -> PedrasEPI -> Bool

A função baixoPedra recebe as coordenadas de um Jogador e as Coordenadas das Pedras e dos PI e verifica se o jogador tem uma Pedra ou um PI por baixo de si.

Propriedade:

baixoPedra (x,y) []  = False

Por exemplo:

>>> baixoPedra (2,1) [(2,2),(4,2),(2,3),(3,3),(2,4),(3,4),(4,4)]
True

cimaPedra :: CoordenadaJ -> PedrasEPI -> Bool

A função cimaPedra recebe as coordenadas de um Jogador e as Coordenadas das Pedras e dos PI e verifica se o jogador tem uma Pedra ou um PI por cima de si.

Propriedade:

cimaPedra (x,y) []  = False

Por exemplo:

>>> cimaPedra (2,3) [(2,2),(4,2),(2,3),(3,3),(2,4),(3,4),(4,4)]
True

explodePU :: PowerUps -> Bombas -> Mapa -> PowerUps

A função explodePU recebe os PowerUps, as Bombas e um Mapa e produz uma lista com os PowerUps que são atingidos pelas Bombas.

Propriedade:

explodePU powerups [] _ = []

explodePU [] _ _ = []

Por exemplo:

>>> explodePU [(5,2),(3,3),(5,5)] [(4,5,1,1)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 1 1 0 1 1", "* 2 1 0 1 10"]
[(4,5)]

arrebentaPU :: Mapa -> Mapa

A função arrebentaPU recebe um Mapa e produz o novo mapa sem os powerups que são atingidos pelas bombas.

Propriedade:

arrebentaPU [] = []

Por exemplo:

>>> arrebentaPU ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 4 5 0 1 1", "0 1 1","1 2 1"]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","0,1,1","1 2 1"]

auxVerify :: Eq a => [a] -> [a] -> Bool

A função auxVerify recebe duas lista de elementos e verifica cada elemento da primeira lista com a segunda lista.

Propriedade:

auxVerify [] _ = False

auxVerify _ [] = False

Por exemplo:

>>> auxVerify [1,2,3,4,5] [5,6,7,8]
False

>>> auxVerify [1,2,3,4,5] [1,6,7,8]
True

removePowerUps :: PowerUps -> Mapa -> Mapa

A função removePowerUps recebe os PowerUps a eliminar, um Mapa e produz o novo mapa sem os powerups que são atingidos pelas bombas.

Propriedade:

removePowerUps [] mapa = mapa

removePowerUps mapa [] = []

Por exemplo:

>>> removePowerUps (5,5) ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 4 5 0 1 1", "0 1 1","1 2 1"]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","0,1,1","1 2 1"]

explodePI :: PontosInt -> Bombas -> Mapa -> PontosInt

A função explodePI recebe os Pontos de Interrogação, as Bombas e um Mapa e produz uma lista com os PontosInt que são atingidos pelas Bombas.

Propriedade:

explodePI [] _ _ = []

explodePI pontos [] _ = []

Por exemplo:

>>> explodePI [(3,2)] [(3,3,1,1)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5", "* 3 3 0 1 1"]
[(3,2)]

mesmoX :: PontosInt -> Int -> PontosInt

A função mesmoX recebe uma lista de Pontos de Interrogação e um Inteiro e fica com as coordendas em que o X é igual a esse Inteiro.

Propriedade:

mesmoX [] _ = []

Por exemplo:

>>> mesmoX [(3,2),(5,2),(3,3)] 2
[(3,2),(5,2)]

verificaCoordSeguinteX :: PontosInt -> Bool

A função verificaCoordSeguinteX recebe uma lista de Pontos de Interrogação e verifica se a coordenada seguinte na lista está na mesma linha ao lado da primeira coordenada.

Propriedade:

verificaCoordSeguinteX ((x,y):[])  = False

Por exemplo:

>>> verificaCoordSeguinteX [(3,2),(4,2),(3,3)]
True

verificaCoordSeguinteY :: PontosInt -> Bool

A função verificaCoordSeguinteY recebe uma lista de Pontos de Interrogação e verifica se a coordenada seguinte na lista está na mesma coluna por baixo da primeira coordenada.

Propriedade:

verificaCoordSeguinteY ((x,y):[])  = False

Por exemplo:

>>> verificaCoordSeguinteY [(3,2),(3,3),(5,3)]
True

tOnlyP :: Mapa -> Pedras

A função tOnlyP recebe um Mapa e origina a lista das coordenadas de todos os cardinais desse mapa.

Por exemplo:

>>> tOnlyP ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]
[(0,0),(1,0),(2,0),(3,0),(4,0),(5,0),(6,0),(0,1),(6,1),(0,2),(2,2),(4,2),(6,2),(0,3),(6,3),(0,4),(2,4),(4,4),(6,4),(0,5),(6,5),(0,6),(1,6),(2,6),(3,6),(4,6),(5,6),(6,6)]

listaSRXb :: PontosInt -> CoordenadaBomba -> PontosInt

A função listaSRXb recebe uma lista de Pontos de Interrogação e uma Coordenada de uma Bomba e remove dessa lista as coordenadas que explodem ou estão à direita de uma coordenada que explode.

Propriedade:

listaSRXb [] _  = []

Por exemplo:

>>> listaSRXb [(3,2),(3,3),(5,3)]  (1,2,3,1)
[(3,3),(5,3)]

listaSRYb :: PontosInt -> CoordenadaBomba -> PontosInt

A função listaSRYb recebe uma lista de Pontos de Interrogação e uma Coordenada de uma Bomba e remove dessa lista as coordenadas que explodem ou estão por baixo de uma coordenada que explode.

Propriedade:

listaSRYb [] _  = []

Por exemplo:

>>> listaSRXb [(3,2),(2,4),(2,5)]  (1,2,3,1)
[(2,4),(2,5)]

tI :: Mapa -> PontosInt

A função tI recebe um Mapa e origina a lista das coordenadas de todos os pontos de interrogação desse mapa.

Propriedade:

Por exemplo:

>>> tI ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]
[(2,3),(3,3),(3,4)]

arrebentaPI :: Mapa -> PontosInt -> Mapa

A função arrebentaPI recebe um Mapa e produz o novo mapa sem os powerups que são atingidos pelas bombas.

Propriedade:

arrebentaPI [] = []

arrebentaPI mapa [] = mapa

Por exemplo:

>>> arrebentaPI ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"] [(2,3)]
["#######","#     #","# # # #","#  ?  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]

arrPIStr :: String -> PontosInt -> String

A função arrebentaPIStr recebe uma String e uma lista de coordenadas de Pontos de Interrogação e apaga-os dessa String.

Por exemplo:

>>> arrebentaPIStr "# ??  #" [(2,3),(3,3)]
"#      #"

apagaUsadosY :: PontosInt -> Int -> PontosInt

A função apagaUsadosY recebe uma lista de coordenadas de Pontos de Interrogação e um Inteiro e apaga-os pontos de interrogação com o y igual a esse inteiro.

Propriedade:

apagaUsadosY [] _ = []

Por exemplo:

>>> apagaUsadosY [(2,3),(3,3),(5,5)] 3
[(5,5)]

mesmoY :: PontosInt -> Int -> PontosInt

A função mesmoY recebe uma lista de coordenadas de Pontos de Interrogação e um Inteiro e origina uma nova lista com os pontos de interrogação com o y igual a esse inteiro.

Propriedade:

mesmoY [] _ = []

Por exemplo:

>>> mesmoY [(2,3),(3,3),(5,5)] 3
[(2,3),(3,3)]

verBomba :: Mapa -> Bool

A função verBomba recebe um Mapa e verifica se há bombas que vão explodir, ou seja, com t = 1.

Propriedade:

verBomba [] = False

Por exemplo:

>>> verBomba ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 0 1 1"]
True

metePontoEspiral :: Mapa -> (Int, Int) -> Mapa

A função 'ḿetePontoEspiral' recebe um Mapa, e uma coordenada e vai ao Mapa colocar um '#' nessa coordenada.

Propriedade:

metePontoEspiral [] _ = []

Por exemplo:

>>> metePontoEspiral ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 0 1 1"] (1,1)
["#######","##    #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","* 1 1 0 1 1"]

substituiPedra :: String -> (Int, Int) -> String

A função substituiPedra recebe uma String (linha) e uma coordenada e coloca uma '#' nessa coordenada.

Por exemplo:

>>> substituiPedra "#     #" (6,1)
"#     ##"

coordespiral :: Int -> Int -> (Int, Int)

A função coordespiral recebe dois inteiros, a dimensão do mapa e nº de instantes de tempo restantes e produz a coordenada onde vai cai a pedra.

Por exemplo:

>>> coordespiral 5 9
(1,1)

tentandoespiral3 :: Int -> Int -> Int -> (Int, Int)

A função tentandoespiral recebe três inteiros, o nº de voltas dadas, a dimensão do mapa e nº de instantes de tempo restantes e produz a coordenada onde vai cai a pedra.

Por exemplo:

>>> tentandoespiral3 0 5 9
(1,1)

espiral :: Int -> Int -> (Int, Int)

A função espiral recebe a dimensão do mapa e o nº de instantes de tempo restantes e origina a coordenada. Esta função é geral e é usada para qualquer dimensão.

Por exemplo:

>>> espiral 7 1
(3,3)

f :: Int -> Int

A função f recebe a dimensão do mapa e dá o nº de voltas até meter a última pedra.

Por exemplo:

>>> f 7
2

adicionavoltas :: (Int, Int) -> Int -> (Int, Int)

A função adicionavoltas recebe um par de coordenadas e o nº da volta e dá o novo par de coordenadas.

Por exemplo:

>>> adicionavoltas (1,1) 1
(2,2)

espiralback :: Int -> Int -> Int -> Int -> (Int, Int)

A função espiralback é uma função auxiliar da coordespiral, recebe o x e o y, a dimensão do mapa e o inteiro e dá a coordenada onde será colocada a próxima pedra. Esta função tem como objetivo dar as coordenadas quando a espiral está a voltar para trás.

Por exemplo:

>>> espiralback 4 5 7 25
(4,5)

espiralup :: Int -> Int -> Int -> Int -> (Int, Int)

A função espiralup é uma função auxiliar da coordespiral, recebe o x e o y, a dimensão do mapa e o inteiro e dá a coordenada onde será colocada a próxima pedra. Esta função tem como objetivo dar as coordenadas quando a espiral está a subir.

Por exemplo:

>>> espiralup 1 4 7 25
(1,4)

removeGeral :: (Int, Int) -> Mapa -> Mapa

A função removeGeral recebe um par de coordenadas e um mapa e elimina essa coordenada do mapa

Propriedade:

removeGeral x []  = []

Por exemplo:

>>> removeGeral (1,1) ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 4","* 1 1 0 1 10","0 1 1"]
["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 4"]

tuploG :: [String] -> (Int, Int)

A função tuploG recebe uma lista de strings e forma um par de coordenadas.

Por exemplo:

>>> tuploG ["0","1","1"]
(1,1)

tuploGeral :: String -> (Int, Int)

A função tuploGeral recebe uma string e dá o par de coordenadas dessa string.

Por exemplo:

>>> tuploGeral "* 1 1 0 1 10"
(1,1)

avanca :: Mapa -> Int -> Mapa

A função avanca é a função principal desta Tarefa. Recebe um Mapa e um Inteiro e origina um novo Mapa efetuando todas as alterações nesse mapa originadas pela passagem de tempo.

Por exemplo:

>>> avanca ["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######","+ 3 3","+ 3 4","0 2 1","* 1 1 0 1 1","* 1 3 0 1 1"] 100
["#######","#     #","# # # #","#  ?  #","# #?# #","#     #","#######","+ 3 3","+ 3 4"]

main :: IO ()