Main

Description

Um módulo contendo definições Haskell para o cálculo de funções que permitam concretizar um mapa do jogo Bomberman sob a forma de lista de strings, partindo de uma dimensão e de uma semente geradora de números aleatórios.

Documentation

grelha :: Int -> Int -> String

A função grelha recebe dois números inteiros, isto é, o número da linha do mapa e a sua dimensão, e devolve a linha y do mapa em forma de string. É de notar que a primeira linha do mapa é a linha 0 e a última é a linha d-1.

Propriedade:

Para a função ter o resultado desejado, tem que respeitar o seguinte: y <= d

Por exemplo:

>>> grelha 1 9
"#  ???  #"
>>> grelha 8 9
"#########"

aux :: Int -> Char -> String

A função aux é uma função auxiliar, utilizada na grelha, que nos dá uma string que representa as linhas em posições ímpares (exceto as linhas 1 para qualquer dimensão superior a 5 e linhas 1 e 3 nos mapas de dimensão 5) de um mapa de dimensão d. Recebe a dimensão do mapa e um Char, que tem que ser ' ' para o propósito desejado.

Propriedade:

aux 0 x = " "

Por exemplo:

>>> aux 9 ' '
"#       #"
>>> aux 13 ' '
"#           #"

aux1 :: Int -> Char -> String

A função aux1 é uma função auxiliar, utilizada na grelha, que nos dá uma string que representa as linhas em posições pares (exceto as linhas 0, (d-1), 2 e (d-3)) de um mapa de dimensão d. Recebe a dimensão do mapa e um Char, que tem que ser '#' para o propósito desejado.

Por exemplo:

>>> aux1 9 '#'
"#?#?#?#?#"
>>> aux1 13 '#'
"#?#?#?#?#?#?#"

unwords1 :: String -> String

A função unwords1 é uma função auxiliar, utilizada na aux1, que dada uma string, devolve uma lista dos elementos que compõem a string intercalados com ?, sob a forma de string.

Propriedades:

unwords1 [] = "?"

unwords1 [x] = [x]

Por exemplo:

>>> unwords1 "####"
"#?#?#?#"
>>> unwords1 '##'
"#?#"

mapaSemN :: Int -> Int -> [String]

A função mapaSemN é uma função que, dados dois números inteiros (posição da linha que encabeça o mapa e a dimensão), devolve uma lista de strings que representam o mapa. Para a lista devolvida não ser vazia, o primeiro inteiro recebido tem que ser inferior ou igual a (d-1).

Por exemplo:

>>> mapaSemN 0 9
["#########","#  ???  #","# #?#?# #","#???????#","#?#?#?#?#","#???????#","# #?#?# #","#  ???  #","#########"]
>>> mapaSemN 2 9
["# #?#?# #","#???????#","#?#?#?#?#","#???????#","# #?#?# #","#  ???  #","#########"]

numeros :: Int -> Int -> [Int]

A função numeros é uma função que nos dá uma lista de números gerados aleatóriamente, cujo tamanho varia consoante a dimensão do mapa. O seu contéudo depende da semende inserida na função.

Por exemplo:

>>> numeros 9 0
[83,93,63,38,0,87,81,1,61,86,13,50,32,80,54,25,90,31,65,92,2,76,70,25,6,29,10,99]
>>> numeros 7 1
[35,26,65,67,58,27,14,35,60]

A função nPontos é uma função auxiliar, definida dentro da numeros, que conta o número de pontos que são necessário de preencher dentro do mapa com a lista de números aleatórios, dada a dimensão do mapa, ou seja, defino o tamanho da lista devolvida pela numeros. Esta função parte da nPedras.

Por exemplo:

>>> nPontos 9
28
>>> nPontos 7
9

A função nPedras é uma função auxiliar, definida dentro da numeros, que conta o número de pedras de cada mapa, dada a sua dimensão.

Por exemplo:

>>> nPedras 9
41
>>> nPedras 7
28

alterarLinha :: String -> [Int] -> String

A função alterarLinha é uma função que, dada uma linha do mapa e a lista de inteiros gerada pela numeros, vai substituir os caracteres ? pelo valor associado a cada número da lista de inteiros, ordenadamente, isto é, o primeiro espaço é substituído pelo valor do primeiro número, o segundo espaço pelo valor do segundo número e assim sucessivamente, devolvendo a lista alterada.

Sendo y um número dessa lista, os valores associados são os seguintes: 1. y = 0 ou y = 1 - + (representa o power-up Bombs) 2. y = 2 ou y = 3 - ! (representa o power-up Flames) 3. 4 <= y <= 39 - ? (representa um tijolo) 4. 40 <= y <= 99 - ' ' (representa um espaço vazio no mapa)

Propriedades:

alterarLinha [] _ = []

alterarLinha l1 [] = l1

Por exemplo:

>>> alterarLinha "#?????#" [1,3,12,67,31]
"#+!? ?#"
>>> alterarLinha "#?#?#?#" [1,3,12,67,31]
"#+#!#?#"

linha :: [String] -> Int -> String

A função linha, dado o mapa (lista de strings) e um número inteiro, dá-nos uma string, ou seja, a linha do mapa que está na posição indicada pelo número inserido. É de notar que a posição da primeira linha corresponde a 0.

Propriedades:

linha [] _ = []

linha (x:xs) 0 = x

linha (x:y:xs) 1 = y

Por exemplo:

>>> linha ["#########","#  ???  #","# #?#?# #","#???????#","#?#?#?#?#","#???????#","# #?#?# #","#  ???  #","#########"] 3
"#???????#"
>>> linha ["#########","#  ???  #","# #?#?# #","#???????#","#?#?#?#?#","#???????#","# #?#?# #","#  ???  #","#########"] 6
"# #?#?# #"

numerotirar :: Int -> Int -> Int

A função numerotirar recebe a posição de uma linha e a dimensão do mapa e devolve o número de inteiros gerados aleatóriamente utilizados nessa linha e nas anteriores.

Por exemplo:

>>> numerotirar 3 9
12
>>> numerotirar 3 11
17

dropar :: Int -> Int -> Int -> [Int]

A função dropar tem como objetivo dar-nos a lista de inteiros que serão utilizados para preencher uma linha, não permitindo que sejam utilizados os números das linhas anteriores. A função devolve-nos essa lista recebendo a dimensão do mapa, a semente da função numeros e a posição da linha. Na base desta função estão portanto as funções numeros e numerotirar. Para a lista devolvida não ser vazia, a posição da lista terá que ser menor ou igual a (d-1).

Por exemplo:

>>> dropar 9 0 3
[87,81,1,61,86,13,50,32,80,54,25,90,31,65,92,2,76,70,25,6,29,10,99]
>>> dropar 9 0 4
[32,80,54,25,90,31,65,92,2,76,70,25,6,29,10,99]

posicao :: String -> Char -> Int

A função posicao recebe uma linha do mapa e um Char e devolve a posição desse Char na string, ou seja, dá-nos um inteiro que representa a posição da primeira ocorrência desse caracter. É importante ter em atenção que o primeiro caracter de uma string ocupa a posição 0.

Por exemplo:

>>> posicao "# #?#?# #" '?'
3
>>> posicao "#?#?#?#?#" '?'
1

tirarPrimeiro1 :: String -> String

A função tirarPrimeiro1 recebe uma linha do mapa e devolve essa mesma linha, mas substituindo a primeira ocorrência do caracter + pelo caracter ?, de maneira a evitar repetir as coordenadas dos power-ups.

Propriedade:

tirarPrimeiro1 [] = []

Por exemplo:

>>> tirarPrimeiro1 "#?# #+#+#"
"#?# #?#+#"
>>> tirarPrimeiro1 "#?# #+#!#"
"#?# #?#!#"

tirarPrimeiro2 :: String -> String

A função tirarPrimeiro2 tem o mesmo objetivo da tirarPrimeiro1, mas funciona para a primeira ocorrência do caracter !.

Propriedade:

tirarPrimeiro2 [] = []

Por exemplo:

>>> tirarPrimeiro2 "#?#+#!#!#"
"#?#+#?#!#"
>>> tirarPrimeiro1 "#?# #!# #"
"#?# #?# #"

darCoordLinha1 :: String -> String -> Int -> [String]

A função darCoordLinha1, que recebe uma linha do mapa (duas vezes) e a sua posição no mapa, tem como objetivo imprimir uma lista de strings com as coordenadas dos power-ups Bombs (+) na linha do mapa inserida. As strings resultantes terão o formato "+ (posicao na linha) (posicao da linha)". Tem por base as funções posicao e tirarPrimeiro1.

Propriedade:

darCoordLinha1 [] _ _ = []

Por exemplo:

>>> darCoordLinha1 "# #+# #" "# #+# #" 0
["+ 3 0"]
>>> darCoordLinha1 "#+#+#+#" "#+#+#+#" 0
["+ 1 0","+ 3 0","+ 5 0"]

darCoordLinha2 :: String -> String -> Int -> [String]

A função darCoordLinha2 irá funcionar da mesma maneira que a darCoordLinha1, mas para os power-ups Flames (!). As strings devolvidas terão o formato "! (posição na linha) (posição da linha)".

Propriedade:

darCoordLinha2 [] _ _ = []

Por exemplo:

>>> darCoordLinha2 "# #!# #" "# #!# #" 0
["! 3 0"]
>>> darCoordLinha2 "#!#!#!#" "#!#!#!#" 0
["! 1 0","! 3 0","! 5 0"]

mapa1 :: Int -> Int -> [String]

A função mapa1 recebe a dimensão do mapa e a semente utilizada para gerar a lista de números aleatórios e devolve uma lista de strings que representam as linhas do mapa. Nestas linhas estão inseridos os power-ups, representados pelos seus caracteres (+ - Bombs,! - Flames), bem como os tijolos ou espaços vazios definidos pela lista de números aleatórios. Na sua base, estão as funções alterarLinha, linha,mapaSemN e dropar.

Por exemplo:

>>> mapa1 9 0
["#########","#       #","# #?#+# #","#  +  ? #","#?# # #?#","# ?  !  #","# #?#?# #","#  ??   #","#########"]
>>> mapa1 9 2
["#########","#    ?  #","# # #?# #","#   ??  #","# # #?#+#","#?    + #","# #?# # #","#  ? ?  #","#########"]

darCoordMapa1 :: Int -> Int -> Int -> [String]

A função darCoordMapa1 tem por base a darCoordLinha1, uma vez que, recebendo a dimensão do mapa, a semente recebida e a posição da primeira linha do mapa (0), devolve uma lista de strings contendo as coordenadas do power-up Bombs (+) existentes em todo o mapa. Utiliza ainda as funções mapa1 e linha.

Por exemplo:

>>> darCoordMapa1 9 0 0
["+ 5 2","+ 3 3"]
>>> darCoordMapa1 9 2 0
["+ 7 4","+ 6 5"]

darCoordMapa2 :: Int -> Int -> Int -> [String]

A função darCoordMapa2 funciona de maneira idêntica à darCoordMapa1, mas devolve a lista das coordenadas dos power-ups Flames (!).

Por exemplo:

>>> darCoordMapa2 9 0 0
["! 5 5"]
>>> darCoordMapa2 9 7 0
["! 5 1","! 5 4"]

substituiLinha :: String -> String

A função substituiLinha, recebendo uma linha do mapa, devolve essa mesma linha alterando os caracteres dos power-ups (+ e !) para o caracter ?.

Propriedade:

substituiLinha [] = []

Por exemplo:

>>> substituiLinha "#! + !#"
"#? ? ?#"
>>> substituiLinha "#!+!+!#"
"#?????#"

substituiMapa :: [String] -> [String]

A função substituiMapa irá realizar o mesmo processo que a substituiLinha, mas desta vez para todo o mapa, ou seja, para toda a lista de strings que o compõem.

Propriedade:

substituiMapa [] = []

Por exemplo:

>>> substituiMapa ["#########","#       #","# #?#+# #","#  +  ? #","#?# # #?#","# ?  !  #","# #?#?# #","#  ??   #","#########"]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]
>>> substituiMapa ["###########","#  ?  ??  #","# #?#?# # #","#? +  ? ?+#","# # # #?# #","#   !   ??#","# #?# # #?#","#  ? ?? ??#","# #?# # # #","#    !    #","###########"]
["###########","#  ?  ??  #","# #?#?# # #","#? ?  ? ??#","# # # #?# #","#   ?   ??#","# #?# # #?#","#  ? ?? ??#","# #?# # # #","#    ?    #","###########"]

mapa :: Int -> Int -> [String]

A função mapa é a função principal, isto é, com todas as funções definidas anteriormente, contrói um mapa (lista de strings) a partir da sua dimensão e da semente utilizada para gerar a lista de números aleatórios referida anteriormente. Funciona concatenando as listas modificas do mapa (através de substituiMapa e mapa1) e as listas das coordenadas dos power-ups (derivadas das funções darCoordMapa1 e darCoordMapa2).

Por exemplo:

>>> mapa 9 0
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]
>>> mapa 11 8
["###########","#  ?  ??  #","# #?#?# # #","#? ?  ? ??#","# # # #?# #","#   ?   ??#","# #?# # #?#","#  ? ?? ??#","# #?# # # #","#    ?    #","###########","+ 3 3","+ 9 3","! 4 5","! 5 9"]

main :: IO ()

O programa main foi-nos facultado para a realização deste projeto e pode ser usado para testar a função mapa, por nós definida, ou seja, após compilado, o programa aceita como parâmetros a dimensão e a semente e depois invoca a função ​ mapa ​ imprimindo o resultado no stdout.