Main

Description

Descrption: funções que permitem a construção do mapa Copyright: Ana Ribeiro Jéssica Lemos

Documentation

mapa :: Int -> Int -> [String]

A função "mapa" faz a junção do mapa com os PowerUps. Dado que o mapa inicia na coordenada (0,0) à dimensão retiramos um.

locPonto :: (Int, Int) -> Int -> [Int] -> [Char]

A função "locPonto" indica para cada ponto do mapa se este é pedra("#") ou vazio (""). Tivemos em atenção que os extremos do mapa, ou seja, a primeira e ultima colunas e linhas são pedras. Posteriormente, recorremos à função "locVazios" que indica as posições dos vazios pré-definidos. Além disto, verificamos que se a linha e a coluna forem pares encontra-se também uma pedra. Nos restantes casos, verificamos se o conteúdo de cada ponto corresponde a um PowerUp (+ ou !), um tijolo (?) ou um espaço vazio (' ').

aConteudo :: [Int] -> [Char]

"aConteudo" associa o número, proveniente do cálculo de numeros aleatório recorrendo à função "semente", a um determinado conteúdo. Caso o número seja menor que 0 e 1 o seu conteúdo é um PowerUp Bombs (+), entre 2 e 3 é um PowerUo Flames (!), entre 3 e 39 é um tijolo e osrestantes é um espaço vazio.

locLinha :: (Int, Int) -> Int -> [Int] -> [Char]

Esta função une todos os pontos de cada linha. Começamos por verificar se a "locPonto" é pedra ou vazio pré-definido. Se o for, mantemos o ponto e recorremos de novo à "locLinha" para verificar o ponto seguinte da linha. Desta forma, somamos 1 ao y e retiramos um valor à sua dimenção uma vez que consideramos que cada linha tem d pontos sendo que o ultimo corresponde a d=0. Daí um dos nossos casos de paragem ser locLinha (x,y) 0 l = "#". Caso a "locPonto" corresponda a um PowerUp, recorresmos à função "converte" uma vez que estes se encontram escondidos num tijolo e recorremos novamente à "locLinha". Note-se que neste caso eliminamos um elemento da lista que contém os números aleatórios, dado que o primeiro número já foi associado. Aquando a "locPonto" está associada a um vazio mantemos o ponto e fazemos a recursiva pelo mesmo métedo anterior. Quando a lista dos números aleatórios se tornar vazia prosseguimos seguindo o mesmo processo.

juntaLinhas :: (Int, Int) -> Int -> [Int] -> [String]

A "juntaLinhas" faz a junção de todas as linhas, recorrendo ao aumento de x. Enquanto a lista de números aleatórios não for vazia vamos eliminando o número de pontos de cada linha que já foram associados. Dando origem ao mapa -

pontoPowerUp :: (Int, Int) -> Int -> [Int] -> [[Char]]

A função dá os PowerUps de determinados pontos. Caso a "locPonto" seja pedra ou um vazio pré-definido devolve-nos a lista vazia, uma vez que não contém PowerUps. O mesmo se verifica caso a "locPonto" não seja um PowerUp. Se "locPonto" corresponder a um PowerUp ao símbolo que lhe correspondo juntamos a coordenada do y e do x, utilizando a função show e seguindo a formatação definida.

locBombs :: (Int, Int) -> Int -> [Int] -> [String]

Esta função junta todos os Bombs de determinada linha. Se a "locPonto" contiver um Bomb então recorre à "pontoPowerUp" e prossegue para o ponto seguinte (daí o y aumentar, a dimensão diminuir e ser eliminada a cabeça da lista). Aquando o ponto corresponde a uma pedras ou vazio pré-definido, a função segue para o seguinte sem eliminar um elemento da lista (por não ser utilizado). Nas seguintes situações recorremos novamente à "locBombs" para verificar o ponto seguinte com a eliminação de cabeças da lista. É de salientar que sempre que chamamos a "locPonto" ao d somamos o y de modo a compensar a redução da dimensão aplicada na recursividade da locBombs.

juntaBombs :: (Int, Int) -> Int -> [Int] -> [String]

Esta junta os Bombs de todas linhas, recorrendo ao aumento de x. Isto é, a função recorre à "locBombs" e faz a recursiva para a linha seguinte. Enquanto a lista de números aleatórios não for vazia vamos eliminando o número de pontos de cada linha que já foram associados. Quando a lista de número aleatórios for vazia devolve a lista vazia dado que, não existem mais Bombs para juntar.

locFlames :: (Int, Int) -> Int -> [Int] -> [String]

Esta função junta todos os Flames de determinada linha. Se a "locPonto" contiver um Flame então recorre à "pontoPowerUp" e prossegue para o ponto seguinte (daí o y aumentar, a dimensão diminuir e ser eliminada a cabeça da lista). Aquando o ponto corresponde a uma pedras ou vazio pré-definido, a função segue para o seguinte sem eliminar um elemento da lista (por não ser utilizado). Nas seguintes situações recorremos novamente à "locBombs" para verificar o ponto seguinte com a eliminação de cabeças da lista. É de salientar que sempre que chamamos a "locPonto" ao d somamos o y de modo a compensar a redução da dimensão aplicada na recursividade da locBombs.

juntaFlames :: (Int, Int) -> Int -> [Int] -> [String]

Esta junta os Flames de todas linhas, recorrendo ao aumento de x. Isto é, a função recorre à "locFlames" e faz a recursiva para a linha seguinte. Enquanto a lista de números aleatórios não for vazia vamos eliminando o número de pontos de cada linha que já foram associados. Quando a lista de número aleatórios for vazia devolve a lista vazia dado que, não existem mais Flames para juntar.

powerUps :: (Int, Int) -> Int -> [Int] -> [String]

A "powerUps" junta todos os Bombs e Flames, ordenados como pedido.

converte :: (Int, Int) -> Int -> [Int] -> [Char]

Esta converte os símbolos dos PowerUps em tijolos, visto que estes aparecem escondidos atrás de um tijolo no mapa.

nCelulasLinha :: (Int, Int) -> Int -> [Int] -> Int

"nCelulasLinha" calcula o número de números aleatórios numa determinada linha. Note-se que a dimensão é chamada com d-1. A primeira e última linha não necessita de nenhum, uma vez que é pedra. Na segunda e penúltima linhas à dimensão retiramos cinco, ou seja, dois vazios pré-definidos e duas pedras. Se a linha for impar à dimensão refiramos um, ou seja, as duas pedras dos extremos. No caso da terceira e antepenúltima linhas calculamos através da expressão (d+1)-((div (d+2) 2) +2), em que d corresponde à dimensão. Nos restantes casos, (d+1)-(div (d+2) 2).

nPedras :: Integral a => a -> a

Esta função indica o número de pedras do mapa, recorrendo apenas ao tamanho do mapa.

nCelulas :: Integral a => a -> a

A função nCelulas dá o número de céluas que irão ser associadas a um número. Para tal ao número total de pontos do mapa retiramos as pedras e os vazios pré-definidos.

semente :: Int -> Int -> [Int]

A "semente" gera um conjunto de números aleatórios tendo em conta a semente. Usamos a função "nCelulas" uma vez que apenas serão associados números aleatórios a estes pontos.

locVazios :: (Int, Int) -> Int -> Bool

A função "locVazios" indica as posições do vazios pré-definidos, tendo em conta que os vazios pré-definidos apenas se encontram nas linha um, dois, d-1 e d-2 e nas colunas um, dois, d-1 e d-2. Tivemos em atenção que as colunas não podem ser ambas dois nem d-2.

main :: IO ()