Main

Description

Descrption: funções que permitem o avanço do jogo Copyright: Ana Ribeiro Jéssica Lemos

Documentation

avanca :: [String] -> Int -> [String]

Esta função dado o estado actual do jogo e o número instantes de tempo que faltam para o jogo terminar, devolve o novo estado do jogo. No caso de existirem bombas prestes a explodir, recorremos à função "juntaAlteraFinal" eliminando a informação das bombas que explodiram. Se existirem bombas prestes a explodir e outras,então, reduzimos o tempo das bombas que não estão prestes a explodir ao mapa já com as alterações das bombas que explodiram. Na eventualidade, de existirem bombas com tempo zero, então, alteramo-lo para um. Na hipótese de não existir bombas no mapa, é devolvido o mesmo estado de jogo. Nos restantes casos diminuimos ao tempo de explosão das bombas.

dimTempo :: [String] -> Int -> [[Char]]

Esta função permite a diminuição do tempo nas bombas à medida que a função "avanca" é chamada. Caso uma das bombas tenha tempo de ação 1 então o seu tempo mantém-se.

me :: [Char] -> [String] -> Int -> ([Char], [Char])

Esta função devolve um tuplo em que o primeiro elemento corresponde ao tempo de explosão da bomba e o segundo à sua restante informação.

dim :: [String] -> Int -> Int

A "dim" devolve a dimensão do mapa.

haBomb0Exp :: [String] -> Int -> [String]

Esta função percorre a lista, através de um map aplicado a "infBombas", caso o tempo de ação da bomba seja 0 então este tempo é substituído por 1.

haBomb0 :: [String] -> Int -> [Bool]

A função verifica se existem bombas com tempo de explosão 0. Deste modo, percorremos a lista procurando se existe o símbolo da bomba e se o tempo de explosão é 0.

haBombExp :: [String] -> Int -> Bool

A "haBombExp" verifica se existe bombas prestes a explodir no mapa.Para tal, percorre a lista comçando pelo fim para ver se existe uma bomba cujo o tempo de explosão é um.

existeBomb :: [String] -> Int -> Bool

Esta função verifica se no mapa existem bombas, assim percorremos este verificando se existe o símbolo de bomba.

existeBNExp :: [String] -> Int -> Bool

Com esta função verificamos se existem bombas sem estarem prestes a explodir. Para tal percorremos a lista verificando se existe o símbolo de bomba e se o seu tempo de explosão é diferente de 1.

infBombas :: [String] -> Int -> [String]

A "infBombas" devolve uma lista com todas as bombas do mapa. Assim, percorremos a lista até encontrarmos o símbolo de bomba, caso encontremos guardamos a sua informação fazendo recursiva desta para verificarmos se existem mais bombas.

raio :: [String] -> Int -> [Int]

A função raio devolve uma lista com os raios das bombas prestes a explodir (ou seja, com tempo igual a 1). Começamos por percorrer a lista até encontrar uma bomba (*) e com tempo de explosão 1. Quando esta é encontrada vamos retirar-lhe a informação do seu raio e fazemos uma recursiva para verificar a existência de mais bombas nesta situação. Para tal, usamos a função break que permitiu separar o tempo de explosão da bomba da sua restante informação. E de seguida, pegamos no segundo elemento do tuplo ao qual retiramos o primeiro elemento visto que correspondia a um ' ' (contém a restante informação da bomba) e recolhemos o primeiro elemento do resultado da função break para obtermos o raio da bomba. Caso não exista nenhuma bombas prestes a explodir devolve uma lista com raio 0.

infBomb :: [String] -> Int -> [String]

Esta função dá uma lista com a informação de todas as bombas prestes a explodir sem o símbolo da bomba e o espaço entre este e o próximo algarismo. Inicialmente percorremos a lista até encontrar uma bomba nestas condições. Quando tal se sucede, retiramos através de "drop 2" o símbolo e o espaço referidos anteriormente à informação da bomba. Fazemos de seguida a recursiva para aplicar o mesmo processo às restantes bombas nestas condições.

cooryBomb :: [String] -> Int -> [Int]

A "cooryBomb" fornece uma lista com todas as coordenadas y de todas as bombas prestes a explodir. Assim, aplicamos a função map à "infBomb". Esta percorre todos os elementos da lista e aplica-lhes a função f, que isola o primeiro elemento resultante do break.

coorxBomb :: [String] -> Int -> [Int]

A "coorxBomb" fornece uma lista com todas as coordenadas x de todas as bombas prestes a explodir. Assim, aplicamos a função map à "infBomb". Esta percorre todos os elementos da lista e aplica-lhes a função f, que isola o primeiro elemento do tuplo resultante do break aplicado à remoção da coordenada y, através de um break, e do espaço que os separava, recorrendo ao drop.

convert :: [String] -> Int -> [Int]

A função multiplica todas as coordenadas x, das bombas prontas a explodir, pela dimensão do mapa. A função "dim" indica a dimensão do mapa recebido. Para tal calculamos o tamanho da cabeça da lista, que corresponde à primeira linha do mapa constituída apenas por pedras.

convertToNum :: [String] -> Int -> [Int]

Esta converte as coordenadas das bombas prestes a explodir num número do mapa. Para tal, recorremos à função auxiliar "aux" que junta cada coordenada y mais 1 com a correspondente coordenada x, que foi anteriormente multiplicada pela dimensão do mapa, na função "convert". Note-se que conseideramos cada ponto do mapa um número sendo que o primeiro ponto do mapa corresponde ao numero 1 e assim sucessivamente.

infJogs :: [String] -> Int -> [String]

Esta função devolve uma lista com toda a informação dos jogadores. Assim, percorremos a lista do mapa invertido até encontrarmos um algarismo. que corresponde ao identificador de um jogador. Fazendo recursividade verificamos se existiam mais jogadores, caso existam guardamos também a sua informação.

coorJog :: [String] -> Int -> [String]

"coorJog" devolve uma lista com a informação dos jogadores sem os Power ups.

tuploCoordJogs :: [String] -> Int -> [([Char], [Char])]

A função dá uma lista de tuplos com as coordenadas x e y dos jogadores. Utilizando um map, aplicamos um break à função que devolvia a informação dos jogadores de modo a obter as suas coordenadas.

cooryJogs :: [String] -> Int -> [Int]

A função dá uma lista com as coordenadas y dos jogadores, para tal convertemos todos os primeiros elementos dos tuplos em algarismos.

coorxJogs :: [String] -> Int -> [Int]

A função dá uma lista com as coordenadas x dos jogadores, para tal convertemos todos os segundos elementos dos tuplos em algarismos.

convertJogs :: [String] -> Int -> [Int]

A função multiplica todas as coordenadas x, dos jogadores, pela dimensão do mapa. A função "dim" indica a dimensão do mapa recebido. Para tal calculamos o tamanho da cabeça da lista, que corresponde à primeira linha do mapa constituída apenas por pedras.

convertToNumJogs :: [String] -> Int -> [Int]

Esta converte as coordenadas dos jogadores num número do mapa. Para tal, recorremos à função auxiliar "aux" que junta cada coordenada y mais 1 com a correspondente coordenada x, que foi anteriormente multiplicada pela dimensão do mapa, na função "convert".

infPU :: [String] -> Int -> [String]

Esta função devolve uma lista com todos os PowerUps do mapa. Assim, percorremos a lista do mapa invertido até encontrarmos o símbolo + ou ! que corresponde a um PowerUp. Fazendo recursividade verificamos se existiam mais PowerUps, caso existam guardamos também a sua informação.

tuploCoord :: [String] -> Int -> [([Char], [Char])]

A função dá uma lista de tuplos com as coordenadas x e y dos PowerUps. Utilizando um map, aplicamos um break à função que devolvia a informação dos PowerUps de modo a obter as suas coordenadas.

cooryPU :: [String] -> Int -> [Int]

A função dá uma lista com as coordenadas y dos PowerUps, para tal convertemos todos os primeiros elementos dos tuplos em algarismos.

coorxPU :: [String] -> Int -> [Int]

A função dá uma lista com as coordenadas x dos PowerUps, para tal convertemos todos os segundos elementos dos tuplos em algarismos.

convertPU :: [String] -> Int -> [Int]

A função multiplica todas as coordenadas x, dos PowerUps, pela dimensão do mapa. A função "dim" indica a dimensão do mapa recebido. Para tal calculamos o tamanho da cabeça da lista, que corresponde à primeira linha do mapa constituída apenas por pedras.

convertToNumPU :: [String] -> Int -> [Int]

Esta converte as coordenadas dos PowerUps num número do mapa. Para tal, recorremos à função auxiliar "aux" que junta cada coordenada y mais 1 com a correspondente coordenada x, que foi anteriormente multiplicada pela dimensão do mapa, na função "convert". Note-se que conseideramos cada ponto do mapa um número sendo que o primeiro ponto do mapa corresponde ao numero 1 e assim sucessivamente.

infTodasBomb :: [String] -> Int -> [String]

Esta função dá uma lista com a informação de todas as bombas sem o símbolo da bomba e o espaço entre este e o próximo algarismo. Inicialmente percorremos a lista até encontrar uma bomba tal como na função "infBomb" sendo que nesta não existe restrição ao tempo que falta para a bomba explodir.

cooryTodasBomb :: [String] -> Int -> [Int]

A "cooryTodasBomb" segue o mesmo procedimento da função "cooryBomb".

coorxTodasBomb :: [String] -> Int -> [Int]

A "coorxTodasBomb" segue o mesmo raciocínio da função "coorxBomb".

convertTodas :: [String] -> Int -> [Int]

Esta também tem o mesmo procedimento da função "convert".

convertToNumTodas :: [String] -> Int -> [Int]

Esta converte as coordenadas das bombas num número do mapa, tendo em conta também a função "convertToNum"

indicaR :: [String] -> Int -> [Int]

A função "indicaR" devolve uma lista com o número máximo que a chama da bomba pode alcançar para a direita. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo somar ao número do mapa onde se encontra a bomba ao raio correspondente.

restricaoR :: [String] -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaR". Uma vez que a dimensão máxima da chama é o limite máximo do mapa à direita, caso o número do mapa correspondente à chama da bomba pertença à linha seguinte, ou seja, maior que

indicaL :: [String] -> Int -> [Int]

A função "indicaL" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para a esquerda. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo subtrair ao número do mapa onde se encontra a bomba ao raio correspondente.

restricaoL :: [String] -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaL".

indicaU :: [String] -> Int -> [Int]

A função "indicaU" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para baixo. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo subtrair ao número do mapa onde se encontra a bomba o resultado do produto da dimensão com o raio correspondente.

restricaoU :: [String] -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaU".

indicaD :: [String] -> Int -> [Int]

A função "indicaD" devolve uma lista com o número máximo que a chama da bomba pode alcançar pode alcançar para a cima. Para tal, recorremos a uma auxiliar que recebia como argumentos a lista com os raios das bombas prestes a explodir e a lista com os números do mapa correspondentes à localização destas bombas. Esta tem como objetivo somar ao número do mapa onde se encontra a bomba o resultado do produto da dimensão com o raio correspondente

restricaoD :: [String] -> Int -> [Int]

Esta função, como o próprio nome indica restringe a informação dada na lista resultante da função "indicaL".

numerosL :: [String] -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Desta forma, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir, pelo que geramos uma lista cujo primeiro elemento é o número máximo da chama e o último é o número da bomba, sendo necessário inverter esta lista uma vez que a chama inicia no local do mapa onde se encontra a bomba. Recorremos à recursiva para as restantes bombas prestes a explodir.

numerosR :: [String] -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Assim, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir, pelo que geramos uma lista cujo primeiro elemento é o número da bomba mais um e o último é o número máxima da chama. Recorremos à recursiva para as restantes bombas prestes a explodir.

numerosU :: [String] -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Pelo que, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir. Para tal, recorremos a uma lista de compreensão, que cada elemento da lista será o número do mapa onde se encontra a bomba menos o produto da dimensão do mapa por um "a" (o que se deve ao facto de a chama se propagar para cima) tal que este "a" vai ser cada elemento de uma lista dos números de um até à divisão da diferença dos números do mapa onde se encontra a bomba e da chama máxima pela dimensão deste. É necessária a recursiva para que o mesmo seja feito para as restantes bombas prestes a explodir.

numerosD :: [String] -> Int -> [[Int]]

Esta devolve uma lista de listas com todos os números do mapa pela qual a chama passa. Pelo que, através de uma auxiliar recebemos o número do mapa onde se encontram as bombas e o número máximo do mapa até onde a chama pode proceguir. Para tal, recorremos a uma lista de compreensão, que cada elemento da lista será o número do mapa onde se encontra a bomba mais o produto da dimensão do mapa por um "a" (o que se deve ao facto de a chama se propagar para baixo) tal que este "a" vai ser cada elemento de uma lista dos números de um até à divisão da diferença dos números do mapa onde se encontra a bomba e da chama máxima pela dimensão deste. É necessária a recursiva para que o mesmo seja feito para as restantes bombas prestes a explodir.

convertToSimbR :: [String] -> Int -> [[String]]

A função transforma a listas devolvida pela funcão "numerosR" nos correspondentes símbolos do mapa: "#"," " e "?". Para tal usamos uma auxiliar que tem com argumento a função "numerosR" (se o último elemento uma vez que é uma lista vazia resultante do caso de paragem). Esta chama a função "convertToSimb" que converte cada número de uma lista no respetivo símbolo. Posteriormente, é feita a recursiva da função "aux" para que sejam convertidas as restantes listas das restantes bombas prestes a explodir.

convertToSimbD :: [String] -> Int -> [[String]]

A função transforma a listas devolvida pela funcão "numerosD" nos correspondentes símbolos do mapa: "#"," " e "?".

convertToSimbL :: [String] -> Int -> [[String]]

A função transforma a listas devolvida pela funcão "numerosL" nos correspondentes símbolos do mapa: "#"," " e "?". Esta segue o mesmo procedimento da função anterior diferindo apenas o argumento que é a função "numerosL".

convertToSimbU :: [String] -> Int -> [[String]]

A função transforma a listas devolvida pela funcão "numerosU" nos correspondentes símbolos do mapa: "#"," " e "?". Esta segue o mesmo procedimento da função anterior diferindo apenas o argumento.

toNumerosU :: [String] -> Int -> [[Int]]

A "toNumerosU" converte a lista devolvida pela "convertToSimbU" através de uma auxiliar que chama a função "toNum" que é aplicada a cada lista das listas recebidas, daqui ser feita a recursividade da função auxiliar.

toNumerosL :: [String] -> Int -> [[Int]]

A "toNumerosL" converte a lista devolvida pela "convertToSimbL" através de uma auxiliar que chama a função "toNum", assim como na função "toNumerosU".

toNumerosD :: [String] -> Int -> [[Int]]

A "toNumerosD" converte a lista devolvida pela "convertToSimbD" através de uma auxiliar que chama a função "toNum", tal como na função anterior.

toNumerosR :: [String] -> Int -> [[Int]]

A "toNumerosR" converte a lista devolvida pela "convertToSimbR" através de uma auxiliar que chama a função "toNum", tal como na função anterior.

toNum :: [String] -> [Int]

A função percorre uma lista convertendo os seus elementos num determinado número. Caso seja "#" (pedra) corresponde ao 0, se for "?" (tijolo) é 1 e se for " " (vazio) é 2.

verificaR :: [String] -> Int -> [(Int, Int)]

Esta função recorre à função "verificaAux" que tem como argumentos a função "toNumerosR" e "numerosR". Sendo o seu resultado uma lista com tuplos em que o primeiro elemento corresponde ao número do mapa onde haverá uma alteração e o segundo elemento é a alteração que irá estar associada.

verificaU :: [String] -> Int -> [(Int, Int)]

Esta função recorre à função "verificaAux" que tem como argumentos a função "toNumerosR" e "numerosR".

verificaD :: [String] -> Int -> [(Int, Int)]

Esta função recorre à função "verificaAux" que tem como argumentos a função "toNumerosR" e "numerosR".

verificaL :: [String] -> Int -> [(Int, Int)]

Esta função recorre à função "verificaAux" que tem como argumentos a função "toNumerosR" e "numerosR".

verificaAux :: [[Int]] -> [[Int]] -> [String] -> Int -> [(Int, Int)]

A função "verificaAux" devolve uma lista de tuplos. Se verificar-se que a função "ve" tem como resutado (0,0) então corresponde a um local do mapa que é apenas vazio pelo que a chama prossegue para o ponto seguinte logo é chamada recursivamente a função sem o primeiro elemento das respetivas listas. Se a "ve" corresponder a um jogador, armazenamos a sua informação dada pelo seu tuplo e repetimos o processo anterior. O mesmo se sucede se a "ve" for uma bomba. Caso não seja nenhum destes casos então guarda-se a informação proveniente da função "ve" numa lista e, mais uma vez, faz-se a recursiva.

confirmaJog :: [String] -> Int -> Bool

A "confirmaJog" verifica se existe algum jogador no local onde foi colocada uma bomba. Assim, verificamos se algum dos elementos da lista com os números do mapa onde se encontram as bombas prestes a explodir se repete na lista com os números do mapa onde estão os jogadores.

lugarJog :: [String] -> Int -> Int

Esta função indica o local onde se encontra o jogador que corresponde ao lugar da bomba.Para tal, verificamos se algum dos elementos da lista com os locais das bombas prestes a explodir se encontra também na lista com o número do mapa onde se estão os jogadores.

ve :: [Int] -> [Int] -> [String] -> Int -> (Int, Int)

Esta função devolve o tuplo associado a um determinado local. Caso o primeiro elemento da lista recebida seja um 0 então é devolvido um tuplo cujo primeiro elemento é o local do mapa associado a este, ou seja, o primeiro elemento de y e o segundo elemento é 0. Se for 1 então verifica-se a mesma situação diferindo apenas nos valores. Caso seja 2 e o número do mapa corresponda a um local com um PowerUp então é devolvido um tuplo cujo primeiro elemento é o número do mapa e o segundo é 3. Se o local da bomba corresponder ao número do mapa onde se encontra um jogador então é devolvido um tuplo cujo primeiro elemento é o número do mapa e o segundo é o número 4. Se for 2 e o número do mapa corresponder a um local com um jogador é devolvido um tuplo com o primeiro elemento igual ao número do mapa e o segundo é 4. Caso seja 2 e nesse local do mapa haja uma bomba então é devolvido um tuplo com o primeiro elemento igual ao número do mapa e o segundo é 5. Senão é devolvido o tuplo (0,0).

coX :: (Int, Int) -> [String] -> Int -> Int

A função devolve a coordenada x correspondente ao número do mapa que é o primeiro elemento do tuplo. Para tal dividimos o número pela dimensão do mapa.

coY :: (Int, Int) -> [String] -> Int -> Int

A função devolve a coordenada Y correspondente ao número do mapa que é o primeiro elemento do tuplo. Logo, ao resto da divisão do número pela dimensão do mapa subtraímos um.

juntaAlteraR :: [String] -> Int -> [String]

Esta função devolve o mapa com as alterações causadas à direita pela explosão da bomba. Para tal percorremos a lista devolvida pela "verificaR" realizando as alterações através da função "altera".

juntaAlteraMaisL :: [String] -> Int -> [String]

Esta função devolve o mapa com as alterações causadas à direita e à esquerda pela explosão da bomba. Para tal percorremos a lista devolvida pela "verificaL" realizando as alterações através da função "altera". Sendo que neste caso o mapa inicial é o mapa devolvido pela função "juntaAlteraR".

juntaAlteraMaisD :: [String] -> Int -> [String]

Esta função devolve o mapa com as alterações causadas à direita, à esquerda e para baixo pela explosão da bomba. Para tal percorremos a lista devolvida pela "verificaD" realizando as alterações através da função "altera". Sendo que neste caso o mapa inicial é o mapa devolvido pela função "juntaAlteraMaisL".

juntaAlteraFinal :: [String] -> Int -> [String]

Esta função devolve o mapa com as alterações causadas, para todos os lados, pela explosão da bomba. Para tal percorremos a lista devolvida pela "verificaU" realizando as alterações através da função "altera". Sendo que neste caso o mapa inicial é o mapa devolvido pela função "juntaAlteraMaisD".

altera :: [(Int, Int)] -> [String] -> Int -> [String]

A "altera" devolve o mapa com as alterações realizadas pela explosão de bombas. Caso a chama encontre uma pedra, o mapa devolvido é o inicial. Se encontrar um tijolo (1) este terá de ser eliminado do mapa. Quando encontra um PowerUp (3) a informação deste tem de ser eliminada. Se encontrar um jogador (4) este irá morrer, logo a sua informação será eliminada. Quando encontra uma bomba (5) esta irá explodir e como tal o seu tempo de ação passa para 1.

infTodosPU :: (Int, Int) -> [String] -> Int -> [String]

Esta função tem como resultado uma lista com a informação de todos os PowerUps presentes no mapa. Assim, percorremos a lista invertida com a informação do mapa até encontrarmos o símbolo dos PowerUps ("!" ou "+"). Quando encontrado é guardada a sua informação e fazemos a recursiva para verificar se existem mais PowerUps.

deletePU :: (Int, Int) -> [String] -> Int -> [String]

A função devolve a lista com a informação do PowerUp que será atingido pela bomba. Para isso, recorrendo à "infTodosPU" verificamos qual destes tem as coordenadas x e y iguais à coordenadas do número do mapa onde irá ocorrer a alteração.

infTodosJog :: (Int, Int) -> [String] -> Int -> [String]

Esta função tem como objetivo uma lista com a informação de todos os jogadores. Assim, percorremos a lista do mapa invertido até encontrarmos um algarismo que corresponde ao identificador de um jogador. Fazendo recursividade verificamos se existiam mais jogadores, caso existam guardamos também a sua informação.

infMorreJog :: (Int, Int) -> [String] -> Int -> [String]

"infMorreJog" devolve uma lista com a informação do jogador que irá morrer, devido ao facto de estar no raio de ação da bomba que irá explodir. Assim, recorrendo à "infTodosJog" verificamos qual destas tem as coordenadas x e y iguais à coordenadas do número do mapa onde irá ocorrer a alteração.

alg :: [Char] -> (Int, Int) -> [String] -> Int -> Char

Esta função devolve a coordenada x do jogador.

tuplo :: (Int, Int) -> [String] -> Int -> ([Char], [Char])

Esta função permite a separação do tempo de explosão da bomba da restante informação, sendo que o segundo elemento corresponde à restante informação.

novaBomb :: (Int, Int) -> [String] -> Int -> [Char]

A função altera a informação da bomba que irá ser afetada pela explosão, sendo alterado o seu tempo de explosão, que passará a ser 1.

infExpBomb :: (Int, Int) -> [String] -> Int -> String

"infExpBomb" devolve uma lista com a informação da bomba que irá explodir, devido ao facto de estar no raio de ação de outra bomba que irá explodir. Assim, recorrendo à "infTodasBomb" verificamos qual destas tem as coordenadas x e y iguais à coordenadas do número do mapa onde irá ocorrer a alteração. Uma vez que, a função à qual recorremos não contém o símbolo da bomba, quando encontrarmos a bomba nas condições anteriores acrescentamos o símbolo.

infListBomb :: [String] -> Int -> [String]

A função "infListBomb" devolve uma lista com a informação das bombas que explodiram. Para tal percorremos a lista do mapa até encontramos o símbolo da bomba "*" e chamamos recursivamente a função para verificar se existem mais bombas que irão explodir.

main :: IO ()

Função que permite testar esta tarefa