Main

Description

Descrption: funções que permitem o movimento dos jogadores no mapa Copyright: Ana Ribeiro Jéssica Lemos

Documentation

move :: [String] -> Int -> Char -> [String]

Caso não seja possível aplicar o movimente quer por não existir o jogador, ou o ponto para o qual pretende se deslocar seja uma pedra ou um tijolo ou o comando bomba seja aplicado a um jogador que não tem PowerUps, é devolvido o próprio mapa. Se for executado o comando bomba é chamada a função "mapaMoveB". Caso o ponto para onde pretende ir o jogador coincida com um PowerUp devolve a "mapaMove" pela ordem contrária. Nos restantes casos, devolve também a "mapaMove" por ordem contrária mais os restantes jogadores.

veJogador :: [String] -> Int -> Char -> Bool

Esta função verifica se o jogador está em jogo. Para tal percorremos a lista, através de uma recursiva, até encontrarmos a lista que contém um algarismo. Caso o algarismo corresponda ao algarismo do jogador ao qual foi aplicado o movimento então a função devolve True. Caso contrário a lista tornar-se-á vazia e desoverá False.

infJogador :: [String] -> Int -> Char -> [String]

"infJogador" dá a informação relativa a um jogador. Para tal percorremos a lista, através de uma recursiva, até encontrarmos a lista que contém um algarismo. Caso o algarismo corresponda ao algarismo do jogador ao qual foi aplicado o movimento então é devolvida a lista com a informação do jogador.

dim :: [String] -> Int -> Char -> Int

Esta indica a dimensão do mapa recebido. Para tal calculamos o tamanho da cabeça da lista, que corresponde à primeira linha do mapa constituída apenas por pedras.

coorJogNum :: [String] -> Int -> Char -> Int

Indica o numero do mapa correspondente à sua coordenada. Assim, multiplicamos o tamanho do mapa pela linha onde se encontra o jogador e somamos com a coluna mais um.

pPonto :: [String] -> Int -> Char -> Char

"pPonto" é o caracter associado à posição para a qual o jogador pretende ir.

preveMov :: [String] -> Int -> Char -> Int

Esta dá a posição do mapa para qual o jogador pretende ir. Se o comando for B então a coordenada correspondente é a mesma do jogador. Caso seja R o jogador movimenta-se para a direita logo a sua posição será o número atual do jogador menos um. O contrário acontece quando o jogador se movimenta para a esquerda L. Se o jogador pretender ir para cima a sua posição será o seu número menos a dimensão do mapa. Quando pretende ir para baixo, é o contrário.

coorNumJog :: [String] -> Int -> Char -> [Char]

A função dá a lista com as novas coordenadas do jogador. O y é dado pelo resto da divisão do número da posição para a qual vai pela dimensão do mapa, menos um. O x é obtido pela divisão inteira do número da posição para a qual vai pela dimensão do mapa.

mapaMove :: [String] -> Int -> Char -> [String]

A "mapaMove" junta a alteração da informação do jogador ao mapa. No caso da posição para a qual o jogador vai corresponder a um PowerUp Bombs e se o jogador tiver PowerUps, então percorremos a lista até encontrarmos o número do jogador que se vai movimentar. Enquanto tal não se verifica a função vai armazenando as listas eliminadas pela recursividade da função que não contém a lista do PowerUp apanhado pelo jogador, o que é assegurado pela utilização da função "mapaSPUb". Quando for encontrado junta o identificador do jogador com as coordenadas para qual vai e o restante da lista acrescentando o símbolo de + mais o restante mapa sem a informação antiga do jogador. Caso o jogador vá para uma posição na qual haja PowerUps e o jogador não tenha PowerUps então recorremos ao mesmo método utilizado anteriormente com a diferença de haver um espaço entre a coordenada x e o símbolo de Bombs. O mesmo acontece no caso de o jogador se movimentar para o local onde houver PowerUps Flames apenas se diferençiando na medida em que o seu símbolo é !. Caso o jogador não se movimente para um local onde haja PowerUps utilizamos uma recursiva até encontrarmos a lista com a informação do jogador. De seguida alteramos esta informação mundando apenas as suas coordenadas, juntando posteriormente ao resto da lista.

coordNumPUB :: [String] -> Int -> Char -> [String]

Esta função devolve-nos uma lista com todos os Bombs descobertos no mapa. Percorremos a lista até encontrar-mos um Bombs, quando tal se sucede armazenamo-lo e verificamos se existem mais tarvés da recursiva. A função termina quando todos elementos foram verificados (lista vazia).

coordNumPUF :: [String] -> Int -> Char -> [String]

Esta função devolve-nos uma lista com todos os Flames descobertos no mapa. Percorremos a lista até encontrar-mos um Flames, quando tal se sucede armazenamo-lo e verificamos se existem mais tarvés da recursiva. A função termina quando todos elementos foram verificados (lista vazia).

coordNumPowerUpB :: [String] -> Int -> Char -> [Int]

A função dá uma lista com os Bombs descobertos. Esta percorre todos os elementos da lista e aplica-lhes a função f, que multiplica o tamanho do mapa pela linha onde se encontra o Bombs e soma-lhe a coluna mais um.

coordNumPowerUpF :: [String] -> Int -> Char -> [Int]

A função dá uma lista com os Flames descobertos. Esta percorre todos os elementos da lista e aplica-lhes a função f, que multiplica o tamanho do mapa pela linha onde se encontra o Flames e soma-lhe a coluna mais um.

mapaSPUb :: [String] -> Int -> Char -> [String]

A "mapaSPUb" devolve o estado do mapa inicial sem a informação referente ao Bombs apanhado pelo jogador, recorrendo à função pré-definida (\).

mapaSPUf :: [String] -> Int -> Char -> [String]

A "mapaSPUf" devolve o estado do mapa inicial sem a informação referente ao Flame apanhado pelo jogador, recorrendo à função pré-definida (\).

existPU :: [String] -> Int -> Char -> Bool

"existPU" verifica se o jogador tem powerUps para colocar bombas.

existBomb :: [String] -> Int -> Char -> Bool

Esta diz se existe bomba no mapa do estado de jogo inicial. Percorremos a lista, através de uma recursiva, caso seja encontrado uma lista que contém o símbolo da bomba, é porque existe caso contrário a lista ficará vazia devolvendo False.

infBomb :: [String] -> Int -> Char -> [String]

Esta função dá a informação da bomba proveniente do mapa do estado do jogo inicial. Para tal percorremos a lista, através de uma recursiva, até encontrarmos a lista que contém o símbolo da bomba (*). Quando tal se sucede é devolvida esta lista. Para evitar problemas com a dimensão do mapa, invertemos a lista.

infNovaBomb :: [String] -> Int -> Char -> [String]

A função indica a informação da bomba que o jogador colocou. Esta é constituída com o símbolo da bomba, a coordenada do jogador que a colocou (dada pela função "coorNumJog"), o caracter coorrespondente ao jogador que a colocou, o raio de ação e tempo de explosão já definidos. Tivemos também em conta de que se o jogador tiver um PowerUp Flame o raio da bomba será 2.

mapaMoveB :: [String] -> Int -> Char -> [String]

Esta função junta as bombas ordenadamente no mapa do estado inicial.Se o jogador pretender colocar uma bomba num local onde já existe entao é lhe devolvido o mapa anterior. Nos restantes casos em que existe bomba, juntamos a "mapaMe" seguida das bombas ordenadas e da "mapaTu" por ordem contrária.Caso contrário unimos o "mapaTi" com a bomba nova e com o "mapaTo" por ordem contrário.

posBomb :: [String] -> Int -> Char -> [String]

A função ordena as bombas, ou seja, ordena a bomba proveniente do estado do mapa inicial e a bomba colocada pelo jogador. Caso não haja bomba no estado do mapa inicial é devolvida apenas a informação da nova bomba. Senão é comparado o número associado à bomba anterior com o número associado à nova bomba. É de notar que quando o jogador pretende colocar uma bomba onde já existe uma bomba é devolvido o estado do mapa anterior.

coorBombNum :: [String] -> Int -> Char -> Int

A função indica o número do mapa da bomba colocada que o jogador quer colocar. Assim, multiplicamos o tamanho de mapa pela linha onde se encontra a bomba que vai ser colocada e somamos com a coluna mais um.

coorBombExisNum :: [String] -> Int -> Char -> Int

A função indica o número do mapa da bomba proveniente do estado inicial. Assim, multiplicamos o tamanho de mapa pela linha onde se encontra a bomba e somamos com a coluna mais um.

mapaMe :: [String] -> Int -> Char -> [String]

A "mapaMe" percorre a lista e guarda a informação até encontrar a informação da bomba ("infBomb").

mapaTu :: [String] -> Int -> Char -> [String]

Esta função percorre todos os elementos da lista, no entanto inverte-a, para armazenar a informação posterior à bomba.

mapaTi :: [String] -> Int -> Char -> [String]

A "mapaTi" percorre a lista e guarda a informação até encontrar um algarismo, fornecendo um mapa sem informações dos jogadores.

mapaTo :: [String] -> Int -> Char -> [String]

A função percorre todos os elementos da lista, no entanto inverte-a, para armazenar a informação dos jogadores.

main :: IO ()