Main

Description

Este módulo contém as funções, da Tarefa 5, que consiste na criação de uma Interface Gráfica do jogo Bomberman e a respetiva jogabilidade.

Documentation

type ActualTime = Int

O type ActualTime, corresponde ao tempo atual do jogo.

type LastTimeReageT = Float

O type LastTimeReageT, corresponde ao somatório de (1/FrameRate).

type Tamanho = Int

O type Tamanho, corresponde ao tamanho do mapa.

type Semente = Int

O type Semente, corresponde à semente do mapa.

type Linha = Int

O type Linha, corresponde à Linha do mapa.

type Coluna = Int

O type Coluna, corresponde à Coluna do mapa.

type WindowAltura = Float

O type WindowAltura, corresponde à altura do mapa em píxeis.

type WindowLargura = Float

O type WindowLargura, corresponde à Largura do mapa em píxeis.

type Gamer = Int

O type Gamer, corresponde ao número do jogador.

type Estado = (InfGame, Tamanho, Semente, LastTimeReageT, Imagens, ActualTime, WindowLargura, WindowAltura, Gamer)

Descrição:

O Estado, contém a informação necessária para executar todas as funções.

Informação do Jogo (Tabuleiro,Power Ups,Bombas,Jogadores);

Tamanho do Mapa;

Semente do Mapa;

Somatório de (1/Frame Rate);

Imagens utilizadas na interface gráfica;

Tempo atual do Jogo;

Largura em Pixeis do Mapa;

Altura em Pixeis do Mapa;

Número do Jogador;

data Imagens

Descrição:

O tipo Imagens contém todas as imagens usadas na interface gráfica.

Constructors

Imagens

Fields

pedra :: Picture

bomba :: Picture

caixa :: Picture

jogador0 :: Picture

jogador1 :: Picture

jogador2 :: Picture

jogador3 :: Picture

powUpBomb :: Picture

powUpFlame :: Picture

fundo :: Picture

explosion :: Picture

estadoInicial :: IO Estado

Descrição:

A função estadoInicial, retorna o estado ínicial do jogo, com a informação do mapa, tamanho, semente, somatório do tempo do frame rate, as imagens, tempo atual de jogo , tamanho da janela (size*50), altura da janela(size*50), número do jogador.

desenhaEstado :: Estado -> Picture

Descrição:

A função desenhaEstado, recebe o estado de jogo, e cria a interface gráfica através desse estado de jogo.

determinaPixels1 :: Int -> Float -> [Float]

Descrição:

A função determinaPixels1, determina a lista dos píxeis por coluna para um dado mapa, no eixo dos xx, em que cada elemento da lista corresponde à coodenada x para colocar uma imagem, portanto o elemento na posição zero corresponde aos píxeis no mapa para a coluna zero.

Exemplos:

>>> determinaPixels1 9 (-200)
[-200.0,-150.0,-100.0,-50.0,0.0,50.0,100.0,150.0,200.0]

Notas:

Como podemos ver no exemplo, para o mapa de tamanho 9, temos uma janela com 450 Pixeis, pois cada imagem tem 50x50 Pixeis, logo o tamanho do mapa na janela é dado por (50*9).

Como o tamanho é 450 Pixeis, então a primeira imagem vai começar nos -200, e a partir desse valor a nossa função determina os restantes, em que cada posição da lista corresponde a uma coluna no mapa bomberman.

determinaPixels2 :: Int -> Float -> [Float]

Descrição:

A função determinaPixels1, determina a lista dos píxeis por coluna para um dado mapa, no eixo dos yy, em que cada elemento da lista corresponde à coodenada y para colocar uma imagem, portanto o elemento na posição zero corresponde aos píxeis no mapa para a linha zero.

Exemplos:

>>> determinaPixels2 9 (200)
[200.0,150.0,100.0,50.0,0.0,-50.0,-100.0,-150.0,-200.0]

Notas:

Como podemos ver no exemplo, para o mapa de tamanho 9, temos uma janela com 450 Pixeis, pois cada imagem tem 50x50 Pixeis, logo o tamanho do mapa na janela é dado por (50*9).

Como o tamanho é 450 Pixeis, então a primeira imagem vai começar nos 200, e a partir desse valor a nossa função determina os restantes, em que cada posição da lista corresponde a uma linha no mapa bomberman.

desenhaPowerUps :: Imagens -> [PowUp] -> InfGame -> Size -> Float -> Float -> [Picture]

Descrição:

A função desenhaPowerUps, coloca as imagens dos Power Ups destapados no mapa, através da sua linha e coluna.

desenhaBombas :: Imagens -> [Bomba] -> Size -> Float -> Float -> InfGame -> ([Picture], [Picture])

Descrição:

A função desenhaBombas, coloca as imagens das bombas na janela, através das posições das mesmas.

desenhaExplosionBomb :: Imagens -> [Posicao] -> Int -> Float -> Float -> [Picture]

Descrição:

A função desenhaExplosionBomb, coloca imagens no flame causado pela explosão das bombas.

desenhaJogadores :: Imagens -> Players -> Size -> Int -> Float -> Float -> [Picture]

A função desenhaJogadores, coloca imagens dos jogadores através das suas posições no mapa.

desenhaTabuleiro :: Imagens -> [Bloco] -> Coluna -> Linha -> Size -> Float -> Float -> [Picture]

Descrição:

A função desenhaTabuleiro, coloca as imagens das Pedras, Tijolos, e vélulas Vazias na Janela, através das coordenadas das mesmas.

reageEvento :: Event -> Estado -> Estado

Descrição:

A função reageEvento, modifica o estado de jogo, através de um evento, ou seja, um comando introduzido por um utilizador.

reageTempo :: Float -> Estado -> Estado

Descrição:

A função reageTempo, modifica o estado do jogo com a passagem do tempo.

Notas:

Aqui usamos o somatório do frame rate, quando este é superior a 1, assumimos que passou 1 segundo, portanto podemos verificar as alterações no estado de jogo, com a passagem de 1 segundo, através da função avanca, definida na T4Lib.

fr :: Int

Frame rate

dm :: Display

Display mode

main :: IO ()

Função principal que invoca o jogo.