Main

Description

Copyrigth: Alexandre Mendonça Pinho a82441@alunos.uminho.pt; Joel Filipe Esteves Gama a82202@alunos.uminho.pt;

O módulo contém as fuñções necessárias para o encode e decode.

Documentation

type Bit = Int

Int que representa um bit por 0 ou 1.

type Bits = [Bit]

Lista de Bits, cada elemento representando um bit.

encode :: [String] -> String

Função que codifica o estado inicial do mapa.

splitMap :: [String] -> ([String], [String])

Função que separa as linhas do mapa que contêm as celulas das que contêm os Power Ups, as Bombas, e os Jogadores

splitExtras :: [String] -> ([String], [String], [String])

Função que separa num tripo os Power Ups, as Bombas, e os Jogadores, respetivamente.

fst3 :: (a, a, a) -> a

Função que retorna o primeiro elemento de um triplo.

snd3 :: (a, a, a) -> a

Função que retorna o segundo elemento de um triplo.

thrd3 :: (a, a, a) -> a

Função que retorna o terceiro elemento de um triplo.

dropPredefined :: [String] -> Int -> [String]

Função que retira as células predefinidas de um mapa.

dropPredefinedAux :: Int -> [String] -> Int -> [String]

Função auxiliar que, dada uma linha, uma lista de Strings e uma dimensão, retira

mkMapBits :: [Char] -> Bits

Função que retorna uma lista de Bits, com False a representar espaços vazios no mapa, e True tijolos.

mkPWsBits :: [String] -> Bits

Função que retorna uma lista de Bits, com False a representar a inexistência de Power Ups, e True a existência de Power Ups.

mkBombsBits :: [String] -> Bits

Função que retorna uma lista de Bits com as informações acerca das bombas no mapa (posição, jogador que a colocou, raio de explosão e intantes de explosão).

mkPlayersBits :: [String] -> Bits

Função que retorna uma lista de Bits com as informações dos jogadores (posição, número e tipo dos Power Ups que cada um possui).

countPWs :: [String] -> Int

Função que conta o número de Power Ups que cada jogador possui.

countBombs :: [String] -> Int

Função que conta o número de bombas que cada jogador tem colocadas no mapa.

countPlayers :: [String] -> Bits

Função que devolve uma lista com os jogadores existentes no jogo e os Power Ups que cada um possui.

playerExists :: Int -> [String] -> Int

Função que verifica se um dado jogador existe no mapa, devolvendo '1' em caso afirmativo ou '0' em caso negativo.

playerHasPWs :: Int -> [String] -> Int

Função que verifica se um jogador possui Power Ups, devolvendo '1' em caso afirmativo ou '0' em caso negativo.

mkChar :: Bits -> Char

Função que recebe uma lista de 8 Bits (1 ou 0) e retorna o Char com o valor do Byte.

mkChars :: Bits -> String

Função que dada uma lista de Ints (1 ou 0), retorna um String cujos caracteres tem valores dos bytes formados ao agrupar a lista em conjuntos de 8. | Se restarem elementos da lista (se forem menos de 8), o último Char do output e contruido adicionando zeros (0) à lista.

intToBits :: Int -> Int -> Bits

Função que converte um número decimal para a sua representação binária, utilizando um número de bits dado como argumento. | Os bits tem ordem decrescente de valorizacao.

iTBAux :: Int -> Int -> Bits

Função auxiliar de intToBits.

decode :: [Char] -> [String]

Função que descodifica a versão comprimida do mapa. decode :: String -> [String]

mkMap :: Int -> Int -> Bits -> [String]

Função que, dada a linha, a dimensão e a lista de Bits, devolve uma lista de Strings que representam as linhas do mapa.

mkMapCols :: Int -> Int -> Int -> Bits -> String

Função que, dada a coluna, a linha, a dimensão e a lista de Bits, devolve uma Strings que representa as colunas do mapa.

lista_vazios :: [(Int, Int)]

Lista de coordenadas das posições do mapa que ficarão obrigatoriamente vazias. | Números negativos significa que se conta a posição a partir do fim da linha ou coluna.

normDim :: Int -> [(Int, Int)] -> [(Int, Int)]

Função que adiciona a dimensao aos números negativos para poder comparar as coordenadas.

>>> normDim 7 [(1,1),(1,(-1)),((-1),(-1))]
              [(1,1),(1,5),(5,5)]

mkBlock :: Int -> Char

Função que dado '0' o converte num espaço ou dado um '1' o converte num Power Up.

mkPWs :: Bits -> [String]

Função que dada uma lista de Ints retorna uma lista de Strings, cada um representando um Power Up. | A função processa a lista de Ints de 15 em 15, pois esse é o número de bits necessário para codificar cada Power Up.

mkPW :: Bits -> String

Função que dada uma lista de 15 Ints retorna um String que representa um Power Up.

mkBombs :: Bits -> [String]

Função que dada uma lista de Ints retorna uma lista de Strings, cada um representando um Power Up. | A função processa a lista de Ints de 27 em 27, pois esse é o número de bits necessário para codificar cada Power Up.

mkBomb :: Bits -> String

Função que dada uma lista de 27 Ints retorna um String que representa um Power Up.

mkPlayers :: Int -> Bits -> Bits -> [String]

Função que dado um número (número do jogador - 0 como valor de acumulação), e duas listas de Ints retorna uma lista de Strings, cada um representando um jogador.

mkPlayer :: Int -> Bits -> Bits -> (String, Bits)

Função que dado um número (número do jogador - 0 como valor de acumulação), e duas listas de Ints retorna um tuplo cujo primeiro elemento e um String e o segundo uma lista de Int, onde o String representa um jogador, e a lista de Ints e a lista contendo todos aqueles que não foram necessários para fazer o String. | O segundo argumento, uma lista de 8 bits, codifica se os jogadores existem e se cada um deles tem Power Ups acumulados (nas posicoes 0, 2, 4 e 6, e 1, 3, 5 e 7, respetivamente).

charToBits :: Char -> Bits

Função que recebe um Char e retorna a sua ordem no formato de Bits.

charsToBits :: [Char] -> Bits

Função que recebe uma lista de Char e retorna a sua ordem no formato de Bits.

xBitsToInt :: Int -> Bits -> (Int, Bits)

Função que dado uma série de Ints (1 ou 0), retorna um tuplo (Int, Bits) em que o primeiro elemento é um número feito a partir de um certo número de Ints na lista (determinado pelo utilizador), e uma outra lista de Ints que não foram utilizados ao calcular o Int.

bitsToInt :: Bits -> Int

Função que dada uma lista de Ints (1 ou 0), converte o número em notação binária por eles representado no correspondente número em sistema de base 10.

getNumCells :: Int -> Int

Função que dada a dimensão de um mapa, calcula o número de casas não fixas desse mapa.

main :: IO ()

Função main.