Main

Description

Documentation

type Bomb = (Pos, Int)

Representação de uma bomba (posição da bomba, tamanho da explosão)

avanca :: Map -> Int -> Map

Avança o estado do jogo um tick para a frente, explodindo as bombas (incluindo os seus efeitos) e colocando os blocos da espiral nos ticks finais.

tickBombs :: Map -> Map

Reduz o contador decrescente de todas as bombas por um tick.

removeExploded :: Map -> Map

Retira do estado do jogo todos os elementos que foram marcados como retirados.

takeExploding :: Map -> ([Bomb], Map)

Função que, dado um mapa, retorna um par de uma lista de bombas que explodem no instante atual, e o mapa sem as linhas dessas bombas.

explode :: Map -> [Bomb] -> Map

Função que dado um mapa e uma lista de bombas retorna um outro mapa onde se sentem os efeitos das explosões das bombas.

explodeBomb :: Map -> Bomb -> Int -> Map

Função que explode a bomba na sua própria posição, e nas 4 direções em que a bomba explode.

explodeBombAux :: Int -> Int -> Pos -> Map -> Pos -> Map

Função que, dado o tamanho da explosão, o tamanho atual, a posição da bomba, um mapa, e a posição atual da explosão, retorna um mapa onde se sentem os efeitos da explosão da bomba apenas na direção atual.

calcNextPos :: Pos -> Pos -> Pos

Função que dadas duas posições alinhados pelo eixo horizontal ou vertical, retorna a próxima posição, ou seja, aquela imediatamente após a segunda no sentido oposto desta para a primeira.

>>> calcNextPos (1,1) (4,1) = (5,1)

inMap :: Int -> Pos -> Bool

Função que dadas a dimensão de um mapa e uma posição, testa se essa posição se encontra dentro desse mapa.

triggerBomb :: Map -> Pos -> Map

Função que dados um mapa e uma posição, retorna um mapa cuja contador da bomba na dada posiçäo é reduzido a 1, fazendo a bomba explodir no próximo tick.

killPlayers :: Map -> Pos -> Map

Função que dados um mapa e uma posição, retorna um outro mapa sem jogadores nessa posição.

main :: IO ()

Função main.