Main

Description

Copyrigth: Alexandre Mendonça Pinho a82441@alunos.uminho.pt; Joel Filipe Esteves Gama a82202@alunos.uminho.pt;

O módulo incorpóra os mecanismos necessários para geração de mapas.

Documentation

lista_vazios :: [(Int, Int)]

Lista de coordenadas das posições do mapa que ficarão obrigatoriamente vazias. | Números negativos significa que se conta a posição a partir do fim da linha ou coluna.

mapa :: Int -> Int -> [String]

Função que, dada a dimensão e semente, deverá devolver uma lista de strings com os elementos do mapa.

mkMap :: Int -> Int -> [Int] -> [String]

Função que, dada a linha, a dimensão e a lista de números aleatórios, devolve uma lista de Strings que representam as linhas do mapa.

mkLine :: Int -> Int -> Int -> [Int] -> String

Função que, dada a coluna, a linha, a dimensão e a lista de números aleatórios, devolde uma String correspondente à linha do mapa.

mkBlock :: Int -> Char

Função que dado um número aleatório o converte num tijolo, espaço ou Power Up.

normDim :: Int -> [(Int, Int)] -> [(Int, Int)]

Função que adiciona a dimensao aos números negativos para poder comparar as coordenadas.

>>> normDim 7 [(1,1),(1,(-1)),((-1),(-1))]
              [(1,1),(1,5),(5,5)]

getBombs :: [String] -> [String]

Função que constroi uma lista de Strings em que cada um representa um Power Up Bomb no mapa.

getLBombs :: Int -> [String] -> [String]

Função que, dado uma linha e uma lista de Strings, devolve uma lista com as linhas das Bombs.

getCBombs :: Int -> Int -> String -> [String]

Função que, dado uma linha, uma coluna e uma String, devolve uma lista com as colunas das Bombs.

getFlames :: [String] -> [String]

Função que constroi uma lista de Strings em que cada um representa um Power Up Flame no mapa.

getLFlames :: Int -> [String] -> [String]

Função que, dado uma linha e uma lista de Strings, devolve uma lista com as linhas dos Flames.

getCFlames :: Int -> Int -> String -> [String]

Função que, dado uma linha, uma coluna e uma String, devolve uma lista com as colunas dos Flames.

replacePowerUps :: [String] -> [String]

Função que procura os caracteres correspondentes aos Power Ups e os subtitui por tijolos.

replaceLPowerUps :: Int -> [String] -> [String]

Função que, dado um número e uma lista de Strings, devolve uma lista com as linhas dos Power Ups.

replaceCPowerUps :: Int -> Int -> String -> String

Função que, dado uma linha, uma coluna e uma String, devolve uma String das colunas dos Power Ups.

main :: IO ()

Função main.